refactor: move all environment generation to teal.environment
