+++ teal/checker/file_checker.tl
@@ -10,8 +10,8 @@ local read_file_skipping_bom = util.read_file_skipping_bom
 local record file_checker
 end
 
-function file_checker.check(filename: string, env?: Env, fd?: FILE): Result, string
-   if env and env.loaded and env.loaded[filename] then
+function file_checker.check(env: Env, filename: string, fd?: FILE): Result, string
+   if env.loaded and env.loaded[filename] then
       return env.loaded[filename]
    end
 
@@ -30,7 +30,7 @@ function file_checker.check(filename: string, env?: Env, fd?: FILE): Result, str
       return nil, "could not read " .. filename .. ": " .. err
    end
 
-   return string_checker.check(input, env, filename)
+   return string_checker.check(env, input, filename)
 end
 
 return file_checker
