+++ teal/checker/require_file.lua
@@ -79,7 +79,7 @@ function require_file.require_module(env, w, module_name, opts)
       }
       env.defaults = defaults
 
-      local found_result, err = file_checker.check(found, env, fd)
+      local found_result, err = file_checker.check(env, found, fd)
       assert(found_result, err)
 
       env.defaults = save_defaults
