+++ teal/checker/type_checker.tl
@@ -6959,16 +6959,14 @@ do
    type_checker.check = function(ast: Node, filename?: string, opts?: CheckOptions, env?: Env): Result, string
       filename = filename or "?"
 
-      opts = opts or {}
-
       if not env then
          local err: string
-         env, err = environment.default()
-         env.defaults = opts
+         env, err = environment.new(opts)
          if err then
             return nil, err
          end
       end
+      opts = opts or env.defaults
 
       local self: TypeChecker = {
          filename = filename,
