+++ teal/precompiled/default_env.lua
@@ -1,9 +1,9 @@
 local K1,K2,K3,K4,K5,K6,K7,K8,K9,K10="typename","typeid","./teal/default/stdlib.d.tl","tuple","typevar","min_arity","function","./teal/default/prelude.d.tl","maybe_method","typearg"
 local K11,K12,K13,K14,K15,K16,K17,K18,K19,K20="string","integer","number","is_va","nominal","names","found","types","typeargs","generic"
 local K21,K22,K23,K24,K25,K26,K27,K28,K29,K30="typedecl","union","fresh","declname","boolean","const","attribute","needs_compat","values","fields"
-local K31,K32,K33,K34,K35,K36,K37,K38,K39,K40="field_order","elements","array","closed","thread","special_function_handler","record","enumset","AnyFunction","typevals"
-local K41,K42,K43,K44,K45,K46,K47,K48,K49,K50="metatable","close","GetInfoTable","is_method","OpenMode","HookFunction","FileStringMode","FileNumberMode","FileMode","DateTable"
-local K={"@prelude","SeekWhence","SetVBufMode","flush","lines","write","interface_list","interfaces_expanded","__close","used_as_type","interface","Function","debug","HookEvent","getmetatable","setmetatable","userdata","Numeric","constraint","DateMode","remove","unpack","SortFunction","PackTable","tupletable"}
+local K31,K32,K33,K34,K35,K36,K37,K38,K39,K40="field_order","elements","array","closed","thread","special_function_handler","record","enumset","AnyFunction","--[[special_function]]"
+local K41,K42,K43,K44,K45,K46,K47,K48,K49,K50="typevals","metatable","close","GetInfoTable","unpack","is_method","OpenMode","HookFunction","FileStringMode","FileNumberMode"
+local K={"@prelude","SeekWhence","SetVBufMode","flush","lines","write","FileMode","interface_list","interfaces_expanded","__close","used_as_type","interface","Function","debug","HookEvent","getmetatable","setmetatable","__newindex","field_comments","userdata","Numeric","constraint","DateTable","DateMode","remove","format","gmatch","packsize","upper","--[[needs_compat]]","SortFunction","PackTable","tupletable"}
 local T1 = {[K34]=true,f=K3,[K2]=694,[K1]=K21,x=1,y=26,}
 local T2 = {f=K3,[K2]=562,[K1]=K21,x=4,y=29,}
 local T3 = {f=K3,[K2]=564,[K1]=K21,x=4,y=33,}
@@ -27,29 +27,29 @@ local T20 = {[K34]=true,f=K3,[K2]=1341,[K1]=K21,x=4,y=227,}
 local T21 = {f=K3,[K2]=1646,[K1]=K21,x=4,y=301,}
 local T22 = {f=K3,[K2]=2298,[K1]=K5,[K5]="A",x=11,y=302,}
 local T23 = {f=K3,[K2]=1639,[K1]=K21,x=24,y=299,}
-local T0 = {["..."]={t={f=K[1],[K14]=true,[K4]={[1]={f=K[1],[K2]=2746,[K1]=K11,x=1,y=1,},},[K2]=2747,[K1]=K4,x=1,y=1,},},["@is_va"]={t={f=K[1],[K2]=2748,[K1]="any",x=1,y=1,},},FILE={t=T1,[K[10]]=true,},["_VERSION"]={[K27]=K26,t={f=K3,[K2]=2175,[K1]=K11,x=14,y=414,},},any={t={[K34]=true,def={[K24]="any",f=K8,[K31]={},[K30]={},[K2]=5,[K1]=K[11],x=1,y=8,},f=K8,[K2]=6,[K1]=K21,x=1,y=8,},},arg={[K27]=K26,t={[K32]={f=K3,[K2]=1860,[K1]=K11,x=10,y=362,},f=K3,[K2]=1861,[K1]=K33,x=9,xend=16,y=362,yend=362,},},["assert"]={[K27]=K26,[K28]=true,t={f=K3,[K23]=true,t={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2495,[K1]=K5,[K5]="A@43",x=27,y=363,},[2]={f=K3,[K2]=2496,[K1]=K5,[K5]="B@43",x=32,y=363,},[3]={f=K3,[K2]=1867,[K1]="any",x=40,y=363,},},[K2]=2497,[K1]=K4,x=44,y=363,},f=K3,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2498,[K1]=K5,[K5]="A@43",x=46,y=363,},},[K2]=2499,[K1]=K4,x=44,y=363,},[K36]="assert",[K2]=2500,[K1]=K7,x=12,y=363,},[K19]={[1]={f=K3,[K10]="A@43",[K2]=2493,[K1]=K10,x=21,y=363,},[2]={f=K3,[K10]="B@43",[K2]=2494,[K1]=K10,x=24,y=363,},},[K2]=2501,[K1]=K20,x=4,y=365,},},["collectgarbage"]={[K27]=K26,t={f=K3,[K2]=1883,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K17]={def={[K24]="CollectGarbageCommand",[K38]={["collect"]=true,["count"]=true,["restart"]=true,stop=true,},f=K3,[K2]=1842,[K1]="enum",x=4,y=337,},f=K3,[K2]=1843,[K1]=K21,x=4,y=337,},[K16]={[1]="CollectGarbageCommand",},[K2]=1873,[K1]=K15,x=31,y=365,},},[K2]=1874,[K1]=K4,x=53,y=365,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1876,[K1]=K13,x=55,y=365,},},[K2]=1875,[K1]=K4,x=53,y=365,},[K2]=1872,[K1]=K7,x=20,y=365,},[2]={args={f=K3,[K4]={[1]={f=K3,[K17]={def={[K24]="CollectGarbageSetValue",[K38]={["setpause"]=true,["setstepmul"]=true,step=true,},f=K3,[K2]=1844,[K1]="enum",x=4,y=344,},f=K3,[K2]=1845,[K1]=K21,x=4,y=344,},[K16]={[1]="CollectGarbageSetValue",},[K2]=1878,[K1]=K15,x=29,y=366,},[2]={f=K3,[K2]=1879,[K1]=K12,x=53,y=366,},},[K2]=1880,[K1]=K4,x=61,y=366,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1882,[K1]=K13,x=63,y=366,},},[K2]=1881,[K1]=K4,x=61,y=366,},[K2]=1877,[K1]=K7,x=20,y=366,},[3]={args={f=K3,[K4]={[1]={f=K3,[K17]={def={[K24]="CollectGarbageIsRunning",[K38]={["isrunning"]=true,},f=K3,[K2]=1846,[K1]="enum",x=4,y=350,},f=K3,[K2]=1847,[K1]=K21,x=4,y=350,},[K16]={[1]="CollectGarbageIsRunning",},[K2]=1885,[K1]=K15,x=29,y=367,},},[K2]=1886,[K1]=K4,x=53,y=367,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1888,[K1]=K25,x=55,y=367,},},[K2]=1887,[K1]=K4,x=53,y=367,},[K2]=1884,[K1]=K7,x=20,y=367,},[4]={args={f=K3,[K4]={[1]={f=K3,[K2]=1890,[K1]=K11,x=29,y=368,},[2]={f=K3,[K2]=1891,[K1]=K13,x=39,y=368,},},[K2]=1892,[K1]=K4,x=46,y=368,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1895,[K1]=K22,[K18]={[1]={f=K3,[K2]=1894,[K1]=K25,x=49,y=368,},[2]={f=K3,[K2]=1896,[K1]=K13,x=59,y=368,},},x=49,y=368,},},[K2]=1893,[K1]=K4,x=46,y=368,},[K2]=1889,[K1]=K7,x=20,y=368,},},x=18,y=366,},},["coroutine"]={[K28]=true,t={[K34]=true,def={[K24]="coroutine",f=K3,[K31]={[1]=K[12],[2]=K42,[3]="create",[4]="isyieldable",[5]="resume",[6]="running",[7]="status",[8]="wrap",[9]="yield",},[K30]={[K[12]]=T7,[K42]={args={f=K3,[K4]={[1]={f=K3,[K2]=703,[K1]=K35,x=20,y=63,},},[K2]=704,[K1]=K4,x=27,y=63,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=706,[K1]=K25,x=29,y=63,},[2]={f=K3,[K2]=707,[K1]=K11,x=38,y=63,},},[K2]=705,[K1]=K4,x=27,y=63,},[K2]=702,[K1]=K7,x=11,y=63,},["create"]={args={f=K3,[K4]={[1]={f=K3,[K17]=T7,[K16]={[1]=K[12],},[K2]=709,[K1]=K15,x=21,y=64,},},[K2]=710,[K1]=K4,x=30,y=64,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=712,[K1]=K35,x=32,y=64,},},[K2]=711,[K1]=K4,x=30,y=64,},[K2]=708,[K1]=K7,x=12,y=64,},["isyieldable"]={args={f=K3,[K4]={},[K2]=714,[K1]=K4,x=27,y=65,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=716,[K1]=K25,x=29,y=65,},},[K2]=715,[K1]=K4,x=27,y=65,},[K2]=713,[K1]=K7,x=17,y=65,},["resume"]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=718,[K1]=K35,x=21,y=66,},[2]={f=K3,[K2]=719,[K1]="any",x=29,y=66,},},[K2]=720,[K1]=K4,x=36,y=66,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=722,[K1]=K25,x=38,y=66,},[2]={f=K3,[K2]=723,[K1]="any",x=47,y=66,},},[K2]=721,[K1]=K4,x=36,y=66,},[K2]=717,[K1]=K7,x=12,y=66,},["running"]={args={f=K3,[K4]={},[K2]=725,[K1]=K4,x=23,y=67,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=727,[K1]=K35,x=25,y=67,},[2]={f=K3,[K2]=728,[K1]=K25,x=33,y=67,},},[K2]=726,[K1]=K4,x=23,y=67,},[K2]=724,[K1]=K7,x=13,y=67,},["status"]={args={f=K3,[K4]={[1]={f=K3,[K2]=730,[K1]=K35,x=21,y=68,},},[K2]=731,[K1]=K4,x=28,y=68,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=733,[K1]=K11,x=30,y=68,},},[K2]=732,[K1]=K4,x=28,y=68,},[K2]=729,[K1]=K7,x=12,y=68,},wrap={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={f=K3,[K2]=2208,[K1]=K5,[K5]="F@23",x=22,y=69,},},[K2]=2209,[K1]=K4,x=24,y=69,},f=K3,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2210,[K1]=K5,[K5]="F@23",x=26,y=69,},},[K2]=2211,[K1]=K4,x=24,y=69,},[K2]=2212,[K1]=K7,x=10,y=69,},[K19]={[1]={f=K3,[K10]="F@23",[K2]=2207,[K1]=K10,x=19,y=69,},},[K2]=2213,[K1]=K20,x=4,y=70,},["yield"]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=742,[K1]="any",x=20,y=70,},},[K2]=743,[K1]=K4,x=27,y=70,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=745,[K1]="any",x=29,y=70,},},[K2]=744,[K1]=K4,x=27,y=70,},[K2]=741,[K1]=K7,x=11,y=70,},},[K2]=695,[K1]=K37,x=1,y=60,},f=K3,[K2]=746,[K1]=K21,x=1,y=60,},},[K[13]]={[K28]=true,t={[K34]=true,def={[K24]=K[13],f=K3,[K31]={[1]=K43,[2]=K[14],[3]=K46,[4]=K39,[5]=K[13],[6]="gethook",[7]="getinfo",[8]="getlocal",[9]=K[15],[10]="getregistry",[11]="getupvalue",[12]="getuservalue",[13]="sethook",[14]="setlocal",[15]=K[16],[16]="setupvalue",[17]="setuservalue",[18]="traceback",[19]="upvalueid",[20]="upvaluejoin",},[K30]={[K39]=T8,[K43]=T9,[K[14]]=T10,[K46]=T11,[K[13]]={args={f=K3,[K4]={},[K2]=781,[K1]=K4,x=4,y=100,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={},[K2]=782,[K1]=K4,x=20,y=99,},[K2]=780,[K1]=K7,x=11,y=99,},["gethook"]={args={f=K3,[K4]={[1]={f=K3,[K2]=784,[K1]=K35,x=24,y=100,},},[K2]=785,[K1]=K4,x=31,y=100,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K17]=T11,[K16]={[1]=K46,},[K2]=787,[K1]=K15,x=33,y=100,},[2]={f=K3,[K2]=788,[K1]=K12,x=47,y=100,},},[K2]=786,[K1]=K4,x=31,y=100,},[K2]=783,[K1]=K7,x=13,y=100,},["getinfo"]={f=K3,[K2]=804,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=791,[K1]=K22,[K18]={[1]={f=K3,[K17]=T8,[K16]={[1]=K39,},[K2]=790,[K1]=K15,x=22,y=102,},[2]={f=K3,[K2]=792,[K1]=K12,x=36,y=102,},},x=22,y=102,},},[K2]=793,[K1]=K4,x=44,y=102,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K17]=T9,[K16]={[1]=K43,},[K2]=795,[K1]=K15,x=46,y=102,},},[K2]=794,[K1]=K4,x=44,y=102,},[K2]=789,[K1]=K7,x=13,y=102,},[2]={args={f=K3,[K4]={[1]={f=K3,[K2]=798,[K1]=K22,[K18]={[1]={f=K3,[K17]=T8,[K16]={[1]=K39,},[K2]=797,[K1]=K15,x=22,y=103,},[2]={f=K3,[K2]=799,[K1]=K12,x=36,y=103,},},x=22,y=103,},[2]={f=K3,[K2]=800,[K1]=K11,x=45,y=103,},},[K2]=801,[K1]=K4,x=52,y=103,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K17]=T9,[K16]={[1]=K43,},[K2]=803,[K1]=K15,x=54,y=103,},},[K2]=802,[K1]=K4,x=52,y=103,},[K2]=796,[K1]=K7,x=13,y=103,},[3]={args={f=K3,[K4]={[1]={f=K3,[K2]=806,[K1]=K35,x=22,y=104,},[2]={f=K3,[K2]=808,[K1]=K22,[K18]={[1]={f=K3,[K17]=T8,[K16]={[1]=K39,},[K2]=807,[K1]=K15,x=30,y=104,},[2]={f=K3,[K2]=809,[K1]=K12,x=44,y=104,},},x=30,y=104,},[3]={f=K3,[K2]=810,[K1]=K11,x=53,y=104,},},[K2]=811,[K1]=K4,x=60,y=104,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,[K17]=T9,[K16]={[1]=K43,},[K2]=813,[K1]=K15,x=62,y=104,},},[K2]=812,[K1]=K4,x=60,y=104,},[K2]=805,[K1]=K7,x=13,y=104,},},x=11,y=103,},["getlocal"]={f=K3,[K2]=829,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=815,[K1]=K35,x=23,y=106,},[2]={f=K3,[K17]=T8,[K16]={[1]=K39,},[K2]=816,[K1]=K15,x=31,y=106,},[3]={f=K3,[K2]=817,[K1]=K12,x=44,y=106,},},[K2]=818,[K1]=K4,x=52,y=106,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,[K2]=820,[K1]=K11,x=54,y=106,},},[K2]=819,[K1]=K4,x=52,y=106,},[K2]=814,[K1]=K7,x=14,y=106,},[2]={args={f=K3,[K4]={[1]={f=K3,[K2]=822,[K1]=K35,x=23,y=107,},[2]={f=K3,[K2]=823,[K1]=K12,x=31,y=107,},[3]={f=K3,[K2]=824,[K1]=K12,x=40,y=107,},},[K2]=825,[K1]=K4,x=48,y=107,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,[K2]=827,[K1]=K11,x=50,y=107,},[2]={f=K3,[K2]=828,[K1]="any",x=58,y=107,},},[K2]=826,[K1]=K4,x=48,y=107,},[K2]=821,[K1]=K7,x=14,y=107,},[3]={args={f=K3,[K4]={[1]={f=K3,[K17]=T8,[K16]={[1]=K39,},[K2]=831,[K1]=K15,x=23,y=108,},[2]={f=K3,[K2]=832,[K1]=K12,x=36,y=108,},},[K2]=833,[K1]=K4,x=44,y=108,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=835,[K1]=K11,x=46,y=108,},},[K2]=834,[K1]=K4,x=44,y=108,},[K2]=830,[K1]=K7,x=14,y=108,},[4]={args={f=K3,[K4]={[1]={f=K3,[K2]=837,[K1]=K12,x=23,y=109,},[2]={f=K3,[K2]=838,[K1]=K12,x=32,y=109,},},[K2]=839,[K1]=K4,x=40,y=109,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=841,[K1]=K11,x=42,y=109,},[2]={f=K3,[K2]=842,[K1]="any",x=50,y=109,},},[K2]=840,[K1]=K4,x=40,y=109,},[K2]=836,[K1]=K7,x=14,y=109,},},x=12,y=107,},[K[15]]={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={f=K3,[K2]=2220,[K1]=K5,[K5]="T@24",x=30,y=111,},},[K2]=2221,[K1]=K4,x=32,y=111,},f=K3,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K17]=T12,[K16]={[1]=K41,},[K2]=2223,[K1]=K15,[K40]={[1]={f=K3,[K2]=2222,[K1]=K5,[K5]="T@24",x=44,y=111,},},x=34,y=111,},},[K2]=2224,[K1]=K4,x=32,y=111,},[K2]=2225,[K1]=K7,x=18,y=111,},[K19]={[1]={f=K3,[K10]="T@24",[K2]=2219,[K1]=K10,x=27,y=111,},},[K2]=2226,[K1]=K20,x=4,y=112,},["getregistry"]={args={f=K3,[K4]={},[K2]=852,[K1]=K4,x=27,y=112,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,keys={f=K3,[K2]=854,[K1]="any",x=30,y=112,},[K2]=855,[K1]="map",[K29]={f=K3,[K2]=856,[K1]="any",x=34,y=112,},x=29,xend=37,y=112,yend=112,},},[K2]=853,[K1]=K4,x=27,y=112,},[K2]=851,[K1]=K7,x=17,y=112,},["getupvalue"]={args={f=K3,[K4]={[1]={f=K3,[K17]=T8,[K16]={[1]=K39,},[K2]=858,[K1]=K15,x=25,y=113,},[2]={f=K3,[K2]=859,[K1]=K12,x=38,y=113,},},[K2]=860,[K1]=K4,x=46,y=113,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=862,[K1]="any",x=48,y=113,},},[K2]=861,[K1]=K4,x=46,y=113,},[K2]=857,[K1]=K7,x=16,y=113,},["getuservalue"]={args={f=K3,[K4]={[1]={f=K3,[K17]=T14,[K16]={[1]=K[17],},[K2]=864,[K1]=K15,x=27,y=114,},[2]={f=K3,[K2]=865,[K1]=K12,x=37,y=114,},},[K2]=866,[K1]=K4,x=45,y=114,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=868,[K1]="any",x=47,y=114,},},[K2]=867,[K1]=K4,x=45,y=114,},[K2]=863,[K1]=K7,x=18,y=114,},["sethook"]={f=K3,[K2]=882,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=870,[K1]=K35,x=22,y=116,},[2]={f=K3,[K17]=T11,[K16]={[1]=K46,},[K2]=871,[K1]=K15,x=30,y=116,},[3]={f=K3,[K2]=872,[K1]=K11,x=44,y=116,},[4]={f=K3,[K2]=873,[K1]=K12,x=54,y=116,},},[K2]=874,[K1]=K4,x=4,y=117,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={},[K2]=875,[K1]=K4,x=61,y=116,},[K2]=869,[K1]=K7,x=13,y=116,},[2]={args={f=K3,[K4]={[1]={f=K3,[K17]=T11,[K16]={[1]=K46,},[K2]=877,[K1]=K15,x=22,y=117,},[2]={f=K3,[K2]=878,[K1]=K11,x=36,y=117,},[3]={f=K3,[K2]=879,[K1]=K12,x=46,y=117,},},[K2]=880,[K1]=K4,x=4,y=119,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={},[K2]=881,[K1]=K4,x=53,y=117,},[K2]=876,[K1]=K7,x=13,y=117,},},x=11,y=117,},["setlocal"]={f=K3,[K2]=898,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=884,[K1]=K35,x=23,y=119,},[2]={f=K3,[K2]=885,[K1]=K12,x=31,y=119,},[3]={f=K3,[K2]=886,[K1]=K12,x=40,y=119,},[4]={f=K3,[K2]=887,[K1]="any",x=49,y=119,},},[K2]=888,[K1]=K4,x=53,y=119,},f=K3,[K9]=false,[K6]=4,rets={f=K3,[K4]={[1]={f=K3,[K2]=890,[K1]=K11,x=55,y=119,},},[K2]=889,[K1]=K4,x=53,y=119,},[K2]=883,[K1]=K7,x=14,y=119,},[2]={args={f=K3,[K4]={[1]={f=K3,[K2]=892,[K1]=K12,x=23,y=120,},[2]={f=K3,[K2]=893,[K1]=K12,x=32,y=120,},[3]={f=K3,[K2]=894,[K1]="any",x=41,y=120,},},[K2]=895,[K1]=K4,x=45,y=120,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,[K2]=897,[K1]=K11,x=47,y=120,},},[K2]=896,[K1]=K4,x=45,y=120,},[K2]=891,[K1]=K7,x=14,y=120,},},x=12,y=120,},[K[16]]={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={f=K3,[K2]=2232,[K1]=K5,[K5]="T@25",x=30,y=122,},[2]={f=K3,[K17]=T12,[K16]={[1]=K41,},[K2]=2234,[K1]=K15,[K40]={[1]={f=K3,[K2]=2233,[K1]=K5,[K5]="T@25",x=43,y=122,},},x=33,y=122,},},[K2]=2235,[K1]=K4,x=46,y=122,},f=K3,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=2236,[K1]=K5,[K5]="T@25",x=48,y=122,},},[K2]=2237,[K1]=K4,x=46,y=122,},[K2]=2238,[K1]=K7,x=18,y=122,},[K19]={[1]={f=K3,[K10]="T@25",[K2]=2231,[K1]=K10,x=27,y=122,},},[K2]=2239,[K1]=K20,x=4,y=123,},["setupvalue"]={args={f=K3,[K4]={[1]={f=K3,[K17]=T8,[K16]={[1]=K39,},[K2]=909,[K1]=K15,x=25,y=123,},[2]={f=K3,[K2]=910,[K1]=K12,x=38,y=123,},[3]={f=K3,[K2]=911,[K1]="any",x=47,y=123,},},[K2]=912,[K1]=K4,x=51,y=123,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,[K2]=914,[K1]=K11,x=53,y=123,},},[K2]=913,[K1]=K4,x=51,y=123,},[K2]=908,[K1]=K7,x=16,y=123,},["setuservalue"]={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={f=K3,[K2]=2244,[K1]=K5,[K5]="U@26",x=30,y=124,},[2]={f=K3,[K2]=918,[K1]="any",x=33,y=124,},[3]={f=K3,[K2]=919,[K1]=K12,x=38,y=124,},},[K2]=2245,[K1]=K4,x=46,y=124,},f=K3,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,[K2]=2246,[K1]=K5,[K5]="U@26",x=48,y=124,},},[K2]=2247,[K1]=K4,x=46,y=124,},[K2]=2248,[K1]=K7,x=18,y=124,},[K19]={[1]={f=K3,[K10]="U@26",[K2]=2243,[K1]=K10,x=27,y=124,},},[K2]=2249,[K1]=K20,x=4,y=126,},["traceback"]={f=K3,[K2]=937,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=925,[K1]=K35,x=24,y=126,},[2]={f=K3,[K2]=926,[K1]=K11,x=34,y=126,},[3]={f=K3,[K2]=927,[K1]=K12,x=44,y=126,},},[K2]=928,[K1]=K4,x=52,y=126,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=930,[K1]=K11,x=54,y=126,},},[K2]=929,[K1]=K4,x=52,y=126,},[K2]=924,[K1]=K7,x=15,y=126,},[2]={args={f=K3,[K4]={[1]={f=K3,[K2]=932,[K1]=K11,x=26,y=127,},[2]={f=K3,[K2]=933,[K1]=K12,x=36,y=127,},},[K2]=934,[K1]=K4,x=44,y=127,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=936,[K1]=K11,x=46,y=127,},},[K2]=935,[K1]=K4,x=44,y=127,},[K2]=931,[K1]=K7,x=15,y=127,},},x=13,y=127,},["upvalueid"]={args={f=K3,[K4]={[1]={f=K3,[K17]=T8,[K16]={[1]=K39,},[K2]=939,[K1]=K15,x=24,y=129,},[2]={f=K3,[K2]=940,[K1]=K12,x=37,y=129,},},[K2]=941,[K1]=K4,x=45,y=129,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K17]=T14,[K16]={[1]=K[17],},[K2]=943,[K1]=K15,x=47,y=129,},},[K2]=942,[K1]=K4,x=45,y=129,},[K2]=938,[K1]=K7,x=15,y=129,},["upvaluejoin"]={args={f=K3,[K4]={[1]={f=K3,[K17]=T8,[K16]={[1]=K39,},[K2]=945,[K1]=K15,x=26,y=130,},[2]={f=K3,[K2]=946,[K1]=K12,x=39,y=130,},[3]={f=K3,[K17]=T8,[K16]={[1]=K39,},[K2]=947,[K1]=K15,x=48,y=130,},[4]={f=K3,[K2]=948,[K1]=K12,x=61,y=130,},},[K2]=949,[K1]=K4,x=1,y=131,},f=K3,[K9]=false,[K6]=4,rets={f=K3,[K4]={},[K2]=950,[K1]=K4,x=68,y=130,},[K2]=944,[K1]=K7,x=17,y=130,},},[K2]=747,[K1]=K37,x=1,y=73,},f=K3,[K2]=951,[K1]=K21,x=1,y=73,},},["dofile"]={[K27]=K26,t={args={f=K3,[K4]={[1]={f=K3,[K2]=1898,[K1]=K11,x=23,y=370,},},[K2]=1899,[K1]=K4,x=30,y=370,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1901,[K1]="any",x=32,y=370,},},[K2]=1900,[K1]=K4,x=30,y=370,},[K2]=1897,[K1]=K7,x=12,y=370,},},["error"]={[K27]=K26,t={args={f=K3,[K4]={[1]={f=K3,[K2]=1903,[K1]="any",x=22,y=372,},[2]={f=K3,[K2]=1904,[K1]=K12,x=29,y=372,},},[K2]=1905,[K1]=K4,x=4,y=373,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={},[K2]=1906,[K1]=K4,x=36,y=372,},[K2]=1902,[K1]=K7,x=11,y=372,},},[K[15]]={[K27]=K26,t={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={f=K3,[K2]=2506,[K1]=K5,[K5]="T@44",x=30,y=373,},},[K2]=2507,[K1]=K4,x=32,y=373,},f=K3,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K17]=T12,[K16]={[1]=K41,},[K2]=2509,[K1]=K15,[K40]={[1]={f=K3,[K2]=2508,[K1]=K5,[K5]="T@44",x=44,y=373,},},x=34,y=373,},},[K2]=2510,[K1]=K4,x=32,y=373,},[K2]=2511,[K1]=K7,x=18,y=373,},[K19]={[1]={f=K3,[K10]="T@44",[K2]=2505,[K1]=K10,x=27,y=373,},},[K2]=2512,[K1]=K20,x=4,y=374,},},io={[K28]=true,t={[K34]=true,def={[K24]="io",f=K3,[K31]={[1]=K45,[2]=K42,[3]="input",[4]=K[4],[5]=K[5],[6]="open",[7]="output",[8]="popen",[9]="read",[10]="stderr",[11]="stdin",[12]="stdout",[13]="tmpfile",[14]="type",[15]=K[6],},[K30]={[K45]=T15,[K42]={args={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=956,[K1]=K15,x=22,y=141,},},[K2]=957,[K1]=K4,x=4,y=142,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={},[K2]=958,[K1]=K4,x=26,y=141,},[K2]=955,[K1]=K7,x=11,y=141,},[K[4]]={args={f=K3,[K4]={},[K2]=967,[K1]=K4,x=4,y=145,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={},[K2]=968,[K1]=K4,x=20,y=143,},[K2]=966,[K1]=K7,x=11,y=143,},["input"]={args={f=K3,[K4]={[1]={f=K3,[K2]=961,[K1]=K22,[K18]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=960,[K1]=K15,x=22,y=142,},[2]={f=K3,[K2]=962,[K1]=K11,x=29,y=142,},},x=22,y=142,},},[K2]=963,[K1]=K4,x=36,y=142,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=965,[K1]=K15,x=38,y=142,},},[K2]=964,[K1]=K4,x=36,y=142,},[K2]=959,[K1]=K7,x=11,y=142,},[K[5]]={f=K3,[K2]=986,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=970,[K1]=K11,x=22,y=145,},},[K2]=971,[K1]=K4,x=29,y=145,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=974,[K1]=K4,x=42,y=145,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=976,[K1]=K11,x=45,y=145,},},[K2]=975,[K1]=K4,x=42,y=145,},[K2]=973,[K1]=K7,x=32,y=145,},},[K2]=972,[K1]=K4,x=29,y=145,},[K2]=969,[K1]=K7,x=11,y=145,},[2]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=978,[K1]=K11,x=22,y=146,},[2]={f=K3,[K17]=T4,[K16]={[1]=K48,},[K2]=979,[K1]=K15,x=30,y=146,},},[K2]=980,[K1]=K4,x=48,y=146,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=983,[K1]=K4,x=61,y=146,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=985,[K1]=K13,x=64,y=146,},},[K2]=984,[K1]=K4,x=61,y=146,},[K2]=982,[K1]=K7,x=51,y=146,},},[K2]=981,[K1]=K4,x=48,y=146,},[K2]=977,[K1]=K7,x=11,y=146,},[3]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=988,[K1]=K11,x=22,y=147,},[2]={f=K3,[K2]=990,[K1]=K22,[K18]={[1]={f=K3,[K2]=989,[K1]=K13,x=31,y=147,},[2]={f=K3,[K17]=T5,[K16]={[1]=K47,},[K2]=991,[K1]=K15,x=40,y=147,},},x=31,y=147,},},[K2]=992,[K1]=K4,x=59,y=147,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=995,[K1]=K4,x=72,y=147,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=997,[K1]=K11,x=75,y=147,},},[K2]=996,[K1]=K4,x=72,y=147,},[K2]=994,[K1]=K7,x=62,y=147,},},[K2]=993,[K1]=K4,x=59,y=147,},[K2]=987,[K1]=K7,x=11,y=147,},[4]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=999,[K1]=K11,x=22,y=148,},[2]={f=K3,[K2]=1001,[K1]=K22,[K18]={[1]={f=K3,[K2]=1000,[K1]=K13,x=31,y=148,},[2]={f=K3,[K17]=T6,[K16]={[1]=K49,},[K2]=1002,[K1]=K15,x=40,y=148,},},x=31,y=148,},},[K2]=1003,[K1]=K4,x=53,y=148,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=1006,[K1]=K4,x=66,y=148,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1009,[K1]=K22,[K18]={[1]={f=K3,[K2]=1008,[K1]=K11,x=70,y=148,},[2]={f=K3,[K2]=1010,[K1]=K13,x=79,y=148,},},x=70,y=148,},},[K2]=1007,[K1]=K4,x=66,y=148,},[K2]=1005,[K1]=K7,x=56,y=148,},},[K2]=1004,[K1]=K4,x=53,y=148,},[K2]=998,[K1]=K7,x=11,y=148,},[5]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1012,[K1]=K11,x=22,y=149,},[2]={f=K3,[K2]=1014,[K1]=K22,[K18]={[1]={f=K3,[K2]=1013,[K1]=K13,x=31,y=149,},[2]={f=K3,[K2]=1015,[K1]=K11,x=40,y=149,},},x=31,y=149,},},[K2]=1016,[K1]=K4,x=51,y=149,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=1019,[K1]=K4,x=64,y=149,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1021,[K1]=K11,x=67,y=149,},},[K2]=1020,[K1]=K4,x=64,y=149,},[K2]=1018,[K1]=K7,x=54,y=149,},},[K2]=1017,[K1]=K4,x=51,y=149,},[K2]=1011,[K1]=K7,x=11,y=149,},},x=9,y=146,},open={args={f=K3,[K4]={[1]={f=K3,[K2]=1023,[K1]=K11,x=19,y=151,},[2]={f=K3,[K17]=T15,[K16]={[1]=K45,},[K2]=1024,[K1]=K15,x=29,y=151,},},[K2]=1025,[K1]=K4,x=38,y=151,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=1027,[K1]=K15,x=40,y=151,},[2]={f=K3,[K2]=1028,[K1]=K11,x=46,y=151,},[3]={f=K3,[K2]=1029,[K1]=K12,x=54,y=151,},},[K2]=1026,[K1]=K4,x=38,y=151,},[K2]=1022,[K1]=K7,x=10,y=151,},["output"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1032,[K1]=K22,[K18]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=1031,[K1]=K15,x=23,y=152,},[2]={f=K3,[K2]=1033,[K1]=K11,x=30,y=152,},},x=23,y=152,},},[K2]=1034,[K1]=K4,x=37,y=152,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=1036,[K1]=K15,x=39,y=152,},},[K2]=1035,[K1]=K4,x=37,y=152,},[K2]=1030,[K1]=K7,x=12,y=152,},["popen"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1038,[K1]=K11,x=20,y=153,},[2]={f=K3,[K17]=T15,[K16]={[1]=K45,},[K2]=1039,[K1]=K15,x=30,y=153,},},[K2]=1040,[K1]=K4,x=39,y=153,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=1042,[K1]=K15,x=41,y=153,},[2]={f=K3,[K2]=1043,[K1]=K11,x=47,y=153,},},[K2]=1041,[K1]=K4,x=39,y=153,},[K2]=1037,[K1]=K7,x=11,y=153,},read={f=K3,[K2]=1053,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={},[K2]=1045,[K1]=K4,x=20,y=155,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1047,[K1]=K11,x=22,y=155,},},[K2]=1046,[K1]=K4,x=20,y=155,},[K2]=1044,[K1]=K7,x=10,y=155,},[2]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K17]=T4,[K16]={[1]=K48,},[K2]=1049,[K1]=K15,x=19,y=156,},},[K2]=1050,[K1]=K4,x=37,y=156,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1052,[K1]=K13,x=39,y=156,},},[K2]=1051,[K1]=K4,x=37,y=156,},[K2]=1048,[K1]=K7,x=10,y=156,},[3]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1056,[K1]=K22,[K18]={[1]={f=K3,[K2]=1055,[K1]=K13,x=20,y=157,},[2]={f=K3,[K17]=T5,[K16]={[1]=K47,},[K2]=1057,[K1]=K15,x=29,y=157,},},x=20,y=157,},},[K2]=1058,[K1]=K4,x=48,y=157,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1060,[K1]=K11,x=50,y=157,},},[K2]=1059,[K1]=K4,x=48,y=157,},[K2]=1054,[K1]=K7,x=10,y=157,},[4]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1063,[K1]=K22,[K18]={[1]={f=K3,[K2]=1062,[K1]=K13,x=20,y=158,},[2]={f=K3,[K17]=T6,[K16]={[1]=K49,},[K2]=1064,[K1]=K15,x=29,y=158,},},x=20,y=158,},},[K2]=1065,[K1]=K4,x=42,y=158,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1068,[K1]=K22,[K18]={[1]={f=K3,[K2]=1067,[K1]=K11,x=46,y=158,},[2]={f=K3,[K2]=1069,[K1]=K13,x=55,y=158,},},x=46,y=158,},},[K2]=1066,[K1]=K4,x=42,y=158,},[K2]=1061,[K1]=K7,x=10,y=158,},[5]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1072,[K1]=K22,[K18]={[1]={f=K3,[K2]=1071,[K1]=K13,x=20,y=159,},[2]={f=K3,[K2]=1073,[K1]=K11,x=29,y=159,},},x=20,y=159,},},[K2]=1074,[K1]=K4,x=40,y=159,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1076,[K1]=K11,x=43,y=159,},},[K2]=1075,[K1]=K4,x=40,y=159,},[K2]=1070,[K1]=K7,x=10,y=159,},},x=8,y=156,},["stderr"]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=1077,[K1]=K15,x=12,y=161,},["stdin"]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=1078,[K1]=K15,x=11,y=162,},["stdout"]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=1079,[K1]=K15,x=12,y=163,},["tmpfile"]={args={f=K3,[K4]={},[K2]=1081,[K1]=K4,x=23,y=164,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=1083,[K1]=K15,x=25,y=164,},},[K2]=1082,[K1]=K4,x=23,y=164,},[K2]=1080,[K1]=K7,x=13,y=164,},type={args={f=K3,[K4]={[1]={f=K3,[K2]=1085,[K1]="any",x=19,y=165,},},[K2]=1086,[K1]=K4,x=23,y=165,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1088,[K1]=K11,x=25,y=165,},},[K2]=1087,[K1]=K4,x=23,y=165,},[K2]=1084,[K1]=K7,x=10,y=165,},[K[6]]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1091,[K1]=K22,[K18]={[1]={f=K3,[K2]=1090,[K1]=K11,x=21,y=166,},[2]={f=K3,[K2]=1092,[K1]=K13,x=30,y=166,},},x=21,y=166,},},[K2]=1093,[K1]=K4,x=41,y=166,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=1095,[K1]=K15,x=43,y=166,},[2]={f=K3,[K2]=1096,[K1]=K11,x=49,y=166,},},[K2]=1094,[K1]=K4,x=41,y=166,},[K2]=1089,[K1]=K7,x=11,y=166,},},[K2]=952,[K1]=K37,x=1,y=133,},f=K3,[K2]=1097,[K1]=K21,x=1,y=133,},},["ipairs"]={[K27]=K26,[K28]=true,t={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={[K32]={f=K3,[K2]=2517,[K1]=K5,[K5]="A@45",x=25,y=374,},f=K3,[K2]=2518,[K1]=K33,x=24,y=374,},},[K2]=2519,[K1]=K4,x=28,y=374,},f=K3,[K6]=1,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=1922,[K1]=K4,x=41,y=374,},f=K3,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1924,[K1]=K12,x=43,y=374,},[2]={f=K3,[K2]=2520,[K1]=K5,[K5]="A@45",x=52,y=374,},},[K2]=2521,[K1]=K4,x=41,y=374,},[K2]=2522,[K1]=K7,x=31,y=374,},},[K2]=2523,[K1]=K4,x=28,y=374,},[K36]="ipairs",[K2]=2524,[K1]=K7,x=12,y=374,},[K19]={[1]={f=K3,[K10]="A@45",[K2]=2516,[K1]=K10,x=21,y=374,},},[K2]=2525,[K1]=K20,x=4,y=376,},},load={[K27]=K26,[K28]=true,t={f=K3,[K2]=1961,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=1929,[K1]=K22,[K18]={[1]={f=K3,[K2]=1928,[K1]=K11,x=20,y=376,},[2]={f=K3,[K17]=T16,[K16]={[1]="LoadFunction",},[K2]=1930,[K1]=K15,x=29,y=376,},},x=20,y=376,},[2]={f=K3,[K2]=1931,[K1]=K11,x=46,y=376,},[3]={f=K3,[K17]={def={[K24]="LoadMode",[K38]={b=true,bt=true,t=true,},f=K3,[K2]=1853,[K1]="enum",x=4,y=356,},f=K3,[K2]=1854,[K1]=K21,x=4,y=356,},[K16]={[1]="LoadMode",},[K2]=1932,[K1]=K15,x=56,y=376,},[4]={f=K3,keys={f=K3,[K2]=1934,[K1]="any",x=68,y=376,},[K2]=1933,[K1]="map",[K29]={f=K3,[K2]=1935,[K1]="any",x=68,y=376,},x=68,y=376,},},[K2]=1936,[K1]=K4,x=74,y=376,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1939,[K1]="any",x=85,y=376,},},[K2]=1940,[K1]=K4,x=85,y=376,},f=K3,[K44]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1941,[K1]="any",x=85,y=376,},},[K2]=1942,[K1]=K4,x=85,y=376,},[K2]=1938,[K1]=K7,x=77,y=376,},[2]={f=K3,[K2]=1943,[K1]=K11,x=87,y=376,},},[K2]=1937,[K1]=K4,x=74,y=376,},[K2]=1927,[K1]=K7,x=10,y=376,},[2]={args={f=K3,[K4]={[1]={f=K3,[K2]=1946,[K1]=K22,[K18]={[1]={f=K3,[K2]=1945,[K1]=K11,x=20,y=377,},[2]={f=K3,[K17]=T16,[K16]={[1]="LoadFunction",},[K2]=1947,[K1]=K15,x=29,y=377,},},x=20,y=377,},[2]={f=K3,[K2]=1948,[K1]=K11,x=46,y=377,},[3]={f=K3,[K2]=1949,[K1]=K11,x=56,y=377,},[4]={f=K3,keys={f=K3,[K2]=1951,[K1]="any",x=66,y=377,},[K2]=1950,[K1]="map",[K29]={f=K3,[K2]=1952,[K1]="any",x=66,y=377,},x=66,y=377,},},[K2]=1953,[K1]=K4,x=72,y=377,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1956,[K1]="any",x=83,y=377,},},[K2]=1957,[K1]=K4,x=83,y=377,},f=K3,[K44]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1958,[K1]="any",x=83,y=377,},},[K2]=1959,[K1]=K4,x=83,y=377,},[K2]=1955,[K1]=K7,x=75,y=377,},[2]={f=K3,[K2]=1960,[K1]=K11,x=85,y=377,},},[K2]=1954,[K1]=K4,x=72,y=377,},[K2]=1944,[K1]=K7,x=10,y=377,},},x=8,y=377,},},["loadfile"]={[K27]=K26,[K28]=true,t={args={f=K3,[K4]={[1]={f=K3,[K2]=1963,[K1]=K11,x=25,y=379,},[2]={f=K3,[K2]=1964,[K1]=K11,x=35,y=379,},[3]={f=K3,keys={f=K3,[K2]=1965,[K1]="any",x=46,y=379,},[K2]=1966,[K1]="map",[K29]={f=K3,[K2]=1967,[K1]="any",x=50,y=379,},x=45,xend=53,y=379,yend=379,},},[K2]=1968,[K1]=K4,x=55,y=379,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1971,[K1]="any",x=66,y=379,},},[K2]=1972,[K1]=K4,x=66,y=379,},f=K3,[K44]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1973,[K1]="any",x=66,y=379,},},[K2]=1974,[K1]=K4,x=66,y=379,},[K2]=1970,[K1]=K7,x=58,y=379,},[2]={f=K3,[K2]=1975,[K1]=K11,x=68,y=379,},},[K2]=1969,[K1]=K4,x=55,y=379,},[K2]=1962,[K1]=K7,x=14,y=379,},},math={[K28]=true,t={[K34]=true,def={[K24]="math",f=K3,[K31]={[1]=K[18],[2]="abs",[3]="acos",[4]="asin",[5]="atan",[6]="atan2",[7]="ceil",[8]="cos",[9]="cosh",[10]="deg",[11]="exp",[12]="floor",[13]="fmod",[14]="frexp",[15]="huge",[16]="ldexp",[17]="log",[18]="log10",[19]="max",[20]="maxinteger",[21]="min",[22]="mininteger",[23]="modf",[24]="pi",[25]="pow",[26]="rad",[27]="random",[28]="randomseed",[29]="sin",[30]="sinh",[31]="sqrt",[32]="tan",[33]="tanh",[34]="tointeger",[35]="type",[36]="ult",},[K30]={[K[18]]=T17,abs={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={[K[19]]=T18,f=K3,[K2]=2256,[K1]=K5,[K5]="N@27",x=32,y=172,},},[K2]=2257,[K1]=K4,x=34,y=172,},f=K3,[K6]=1,rets={f=K3,[K4]={[1]={[K[19]]=T18,f=K3,[K2]=2258,[K1]=K5,[K5]="N@27",x=36,y=172,},},[K2]=2259,[K1]=K4,x=34,y=172,},[K2]=2260,[K1]=K7,x=9,y=172,},[K19]={[1]={[K[19]]=T18,f=K3,[K10]="N@27",[K2]=2255,[K1]=K10,x=18,y=172,},},[K2]=2261,[K1]=K20,x=4,y=173,},acos={args={f=K3,[K4]={[1]={f=K3,[K2]=1112,[K1]=K13,x=19,y=173,},},[K2]=1113,[K1]=K4,x=26,y=173,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1115,[K1]=K13,x=28,y=173,},},[K2]=1114,[K1]=K4,x=26,y=173,},[K2]=1111,[K1]=K7,x=10,y=173,},asin={args={f=K3,[K4]={[1]={f=K3,[K2]=1117,[K1]=K13,x=19,y=174,},},[K2]=1118,[K1]=K4,x=26,y=174,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1120,[K1]=K13,x=28,y=174,},},[K2]=1119,[K1]=K4,x=26,y=174,},[K2]=1116,[K1]=K7,x=10,y=174,},atan={args={f=K3,[K4]={[1]={f=K3,[K2]=1122,[K1]=K13,x=19,y=175,},[2]={f=K3,[K2]=1123,[K1]=K13,x=29,y=175,},},[K2]=1124,[K1]=K4,x=36,y=175,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1126,[K1]=K13,x=38,y=175,},},[K2]=1125,[K1]=K4,x=36,y=175,},[K2]=1121,[K1]=K7,x=10,y=175,},["atan2"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1128,[K1]=K13,x=20,y=176,},[2]={f=K3,[K2]=1129,[K1]=K13,x=28,y=176,},},[K2]=1130,[K1]=K4,x=35,y=176,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1132,[K1]=K13,x=37,y=176,},},[K2]=1131,[K1]=K4,x=35,y=176,},[K2]=1127,[K1]=K7,x=11,y=176,},ceil={args={f=K3,[K4]={[1]={f=K3,[K2]=1134,[K1]=K13,x=19,y=177,},},[K2]=1135,[K1]=K4,x=26,y=177,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1137,[K1]=K12,x=28,y=177,},},[K2]=1136,[K1]=K4,x=26,y=177,},[K2]=1133,[K1]=K7,x=10,y=177,},cos={args={f=K3,[K4]={[1]={f=K3,[K2]=1139,[K1]=K13,x=18,y=178,},},[K2]=1140,[K1]=K4,x=25,y=178,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1142,[K1]=K13,x=27,y=178,},},[K2]=1141,[K1]=K4,x=25,y=178,},[K2]=1138,[K1]=K7,x=9,y=178,},cosh={args={f=K3,[K4]={[1]={f=K3,[K2]=1144,[K1]=K13,x=19,y=179,},},[K2]=1145,[K1]=K4,x=26,y=179,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1147,[K1]=K13,x=28,y=179,},},[K2]=1146,[K1]=K4,x=26,y=179,},[K2]=1143,[K1]=K7,x=10,y=179,},deg={args={f=K3,[K4]={[1]={f=K3,[K2]=1149,[K1]=K13,x=18,y=180,},},[K2]=1150,[K1]=K4,x=25,y=180,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1152,[K1]=K13,x=27,y=180,},},[K2]=1151,[K1]=K4,x=25,y=180,},[K2]=1148,[K1]=K7,x=9,y=180,},exp={args={f=K3,[K4]={[1]={f=K3,[K2]=1154,[K1]=K13,x=18,y=181,},},[K2]=1155,[K1]=K4,x=25,y=181,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1157,[K1]=K13,x=27,y=181,},},[K2]=1156,[K1]=K4,x=25,y=181,},[K2]=1153,[K1]=K7,x=9,y=181,},["floor"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1159,[K1]=K13,x=20,y=182,},},[K2]=1160,[K1]=K4,x=27,y=182,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1162,[K1]=K12,x=29,y=182,},},[K2]=1161,[K1]=K4,x=27,y=182,},[K2]=1158,[K1]=K7,x=11,y=182,},fmod={f=K3,[K2]=1175,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=1164,[K1]=K12,x=19,y=184,},[2]={f=K3,[K2]=1165,[K1]=K12,x=28,y=184,},},[K2]=1166,[K1]=K4,x=36,y=184,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1168,[K1]=K12,x=38,y=184,},},[K2]=1167,[K1]=K4,x=36,y=184,},[K2]=1163,[K1]=K7,x=10,y=184,},[2]={args={f=K3,[K4]={[1]={f=K3,[K2]=1170,[K1]=K13,x=19,y=185,},[2]={f=K3,[K2]=1171,[K1]=K13,x=27,y=185,},},[K2]=1172,[K1]=K4,x=34,y=185,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1174,[K1]=K13,x=36,y=185,},},[K2]=1173,[K1]=K4,x=34,y=185,},[K2]=1169,[K1]=K7,x=10,y=185,},},x=8,y=185,},["frexp"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1177,[K1]=K13,x=20,y=187,},},[K2]=1178,[K1]=K4,x=27,y=187,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1180,[K1]=K13,x=29,y=187,},[2]={f=K3,[K2]=1181,[K1]=K12,x=37,y=187,},},[K2]=1179,[K1]=K4,x=27,y=187,},[K2]=1176,[K1]=K7,x=11,y=187,},huge={f=K3,[K2]=1182,[K1]=K13,x=10,y=188,},["ldexp"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1184,[K1]=K13,x=20,y=189,},[2]={f=K3,[K2]=1185,[K1]=K12,x=28,y=189,},},[K2]=1186,[K1]=K4,x=36,y=189,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1188,[K1]=K13,x=38,y=189,},},[K2]=1187,[K1]=K4,x=36,y=189,},[K2]=1183,[K1]=K7,x=11,y=189,},log={args={f=K3,[K4]={[1]={f=K3,[K2]=1190,[K1]=K13,x=18,y=190,},[2]={f=K3,[K2]=1191,[K1]=K13,x=28,y=190,},},[K2]=1192,[K1]=K4,x=35,y=190,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1194,[K1]=K13,x=37,y=190,},},[K2]=1193,[K1]=K4,x=35,y=190,},[K2]=1189,[K1]=K7,x=9,y=190,},["log10"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1196,[K1]=K13,x=20,y=191,},},[K2]=1197,[K1]=K4,x=27,y=191,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1199,[K1]=K13,x=29,y=191,},},[K2]=1198,[K1]=K4,x=27,y=191,},[K2]=1195,[K1]=K7,x=11,y=191,},max={f=K3,[K2]=1212,[K1]="poly",[K18]={[1]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1201,[K1]=K12,x=18,y=193,},},[K2]=1202,[K1]=K4,x=29,y=193,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1204,[K1]=K12,x=31,y=193,},},[K2]=1203,[K1]=K4,x=29,y=193,},[K2]=1200,[K1]=K7,x=9,y=193,},[2]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1207,[K1]=K22,[K18]={[1]={f=K3,[K2]=1206,[K1]=K13,x=19,y=194,},[2]={f=K3,[K2]=1208,[K1]=K12,x=28,y=194,},},x=19,y=194,},},[K2]=1209,[K1]=K4,x=40,y=194,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1211,[K1]=K13,x=42,y=194,},},[K2]=1210,[K1]=K4,x=40,y=194,},[K2]=1205,[K1]=K7,x=9,y=194,},[3]={f=K3,t={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2263,[K1]=K5,[K5]="T",x=21,y=195,},},[K2]=1216,[K1]=K4,x=26,y=195,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=2264,[K1]=K5,[K5]="T",x=28,y=195,},},[K2]=1217,[K1]=K4,x=26,y=195,},[K2]=1213,[K1]=K7,x=9,y=195,},[K19]={[1]={f=K3,[K10]="T",[K2]=1214,[K1]=K10,x=18,y=195,},xend=19,yend=195,},[K2]=1219,[K1]=K20,x=4,y=196,},[4]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1221,[K1]="any",x=18,y=196,},},[K2]=1222,[K1]=K4,x=25,y=196,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1224,[K1]="any",x=27,y=196,},},[K2]=1223,[K1]=K4,x=25,y=196,},[K2]=1220,[K1]=K7,x=9,y=196,},},x=7,y=194,},["maxinteger"]={f=K3,[K28]=true,[K2]=1225,[K1]=K12,x=16,y=198,},min={f=K3,[K2]=1238,[K1]="poly",[K18]={[1]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1227,[K1]=K12,x=18,y=200,},},[K2]=1228,[K1]=K4,x=29,y=200,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1230,[K1]=K12,x=31,y=200,},},[K2]=1229,[K1]=K4,x=29,y=200,},[K2]=1226,[K1]=K7,x=9,y=200,},[2]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1233,[K1]=K22,[K18]={[1]={f=K3,[K2]=1232,[K1]=K13,x=19,y=201,},[2]={f=K3,[K2]=1234,[K1]=K12,x=28,y=201,},},x=19,y=201,},},[K2]=1235,[K1]=K4,x=40,y=201,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1237,[K1]=K13,x=42,y=201,},},[K2]=1236,[K1]=K4,x=40,y=201,},[K2]=1231,[K1]=K7,x=9,y=201,},[3]={f=K3,t={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2273,[K1]=K5,[K5]="T",x=21,y=202,},},[K2]=1242,[K1]=K4,x=26,y=202,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=2274,[K1]=K5,[K5]="T",x=28,y=202,},},[K2]=1243,[K1]=K4,x=26,y=202,},[K2]=1239,[K1]=K7,x=9,y=202,},[K19]={[1]={f=K3,[K10]="T",[K2]=1240,[K1]=K10,x=18,y=202,},xend=19,yend=202,},[K2]=1245,[K1]=K20,x=4,y=203,},[4]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1247,[K1]="any",x=18,y=203,},},[K2]=1248,[K1]=K4,x=25,y=203,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1250,[K1]="any",x=27,y=203,},},[K2]=1249,[K1]=K4,x=25,y=203,},[K2]=1246,[K1]=K7,x=9,y=203,},},x=7,y=201,},["mininteger"]={f=K3,[K28]=true,[K2]=1251,[K1]=K12,x=16,y=205,},modf={args={f=K3,[K4]={[1]={f=K3,[K2]=1253,[K1]=K13,x=19,y=207,},},[K2]=1254,[K1]=K4,x=26,y=207,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1256,[K1]=K12,x=28,y=207,},[2]={f=K3,[K2]=1257,[K1]=K13,x=37,y=207,},},[K2]=1255,[K1]=K4,x=26,y=207,},[K2]=1252,[K1]=K7,x=10,y=207,},pi={f=K3,[K2]=1258,[K1]=K13,x=8,y=208,},pow={args={f=K3,[K4]={[1]={f=K3,[K2]=1260,[K1]=K13,x=18,y=209,},[2]={f=K3,[K2]=1261,[K1]=K13,x=26,y=209,},},[K2]=1262,[K1]=K4,x=33,y=209,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1264,[K1]=K13,x=35,y=209,},},[K2]=1263,[K1]=K4,x=33,y=209,},[K2]=1259,[K1]=K7,x=9,y=209,},rad={args={f=K3,[K4]={[1]={f=K3,[K2]=1266,[K1]=K13,x=18,y=210,},},[K2]=1267,[K1]=K4,x=25,y=210,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1269,[K1]=K13,x=27,y=210,},},[K2]=1268,[K1]=K4,x=25,y=210,},[K2]=1265,[K1]=K7,x=9,y=210,},["random"]={f=K3,[K2]=1280,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=1271,[K1]=K12,x=21,y=212,},[2]={f=K3,[K2]=1272,[K1]=K12,x=32,y=212,},},[K2]=1273,[K1]=K4,x=40,y=212,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1275,[K1]=K12,x=42,y=212,},},[K2]=1274,[K1]=K4,x=40,y=212,},[K2]=1270,[K1]=K7,x=12,y=212,},[2]={args={f=K3,[K4]={},[K2]=1277,[K1]=K4,x=22,y=213,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1279,[K1]=K13,x=24,y=213,},},[K2]=1278,[K1]=K4,x=22,y=213,},[K2]=1276,[K1]=K7,x=12,y=213,},},x=10,y=213,},["randomseed"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1282,[K1]=K12,x=27,y=215,},[2]={f=K3,[K2]=1283,[K1]=K12,x=38,y=215,},},[K2]=1284,[K1]=K4,x=46,y=215,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1286,[K1]=K12,x=48,y=215,},[2]={f=K3,[K2]=1287,[K1]=K12,x=57,y=215,},},[K2]=1285,[K1]=K4,x=46,y=215,},[K2]=1281,[K1]=K7,x=16,y=215,},sin={args={f=K3,[K4]={[1]={f=K3,[K2]=1289,[K1]=K13,x=18,y=216,},},[K2]=1290,[K1]=K4,x=25,y=216,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1292,[K1]=K13,x=27,y=216,},},[K2]=1291,[K1]=K4,x=25,y=216,},[K2]=1288,[K1]=K7,x=9,y=216,},sinh={args={f=K3,[K4]={[1]={f=K3,[K2]=1294,[K1]=K13,x=19,y=217,},},[K2]=1295,[K1]=K4,x=26,y=217,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1297,[K1]=K13,x=28,y=217,},},[K2]=1296,[K1]=K4,x=26,y=217,},[K2]=1293,[K1]=K7,x=10,y=217,},sqrt={args={f=K3,[K4]={[1]={f=K3,[K2]=1299,[K1]=K13,x=19,y=218,},},[K2]=1300,[K1]=K4,x=26,y=218,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1302,[K1]=K13,x=28,y=218,},},[K2]=1301,[K1]=K4,x=26,y=218,},[K2]=1298,[K1]=K7,x=10,y=218,},tan={args={f=K3,[K4]={[1]={f=K3,[K2]=1304,[K1]=K13,x=18,y=219,},},[K2]=1305,[K1]=K4,x=25,y=219,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1307,[K1]=K13,x=27,y=219,},},[K2]=1306,[K1]=K4,x=25,y=219,},[K2]=1303,[K1]=K7,x=9,y=219,},tanh={args={f=K3,[K4]={[1]={f=K3,[K2]=1309,[K1]=K13,x=19,y=220,},},[K2]=1310,[K1]=K4,x=26,y=220,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1312,[K1]=K13,x=28,y=220,},},[K2]=1311,[K1]=K4,x=26,y=220,},[K2]=1308,[K1]=K7,x=10,y=220,},["tointeger"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1314,[K1]="any",x=24,y=221,},},[K2]=1315,[K1]=K4,x=28,y=221,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1317,[K1]=K12,x=30,y=221,},},[K2]=1316,[K1]=K4,x=28,y=221,},[K2]=1313,[K1]=K7,x=15,y=221,},type={args={f=K3,[K4]={[1]={f=K3,[K2]=1319,[K1]="any",x=19,y=222,},},[K2]=1320,[K1]=K4,x=23,y=222,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1322,[K1]=K11,x=25,y=222,},},[K2]=1321,[K1]=K4,x=23,y=222,},[K2]=1318,[K1]=K7,x=10,y=222,},ult={args={f=K3,[K4]={[1]={f=K3,[K2]=1324,[K1]=K13,x=18,y=223,},[2]={f=K3,[K2]=1325,[K1]=K13,x=26,y=223,},},[K2]=1326,[K1]=K4,x=33,y=223,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1328,[K1]=K25,x=35,y=223,},},[K2]=1327,[K1]=K4,x=33,y=223,},[K2]=1323,[K1]=K7,x=9,y=223,},},[K2]=1098,[K1]=K37,x=1,y=169,},f=K3,[K2]=1329,[K1]=K21,x=1,y=169,},},[K41]={t=T12,[K[10]]=true,},next={[K27]=K26,t={f=K3,[K2]=1998,[K1]="poly",[K18]={[1]={f=K3,t={args={f=K3,[K4]={[1]={f=K3,keys={f=K3,[K2]=2528,[K1]=K5,[K5]="K",x=26,y=381,},[K2]=1980,[K1]="map",[K29]={f=K3,[K2]=2529,[K1]=K5,[K5]="V",x=28,y=381,},x=25,xend=29,y=381,yend=381,},[2]={f=K3,[K2]=2530,[K1]=K5,[K5]="K",x=34,y=381,},},[K2]=1983,[K1]=K4,x=36,y=381,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2531,[K1]=K5,[K5]="K",x=39,y=381,},[2]={f=K3,[K2]=2532,[K1]=K5,[K5]="V",x=42,y=381,},},[K2]=1984,[K1]=K4,x=36,y=381,},[K2]=1976,[K1]=K7,x=10,y=381,},[K19]={[1]={f=K3,[K10]="K",[K2]=1977,[K1]=K10,x=19,y=381,},[2]={f=K3,[K10]="V",[K2]=1978,[K1]=K10,x=22,y=381,},xend=23,yend=381,},[K2]=1987,[K1]=K20,x=4,y=382,},[2]={f=K3,t={args={f=K3,[K4]={[1]={[K32]={f=K3,[K2]=2546,[K1]=K5,[K5]="A",x=23,y=382,},f=K3,[K2]=1991,[K1]=K33,x=22,xend=24,y=382,yend=382,},[2]={f=K3,[K2]=1992,[K1]=K12,x=29,y=382,},},[K2]=1993,[K1]=K4,x=37,y=382,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1995,[K1]=K12,x=40,y=382,},[2]={f=K3,[K2]=2547,[K1]=K5,[K5]="A",x=49,y=382,},},[K2]=1994,[K1]=K4,x=37,y=382,},[K2]=1988,[K1]=K7,x=10,y=382,},[K19]={[1]={f=K3,[K10]="A",[K2]=1989,[K1]=K10,x=19,y=382,},xend=20,yend=382,},[K2]=1997,[K1]=K20,x=4,y=384,},},x=8,y=382,},},os={[K28]=true,t={[K34]=true,def={[K24]="os",f=K3,[K31]={[1]=K50,[2]=K[20],[3]="clock",[4]="date",[5]="difftime",[6]="execute",[7]="exit",[8]="getenv",[9]=K[21],[10]="rename",[11]="setlocale",[12]="time",[13]="tmpname",},[K30]={[K[20]]=T19,[K50]=T20,["clock"]={args={f=K3,[K4]={},[K2]=1345,[K1]=K4,x=21,y=243,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1347,[K1]=K13,x=23,y=243,},},[K2]=1346,[K1]=K4,x=21,y=243,},[K2]=1344,[K1]=K7,x=11,y=243,},date={f=K3,[K2]=1360,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K17]=T19,[K16]={[1]=K[20],},[K2]=1349,[K1]=K15,x=19,y=245,},[2]={f=K3,[K2]=1350,[K1]=K13,x=31,y=245,},},[K2]=1351,[K1]=K4,x=38,y=245,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K17]=T20,[K16]={[1]=K50,},[K2]=1353,[K1]=K15,x=40,y=245,},},[K2]=1352,[K1]=K4,x=38,y=245,},[K2]=1348,[K1]=K7,x=10,y=245,},[2]={args={f=K3,[K4]={[1]={f=K3,[K2]=1355,[K1]=K11,x=21,y=246,},[2]={f=K3,[K2]=1356,[K1]=K13,x=31,y=246,},},[K2]=1357,[K1]=K4,x=38,y=246,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1359,[K1]=K11,x=40,y=246,},},[K2]=1358,[K1]=K4,x=38,y=246,},[K2]=1354,[K1]=K7,x=10,y=246,},},x=8,y=246,},["difftime"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1362,[K1]=K12,x=23,y=248,},[2]={f=K3,[K2]=1363,[K1]=K12,x=32,y=248,},},[K2]=1364,[K1]=K4,x=40,y=248,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1366,[K1]=K13,x=42,y=248,},},[K2]=1365,[K1]=K4,x=40,y=248,},[K2]=1361,[K1]=K7,x=14,y=248,},["execute"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1368,[K1]=K11,x=22,y=249,},},[K2]=1369,[K1]=K4,x=29,y=249,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1371,[K1]=K25,x=31,y=249,},[2]={f=K3,[K2]=1372,[K1]=K11,x=40,y=249,},[3]={f=K3,[K2]=1373,[K1]=K12,x=48,y=249,},},[K2]=1370,[K1]=K4,x=29,y=249,},[K2]=1367,[K1]=K7,x=13,y=249,},exit={args={f=K3,[K4]={[1]={f=K3,[K2]=1376,[K1]=K22,[K18]={[1]={f=K3,[K2]=1375,[K1]=K12,x=22,y=250,},[2]={f=K3,[K2]=1377,[K1]=K25,x=32,y=250,},},x=22,y=250,},[2]={f=K3,[K2]=1378,[K1]=K25,x=44,y=250,},},[K2]=1379,[K1]=K4,x=4,y=251,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={},[K2]=1380,[K1]=K4,x=51,y=250,},[K2]=1374,[K1]=K7,x=10,y=250,},["getenv"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1382,[K1]=K11,x=21,y=251,},},[K2]=1383,[K1]=K4,x=28,y=251,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1385,[K1]=K11,x=30,y=251,},},[K2]=1384,[K1]=K4,x=28,y=251,},[K2]=1381,[K1]=K7,x=12,y=251,},[K[21]]={args={f=K3,[K4]={[1]={f=K3,[K2]=1387,[K1]=K11,x=21,y=252,},},[K2]=1388,[K1]=K4,x=28,y=252,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1390,[K1]=K25,x=30,y=252,},[2]={f=K3,[K2]=1391,[K1]=K11,x=39,y=252,},},[K2]=1389,[K1]=K4,x=28,y=252,},[K2]=1386,[K1]=K7,x=12,y=252,},["rename"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1393,[K1]=K11,x=21,y=253,},[2]={f=K3,[K2]=1394,[K1]=K11,x=29,y=253,},},[K2]=1395,[K1]=K4,x=36,y=253,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1397,[K1]=K25,x=38,y=253,},[2]={f=K3,[K2]=1398,[K1]=K11,x=47,y=253,},},[K2]=1396,[K1]=K4,x=36,y=253,},[K2]=1392,[K1]=K7,x=12,y=253,},["setlocale"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1400,[K1]=K11,x=24,y=254,},[2]={f=K3,[K2]=1401,[K1]=K11,x=34,y=254,},},[K2]=1402,[K1]=K4,x=41,y=254,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1404,[K1]=K11,x=43,y=254,},},[K2]=1403,[K1]=K4,x=41,y=254,},[K2]=1399,[K1]=K7,x=15,y=254,},time={args={f=K3,[K4]={[1]={f=K3,[K17]=T20,[K16]={[1]=K50,},[K2]=1406,[K1]=K15,x=21,y=255,},},[K2]=1407,[K1]=K4,x=31,y=255,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1409,[K1]=K12,x=33,y=255,},},[K2]=1408,[K1]=K4,x=31,y=255,},[K2]=1405,[K1]=K7,x=10,y=255,},["tmpname"]={args={f=K3,[K4]={},[K2]=1411,[K1]=K4,x=23,y=256,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1413,[K1]=K11,x=25,y=256,},},[K2]=1412,[K1]=K4,x=23,y=256,},[K2]=1410,[K1]=K7,x=13,y=256,},},[K2]=1330,[K1]=K37,x=1,y=226,},f=K3,[K2]=1414,[K1]=K21,x=1,y=226,},},["package"]={[K28]=true,t={[K34]=true,def={[K24]="package",f=K3,[K31]={[1]="config",[2]="cpath",[3]="loaded",[4]="loadlib",[5]="loaders",[6]="path",[7]="preload",[8]="searchers",[9]="searchpath",},[K30]={["config"]={f=K3,[K2]=1416,[K1]=K11,x=12,y=260,},["cpath"]={f=K3,[K2]=1417,[K1]=K11,x=11,y=261,},["loaded"]={f=K3,keys={f=K3,[K2]=1418,[K1]=K11,x=13,y=262,},[K2]=1419,[K1]="map",[K29]={f=K3,[K2]=1420,[K1]="any",x=20,y=262,},x=12,xend=23,y=262,yend=262,},["loaders"]={[K32]={args={f=K3,[K4]={[1]={f=K3,[K2]=1432,[K1]=K11,x=24,y=264,},},[K2]=1433,[K1]=K4,x=31,y=264,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1435,[K1]="any",x=33,y=264,},[2]={f=K3,[K2]=1436,[K1]="any",x=38,y=264,},},[K2]=1434,[K1]=K4,x=31,y=264,},[K2]=1431,[K1]=K7,x=15,y=264,},f=K3,[K2]=1437,[K1]=K33,x=13,xend=42,y=264,yend=264,},["loadlib"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1422,[K1]=K11,x=22,y=263,},[2]={f=K3,[K2]=1423,[K1]=K11,x=30,y=263,},},[K2]=1424,[K1]=K4,x=37,y=263,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1427,[K1]="any",x=48,y=263,},},[K2]=1428,[K1]=K4,x=48,y=263,},f=K3,[K44]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1429,[K1]="any",x=48,y=263,},},[K2]=1430,[K1]=K4,x=48,y=263,},[K2]=1426,[K1]=K7,x=40,y=263,},},[K2]=1425,[K1]=K4,x=37,y=263,},[K2]=1421,[K1]=K7,x=13,y=263,},path={f=K3,[K2]=1438,[K1]=K11,x=10,y=265,},["preload"]={f=K3,keys={f=K3,[K2]=1439,[K1]="any",x=14,y=266,},[K2]=1440,[K1]="map",[K29]={f=K3,[K2]=1441,[K1]="any",x=18,y=266,},x=13,xend=21,y=266,yend=266,},["searchers"]={[K32]={args={f=K3,[K4]={[1]={f=K3,[K2]=1443,[K1]=K11,x=26,y=267,},},[K2]=1444,[K1]=K4,x=33,y=267,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1446,[K1]="any",x=35,y=267,},},[K2]=1445,[K1]=K4,x=33,y=267,},[K2]=1442,[K1]=K7,x=17,y=267,},f=K3,[K2]=1447,[K1]=K33,x=15,xend=39,y=267,yend=267,},["searchpath"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1449,[K1]=K11,x=25,y=268,},[2]={f=K3,[K2]=1450,[K1]=K11,x=33,y=268,},[3]={f=K3,[K2]=1451,[K1]=K11,x=43,y=268,},[4]={f=K3,[K2]=1452,[K1]=K11,x=53,y=268,},},[K2]=1453,[K1]=K4,x=60,y=268,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1455,[K1]=K11,x=62,y=268,},[2]={f=K3,[K2]=1456,[K1]=K11,x=70,y=268,},},[K2]=1454,[K1]=K4,x=60,y=268,},[K2]=1448,[K1]=K7,x=16,y=268,},},[K2]=1415,[K1]=K37,x=1,y=259,},f=K3,[K2]=1457,[K1]=K21,x=1,y=259,},},["pairs"]={[K27]=K26,[K28]=true,t={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={f=K3,keys={f=K3,[K2]=2564,[K1]=K5,[K5]="K@48",x=27,y=384,},[K2]=2566,[K1]="map",[K29]={f=K3,[K2]=2565,[K1]=K5,[K5]="V@48",x=29,y=384,},x=26,y=384,},},[K2]=2567,[K1]=K4,x=32,y=384,},f=K3,[K6]=1,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=2008,[K1]=K4,x=45,y=384,},f=K3,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=2568,[K1]=K5,[K5]="K@48",x=47,y=384,},[2]={f=K3,[K2]=2569,[K1]=K5,[K5]="V@48",x=50,y=384,},},[K2]=2570,[K1]=K4,x=45,y=384,},[K2]=2571,[K1]=K7,x=35,y=384,},},[K2]=2572,[K1]=K4,x=32,y=384,},[K36]="pairs",[K2]=2573,[K1]=K7,x=11,y=384,},[K19]={[1]={f=K3,[K10]="K@48",[K2]=2562,[K1]=K10,x=20,y=384,},[2]={f=K3,[K10]="V@48",[K2]=2563,[K1]=K10,x=23,y=384,},},[K2]=2574,[K1]=K20,x=4,y=385,},},["pcall"]={[K27]=K26,[K28]=true,t={args={f=K3,[K14]=true,[K4]={[1]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2015,[K1]="any",x=29,y=385,},},[K2]=2016,[K1]=K4,x=36,y=385,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2018,[K1]="any",x=38,y=385,},},[K2]=2017,[K1]=K4,x=36,y=385,},[K2]=2014,[K1]=K7,x=20,y=385,},[2]={f=K3,[K2]=2019,[K1]="any",x=47,y=385,},},[K2]=2020,[K1]=K4,x=54,y=385,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2022,[K1]=K25,x=56,y=385,},[2]={f=K3,[K2]=2023,[K1]="any",x=65,y=385,},},[K2]=2021,[K1]=K4,x=54,y=385,},[K36]="pcall",[K2]=2013,[K1]=K7,x=11,y=385,},},["print"]={[K27]=K26,t={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2025,[K1]="any",x=20,y=386,},},[K2]=2026,[K1]=K4,x=4,y=387,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={},[K2]=2027,[K1]=K4,x=26,y=386,},[K2]=2024,[K1]=K7,x=11,y=386,},},["rawequal"]={[K27]=K26,t={args={f=K3,[K4]={[1]={f=K3,[K2]=2029,[K1]="any",x=23,y=387,},[2]={f=K3,[K2]=2030,[K1]="any",x=28,y=387,},},[K2]=2031,[K1]=K4,x=32,y=387,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=2033,[K1]=K25,x=34,y=387,},},[K2]=2032,[K1]=K4,x=32,y=387,},[K2]=2028,[K1]=K7,x=14,y=387,},},["rawget"]={[K27]=K26,t={f=K3,[K2]=2053,[K1]="poly",[K18]={[1]={f=K3,t={args={f=K3,[K4]={[1]={f=K3,keys={f=K3,[K2]=2577,[K1]=K5,[K5]="K",x=28,y=389,},[K2]=2038,[K1]="map",[K29]={f=K3,[K2]=2578,[K1]=K5,[K5]="V",x=30,y=389,},x=27,xend=31,y=389,yend=389,},[2]={f=K3,[K2]=2579,[K1]=K5,[K5]="K",x=34,y=389,},},[K2]=2041,[K1]=K4,x=36,y=389,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=2580,[K1]=K5,[K5]="V",x=38,y=389,},},[K2]=2042,[K1]=K4,x=36,y=389,},[K36]="rawget",[K2]=2034,[K1]=K7,x=12,y=389,},[K19]={[1]={f=K3,[K10]="K",[K2]=2035,[K1]=K10,x=21,y=389,},[2]={f=K3,[K10]="V",[K2]=2036,[K1]=K10,x=24,y=389,},xend=25,yend=389,},[K2]=2044,[K1]=K20,x=4,y=390,},[2]={args={f=K3,[K4]={[1]={f=K3,keys={f=K3,[K2]=2046,[K1]="any",x=22,y=390,},[K2]=2047,[K1]="map",[K29]={f=K3,[K2]=2048,[K1]="any",x=26,y=390,},x=21,xend=29,y=390,yend=390,},[2]={f=K3,[K2]=2049,[K1]="any",x=32,y=390,},},[K2]=2050,[K1]=K4,x=36,y=390,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=2052,[K1]="any",x=38,y=390,},},[K2]=2051,[K1]=K4,x=36,y=390,},[K2]=2045,[K1]=K7,x=12,y=390,},[3]={args={f=K3,[K4]={[1]={f=K3,[K2]=2055,[K1]="any",x=21,y=391,},[2]={f=K3,[K2]=2056,[K1]="any",x=26,y=391,},},[K2]=2057,[K1]=K4,x=30,y=391,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=2059,[K1]="any",x=32,y=391,},},[K2]=2058,[K1]=K4,x=30,y=391,},[K2]=2054,[K1]=K7,x=12,y=391,},},x=10,y=390,},},["rawlen"]={[K27]=K26,[K28]=true,t={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={[K32]={f=K3,[K2]=2595,[K1]=K5,[K5]="A@50",x=25,y=393,},f=K3,[K2]=2596,[K1]=K33,x=24,y=393,},},[K2]=2597,[K1]=K4,x=28,y=393,},f=K3,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2066,[K1]=K12,x=30,y=393,},},[K2]=2065,[K1]=K4,x=28,y=393,},[K2]=2598,[K1]=K7,x=12,y=393,},[K19]={[1]={f=K3,[K10]="A@50",[K2]=2594,[K1]=K10,x=21,y=393,},},[K2]=2599,[K1]=K20,x=4,y=395,},},["rawset"]={[K27]=K26,t={f=K3,[K2]=2093,[K1]="poly",[K18]={[1]={f=K3,t={args={f=K3,[K4]={[1]={f=K3,keys={f=K3,[K2]=2602,[K1]=K5,[K5]="K",x=28,y=395,},[K2]=2072,[K1]="map",[K29]={f=K3,[K2]=2603,[K1]=K5,[K5]="V",x=30,y=395,},x=27,xend=31,y=395,yend=395,},[2]={f=K3,[K2]=2604,[K1]=K5,[K5]="K",x=34,y=395,},[3]={f=K3,[K2]=2605,[K1]=K5,[K5]="V",x=37,y=395,},},[K2]=2076,[K1]=K4,x=39,y=395,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,keys={f=K3,[K2]=2606,[K1]=K5,[K5]="K",x=42,y=395,},[K2]=2079,[K1]="map",[K29]={f=K3,[K2]=2607,[K1]=K5,[K5]="V",x=44,y=395,},x=41,xend=45,y=395,yend=395,},},[K2]=2077,[K1]=K4,x=39,y=395,},[K2]=2068,[K1]=K7,x=12,y=395,},[K19]={[1]={f=K3,[K10]="K",[K2]=2069,[K1]=K10,x=21,y=395,},[2]={f=K3,[K10]="V",[K2]=2070,[K1]=K10,x=24,y=395,},xend=25,yend=395,},[K2]=2081,[K1]=K20,x=4,y=396,},[2]={args={f=K3,[K4]={[1]={f=K3,keys={f=K3,[K2]=2083,[K1]="any",x=22,y=396,},[K2]=2084,[K1]="map",[K29]={f=K3,[K2]=2085,[K1]="any",x=26,y=396,},x=21,xend=29,y=396,yend=396,},[2]={f=K3,[K2]=2086,[K1]="any",x=32,y=396,},[3]={f=K3,[K2]=2087,[K1]="any",x=37,y=396,},},[K2]=2088,[K1]=K4,x=41,y=396,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,keys={f=K3,[K2]=2090,[K1]="any",x=44,y=396,},[K2]=2091,[K1]="map",[K29]={f=K3,[K2]=2092,[K1]="any",x=48,y=396,},x=43,xend=51,y=396,yend=396,},},[K2]=2089,[K1]=K4,x=41,y=396,},[K2]=2082,[K1]=K7,x=12,y=396,},[3]={args={f=K3,[K4]={[1]={f=K3,[K2]=2095,[K1]="any",x=21,y=397,},[2]={f=K3,[K2]=2096,[K1]="any",x=26,y=397,},[3]={f=K3,[K2]=2097,[K1]="any",x=31,y=397,},},[K2]=2098,[K1]=K4,x=35,y=397,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,[K2]=2100,[K1]="any",x=37,y=397,},},[K2]=2099,[K1]=K4,x=35,y=397,},[K2]=2094,[K1]=K7,x=12,y=397,},},x=10,y=396,},},["require"]={[K27]=K26,t={args={f=K3,[K4]={[1]={f=K3,[K2]=2102,[K1]=K11,x=22,y=399,},},[K2]=2103,[K1]=K4,x=29,y=399,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2105,[K1]="any",x=31,y=399,},},[K2]=2104,[K1]=K4,x=29,y=399,},[K36]="require",[K2]=2101,[K1]=K7,x=13,y=399,},},["select"]={[K27]=K26,t={f=K3,[K2]=2120,[K1]="poly",[K18]={[1]={f=K3,t={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2108,[K1]=K12,x=24,y=401,},[2]={f=K3,[K2]=2623,[K1]=K5,[K5]="T",x=33,y=401,},},[K2]=2110,[K1]=K4,x=38,y=401,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2624,[K1]=K5,[K5]="T",x=40,y=401,},},[K2]=2111,[K1]=K4,x=38,y=401,},[K2]=2106,[K1]=K7,x=12,y=401,},[K19]={[1]={f=K3,[K10]="T",[K2]=2107,[K1]=K10,x=21,y=401,},xend=22,yend=401,},[K2]=2113,[K1]=K20,x=4,y=402,},[2]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2115,[K1]=K12,x=21,y=402,},[2]={f=K3,[K2]=2116,[K1]="any",x=30,y=402,},},[K2]=2117,[K1]=K4,x=37,y=402,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2119,[K1]="any",x=39,y=402,},},[K2]=2118,[K1]=K4,x=37,y=402,},[K2]=2114,[K1]=K7,x=12,y=402,},[3]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2122,[K1]=K11,x=21,y=403,},[2]={f=K3,[K2]=2123,[K1]="any",x=29,y=403,},},[K2]=2124,[K1]=K4,x=36,y=403,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2126,[K1]=K12,x=38,y=403,},},[K2]=2125,[K1]=K4,x=36,y=403,},[K2]=2121,[K1]=K7,x=12,y=403,},},x=10,y=402,},},[K[16]]={[K27]=K26,t={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={f=K3,[K2]=2637,[K1]=K5,[K5]="T@53",x=30,y=405,},[2]={f=K3,[K17]=T12,[K16]={[1]=K41,},[K2]=2639,[K1]=K15,[K40]={[1]={f=K3,[K2]=2638,[K1]=K5,[K5]="T@53",x=43,y=405,},},x=33,y=405,},},[K2]=2640,[K1]=K4,x=46,y=405,},f=K3,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=2641,[K1]=K5,[K5]="T@53",x=48,y=405,},},[K2]=2642,[K1]=K4,x=46,y=405,},[K2]=2643,[K1]=K7,x=18,y=405,},[K19]={[1]={f=K3,[K10]="T@53",[K2]=2636,[K1]=K10,x=27,y=405,},},[K2]=2644,[K1]=K20,x=4,y=407,},},[K11]={[K28]=true,t={[K34]=true,def={[K24]=K11,f=K3,[K31]={[1]="byte",[2]="char",[3]="dump",[4]="find",[5]="format",[6]="gmatch",[7]="gsub",[8]="len",[9]="lower",[10]="match",[11]="pack",[12]="packsize",[13]="rep",[14]="reverse",[15]="sub",[16]=K[22],[17]="upper",},[K30]={byte={f=K3,[K2]=1472,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=1460,[K1]=K11,x=19,y=272,},[2]={f=K3,[K2]=1461,[K1]=K12,x=29,y=272,},},[K2]=1462,[K1]=K4,x=37,y=272,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1464,[K1]=K12,x=39,y=272,},},[K2]=1463,[K1]=K4,x=37,y=272,},[K2]=1459,[K1]=K7,x=10,y=272,},[2]={args={f=K3,[K4]={[1]={f=K3,[K2]=1466,[K1]=K11,x=19,y=273,},[2]={f=K3,[K2]=1467,[K1]=K12,x=27,y=273,},[3]={f=K3,[K2]=1468,[K1]=K12,x=38,y=273,},},[K2]=1469,[K1]=K4,x=46,y=273,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1471,[K1]=K12,x=48,y=273,},},[K2]=1470,[K1]=K4,x=46,y=273,},[K2]=1465,[K1]=K7,x=10,y=273,},},x=8,y=273,},char={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1474,[K1]=K12,x=19,y=275,},},[K2]=1475,[K1]=K4,x=30,y=275,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1477,[K1]=K11,x=32,y=275,},},[K2]=1476,[K1]=K4,x=30,y=275,},[K2]=1473,[K1]=K7,x=10,y=275,},dump={args={f=K3,[K4]={[1]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1480,[K1]="any",x=28,y=276,},},[K2]=1481,[K1]=K4,x=35,y=276,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1483,[K1]="any",x=38,y=276,},},[K2]=1482,[K1]=K4,x=35,y=276,},[K2]=1479,[K1]=K7,x=19,y=276,},[2]={f=K3,[K2]=1484,[K1]=K25,x=46,y=276,},},[K2]=1485,[K1]=K4,x=54,y=276,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1487,[K1]=K11,x=56,y=276,},},[K2]=1486,[K1]=K4,x=54,y=276,},[K2]=1478,[K1]=K7,x=10,y=276,},find={args={f=K3,[K4]={[1]={f=K3,[K2]=1489,[K1]=K11,x=19,y=277,},[2]={f=K3,[K2]=1490,[K1]=K11,x=27,y=277,},[3]={f=K3,[K2]=1491,[K1]=K12,x=37,y=277,},[4]={f=K3,[K2]=1492,[K1]=K25,x=48,y=277,},},[K2]=1493,[K1]=K4,x=56,y=277,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1495,[K1]=K12,x=58,y=277,},[2]={f=K3,[K2]=1496,[K1]=K12,x=67,y=277,},[3]={f=K3,[K2]=1497,[K1]=K11,x=76,y=277,},},[K2]=1494,[K1]=K4,x=56,y=277,},[K36]="string.find",[K2]=1488,[K1]=K7,x=10,y=277,},["format"]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1499,[K1]=K11,x=21,y=278,},[2]={f=K3,[K2]=1500,[K1]="any",x=29,y=278,},},[K2]=1501,[K1]=K4,x=36,y=278,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1503,[K1]=K11,x=38,y=278,},},[K2]=1502,[K1]=K4,x=36,y=278,},[K36]="string.format",[K2]=1498,[K1]=K7,x=12,y=278,},["gmatch"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1505,[K1]=K11,x=21,y=279,},[2]={f=K3,[K2]=1506,[K1]=K11,x=29,y=279,},[3]={f=K3,[K2]=1507,[K1]=K12,x=39,y=279,},},[K2]=1508,[K1]=K4,x=47,y=279,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=1511,[K1]=K4,x=60,y=279,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1513,[K1]=K11,x=62,y=279,},},[K2]=1512,[K1]=K4,x=60,y=279,},[K2]=1510,[K1]=K7,x=50,y=279,},},[K2]=1509,[K1]=K4,x=47,y=279,},[K36]="string.gmatch",[K2]=1504,[K1]=K7,x=12,y=279,},gsub={f=K3,[K2]=1537,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=1515,[K1]=K11,x=19,y=281,},[2]={f=K3,[K2]=1516,[K1]=K11,x=27,y=281,},[3]={f=K3,[K2]=1517,[K1]=K11,x=35,y=281,},[4]={f=K3,[K2]=1518,[K1]=K12,x=45,y=281,},},[K2]=1519,[K1]=K4,x=53,y=281,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,[K2]=1521,[K1]=K11,x=55,y=281,},[2]={f=K3,[K2]=1522,[K1]=K12,x=63,y=281,},},[K2]=1520,[K1]=K4,x=53,y=281,},[K36]="string.gsub",[K2]=1514,[K1]=K7,x=10,y=281,},[2]={args={f=K3,[K4]={[1]={f=K3,[K2]=1524,[K1]=K11,x=19,y=282,},[2]={f=K3,[K2]=1525,[K1]=K11,x=27,y=282,},[3]={f=K3,keys={f=K3,[K2]=1526,[K1]=K11,x=36,y=282,},[K2]=1527,[K1]="map",[K29]={f=K3,[K2]=1529,[K1]=K22,[K18]={[1]={f=K3,[K2]=1528,[K1]=K11,x=43,y=282,},[2]={f=K3,[K2]=1530,[K1]=K12,x=50,y=282,},[3]={f=K3,[K2]=1531,[K1]=K13,x=58,y=282,},},x=43,y=282,},x=35,xend=64,y=282,yend=282,},[4]={f=K3,[K2]=1532,[K1]=K12,x=69,y=282,},},[K2]=1533,[K1]=K4,x=77,y=282,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,[K2]=1535,[K1]=K11,x=79,y=282,},[2]={f=K3,[K2]=1536,[K1]=K12,x=87,y=282,},},[K2]=1534,[K1]=K4,x=77,y=282,},[K2]=1523,[K1]=K7,x=10,y=282,},[3]={args={f=K3,[K4]={[1]={f=K3,[K2]=1539,[K1]=K11,x=19,y=283,},[2]={f=K3,[K2]=1540,[K1]=K11,x=27,y=283,},[3]={f=K3,keys={f=K3,[K2]=1541,[K1]=K12,x=36,y=283,},[K2]=1542,[K1]="map",[K29]={f=K3,[K2]=1544,[K1]=K22,[K18]={[1]={f=K3,[K2]=1543,[K1]=K11,x=44,y=283,},[2]={f=K3,[K2]=1545,[K1]=K12,x=51,y=283,},[3]={f=K3,[K2]=1546,[K1]=K13,x=59,y=283,},},x=44,y=283,},x=35,xend=65,y=283,yend=283,},[4]={f=K3,[K2]=1547,[K1]=K12,x=70,y=283,},},[K2]=1548,[K1]=K4,x=78,y=283,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,[K2]=1550,[K1]=K11,x=80,y=283,},[2]={f=K3,[K2]=1551,[K1]=K12,x=88,y=283,},},[K2]=1549,[K1]=K4,x=78,y=283,},[K2]=1538,[K1]=K7,x=10,y=283,},[4]={args={f=K3,[K4]={[1]={f=K3,[K2]=1553,[K1]=K11,x=19,y=284,},[2]={f=K3,[K2]=1554,[K1]=K11,x=27,y=284,},[3]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1557,[K1]=K22,[K18]={[1]={f=K3,[K2]=1556,[K1]=K11,x=45,y=284,},[2]={f=K3,[K2]=1558,[K1]=K12,x=52,y=284,},},x=45,y=284,},},[K2]=1559,[K1]=K4,x=64,y=284,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1562,[K1]=K22,[K18]={[1]={f=K3,[K2]=1561,[K1]=K11,x=68,y=284,},[2]={f=K3,[K2]=1563,[K1]=K12,x=75,y=284,},[3]={f=K3,[K2]=1564,[K1]=K13,x=83,y=284,},},x=68,y=284,},},[K2]=1560,[K1]=K4,x=64,y=284,},[K2]=1555,[K1]=K7,x=35,y=284,},[4]={f=K3,[K2]=1565,[K1]=K12,x=98,y=284,},},[K2]=1566,[K1]=K4,x=106,y=284,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,[K2]=1568,[K1]=K11,x=108,y=284,},[2]={f=K3,[K2]=1569,[K1]=K12,x=116,y=284,},},[K2]=1567,[K1]=K4,x=106,y=284,},[K2]=1552,[K1]=K7,x=10,y=284,},},x=8,y=282,},len={args={f=K3,[K4]={[1]={f=K3,[K2]=1571,[K1]=K11,x=18,y=286,},},[K2]=1572,[K1]=K4,x=25,y=286,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1574,[K1]=K12,x=27,y=286,},},[K2]=1573,[K1]=K4,x=25,y=286,},[K2]=1570,[K1]=K7,x=9,y=286,},["lower"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1576,[K1]=K11,x=20,y=287,},},[K2]=1577,[K1]=K4,x=27,y=287,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1579,[K1]=K11,x=29,y=287,},},[K2]=1578,[K1]=K4,x=27,y=287,},[K2]=1575,[K1]=K7,x=11,y=287,},["match"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1581,[K1]=K11,x=20,y=288,},[2]={f=K3,[K2]=1582,[K1]=K11,x=28,y=288,},[3]={f=K3,[K2]=1583,[K1]=K12,x=38,y=288,},},[K2]=1584,[K1]=K4,x=46,y=288,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1586,[K1]=K11,x=48,y=288,},},[K2]=1585,[K1]=K4,x=46,y=288,},[K36]="string.match",[K2]=1580,[K1]=K7,x=11,y=288,},pack={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1588,[K1]=K11,x=19,y=289,},[2]={f=K3,[K2]=1589,[K1]="any",x=27,y=289,},},[K2]=1590,[K1]=K4,x=34,y=289,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1592,[K1]=K11,x=36,y=289,},},[K2]=1591,[K1]=K4,x=34,y=289,},[K36]="string.pack",[K2]=1587,[K1]=K7,x=10,y=289,},["packsize"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1594,[K1]=K11,x=23,y=290,},},[K2]=1595,[K1]=K4,x=30,y=290,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1597,[K1]=K12,x=32,y=290,},},[K2]=1596,[K1]=K4,x=30,y=290,},[K2]=1593,[K1]=K7,x=14,y=290,},rep={args={f=K3,[K4]={[1]={f=K3,[K2]=1599,[K1]=K11,x=18,y=291,},[2]={f=K3,[K2]=1600,[K1]=K12,x=26,y=291,},[3]={f=K3,[K2]=1601,[K1]=K11,x=37,y=291,},},[K2]=1602,[K1]=K4,x=44,y=291,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1604,[K1]=K11,x=46,y=291,},},[K2]=1603,[K1]=K4,x=44,y=291,},[K2]=1598,[K1]=K7,x=9,y=291,},["reverse"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1606,[K1]=K11,x=22,y=292,},},[K2]=1607,[K1]=K4,x=29,y=292,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1609,[K1]=K11,x=31,y=292,},},[K2]=1608,[K1]=K4,x=29,y=292,},[K2]=1605,[K1]=K7,x=13,y=292,},sub={args={f=K3,[K4]={[1]={f=K3,[K2]=1611,[K1]=K11,x=18,y=293,},[2]={f=K3,[K2]=1612,[K1]=K12,x=26,y=293,},[3]={f=K3,[K2]=1613,[K1]=K12,x=37,y=293,},},[K2]=1614,[K1]=K4,x=45,y=293,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1616,[K1]=K11,x=47,y=293,},},[K2]=1615,[K1]=K4,x=45,y=293,},[K2]=1610,[K1]=K7,x=9,y=293,},[K[22]]={args={f=K3,[K4]={[1]={f=K3,[K2]=1618,[K1]=K11,x=21,y=294,},[2]={f=K3,[K2]=1619,[K1]=K11,x=29,y=294,},[3]={f=K3,[K2]=1620,[K1]=K12,x=39,y=294,},},[K2]=1621,[K1]=K4,x=47,y=294,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1623,[K1]="any",x=49,y=294,},},[K2]=1622,[K1]=K4,x=47,y=294,},[K36]="string.unpack",[K2]=1617,[K1]=K7,x=12,y=294,},["upper"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1625,[K1]=K11,x=20,y=295,},},[K2]=1626,[K1]=K4,x=27,y=295,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1628,[K1]=K11,x=29,y=295,},},[K2]=1627,[K1]=K4,x=27,y=295,},[K2]=1624,[K1]=K7,x=11,y=295,},},[K2]=1458,[K1]=K37,x=1,y=271,},f=K3,[K2]=1629,[K1]=K21,x=1,y=271,},},["table"]={[K28]=true,t={[K34]=true,def={[K24]="table",f=K3,[K31]={[1]=K[23],[2]=K[24],[3]="concat",[4]="insert",[5]="move",[6]="pack",[7]=K[21],[8]="sort",[9]=K[22],},[K30]={[K[24]]=T21,[K[23]]=T23,["concat"]={args={f=K3,[K4]={[1]={[K32]={f=K3,[K2]=1649,[K1]=K22,[K18]={[1]={f=K3,[K2]=1648,[K1]=K11,x=23,y=307,},[2]={f=K3,[K2]=1650,[K1]=K13,x=32,y=307,},},x=23,y=307,},f=K3,[K2]=1651,[K1]=K33,x=21,xend=39,y=307,yend=307,},[2]={f=K3,[K2]=1652,[K1]=K11,x=44,y=307,},[3]={f=K3,[K2]=1653,[K1]=K12,x=54,y=307,},[4]={f=K3,[K2]=1654,[K1]=K12,x=65,y=307,},},[K2]=1655,[K1]=K4,x=73,y=307,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1657,[K1]=K11,x=75,y=307,},},[K2]=1656,[K1]=K4,x=73,y=307,},[K2]=1647,[K1]=K7,x=12,y=307,},["insert"]={f=K3,[K2]=1675,[K1]="poly",[K18]={[1]={f=K3,t={args={f=K3,[K4]={[1]={[K32]={f=K3,[K2]=2304,[K1]=K5,[K5]="A",x=25,y=309,},f=K3,[K2]=1661,[K1]=K33,x=24,xend=26,y=309,yend=309,},[2]={f=K3,[K2]=1662,[K1]=K12,x=29,y=309,},[3]={f=K3,[K2]=2305,[K1]=K5,[K5]="A",x=38,y=309,},},[K2]=1664,[K1]=K4,x=4,y=310,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={},[K2]=1665,[K1]=K4,x=39,y=309,},[K2]=1658,[K1]=K7,x=12,y=309,},[K19]={[1]={f=K3,[K10]="A",[K2]=1659,[K1]=K10,x=21,y=309,},xend=22,yend=309,},[K2]=1666,[K1]=K20,x=4,y=310,},[2]={f=K3,t={args={f=K3,[K4]={[1]={[K32]={f=K3,[K2]=2314,[K1]=K5,[K5]="A",x=25,y=310,},f=K3,[K2]=1670,[K1]=K33,x=24,xend=26,y=310,yend=310,},[2]={f=K3,[K2]=2315,[K1]=K5,[K5]="A",x=29,y=310,},},[K2]=1672,[K1]=K4,x=4,y=312,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={},[K2]=1673,[K1]=K4,x=30,y=310,},[K2]=1667,[K1]=K7,x=12,y=310,},[K19]={[1]={f=K3,[K10]="A",[K2]=1668,[K1]=K10,x=21,y=310,},xend=22,yend=310,},[K2]=1674,[K1]=K20,x=4,y=312,},},x=10,y=310,},move={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={[K32]={f=K3,[K2]=2328,[K1]=K5,[K5]="A@34",x=23,y=312,},f=K3,[K2]=2329,[K1]=K33,x=22,y=312,},[2]={f=K3,[K2]=1680,[K1]=K12,x=27,y=312,},[3]={f=K3,[K2]=1681,[K1]=K12,x=36,y=312,},[4]={f=K3,[K2]=1682,[K1]=K12,x=45,y=312,},[5]={[K32]={f=K3,[K2]=2330,[K1]=K5,[K5]="A@34",x=57,y=312,},f=K3,[K2]=2331,[K1]=K33,x=56,y=312,},},[K2]=2332,[K1]=K4,x=60,y=312,},f=K3,[K6]=4,rets={f=K3,[K4]={[1]={[K32]={f=K3,[K2]=2333,[K1]=K5,[K5]="A@34",x=63,y=312,},f=K3,[K2]=2334,[K1]=K33,x=62,y=312,},},[K2]=2335,[K1]=K4,x=60,y=312,},[K2]=2336,[K1]=K7,x=10,y=312,},[K19]={[1]={f=K3,[K10]="A@34",[K2]=2327,[K1]=K10,x=19,y=312,},},[K2]=2337,[K1]=K20,x=4,y=314,},pack={f=K3,[K28]=true,[K2]=1705,[K1]="poly",[K18]={[1]={f=K3,t={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2339,[K1]=K5,[K5]="T",x=22,y=314,},},[K2]=1693,[K1]=K4,x=27,y=314,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K17]=T21,[K16]={[1]=K[24],},[K2]=1695,[K1]=K15,[K40]={[1]={f=K3,[K2]=2340,[K1]=K5,[K5]="T",x=39,y=314,},xend=40,yend=314,},x=29,y=314,},},[K2]=1694,[K1]=K4,x=27,y=314,},[K2]=1690,[K1]=K7,x=10,y=314,},[K19]={[1]={f=K3,[K10]="T",[K2]=1691,[K1]=K10,x=19,y=314,},xend=20,yend=314,},[K2]=1697,[K1]=K20,x=4,y=315,},[2]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1699,[K1]="any",x=19,y=315,},},[K2]=1700,[K1]=K4,x=26,y=315,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,keys={f=K3,[K2]=1702,[K1]="any",x=29,y=315,},[K2]=1703,[K1]="map",[K29]={f=K3,[K2]=1704,[K1]="any",x=33,y=315,},x=28,xend=36,y=315,yend=315,},},[K2]=1701,[K1]=K4,x=26,y=315,},[K2]=1698,[K1]=K7,x=10,y=315,},},x=8,y=315,},[K[21]]={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={[K32]={f=K3,[K2]=2353,[K1]=K5,[K5]="A@36",x=25,y=317,},f=K3,[K2]=2354,[K1]=K33,x=24,y=317,},[2]={f=K3,[K2]=1710,[K1]=K12,x=31,y=317,},},[K2]=2355,[K1]=K4,x=39,y=317,},f=K3,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2356,[K1]=K5,[K5]="A@36",x=41,y=317,},},[K2]=2357,[K1]=K4,x=39,y=317,},[K2]=2358,[K1]=K7,x=12,y=317,},[K19]={[1]={f=K3,[K10]="A@36",[K2]=2352,[K1]=K10,x=21,y=317,},},[K2]=2359,[K1]=K20,x=4,y=318,},sort={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={[K32]={f=K3,[K2]=2364,[K1]=K5,[K5]="A@37",x=23,y=318,},f=K3,[K2]=2365,[K1]=K33,x=22,y=318,},[2]={f=K3,[K17]=T23,[K16]={[1]=K[23],},[K2]=2367,[K1]=K15,[K40]={[1]={f=K3,[K2]=2366,[K1]=K5,[K5]="A@37",x=42,y=318,},},x=29,y=318,},},[K2]=2368,[K1]=K4,x=4,y=320,},f=K3,[K6]=1,rets={f=K3,[K4]={},[K2]=1722,[K1]=K4,x=44,y=318,},[K2]=2369,[K1]=K7,x=10,y=318,},[K19]={[1]={f=K3,[K10]="A@37",[K2]=2363,[K1]=K10,x=19,y=318,},},[K2]=2370,[K1]=K20,x=4,y=320,},[K[22]]={f=K3,[K28]=true,[K2]=1745,[K1]="poly",[K18]={[1]={f=K3,t={args={f=K3,[K4]={[1]={[K32]={f=K3,[K2]=2372,[K1]=K5,[K5]="A",x=25,y=320,},f=K3,[K2]=1727,[K1]=K33,x=24,xend=26,y=320,yend=320,},[2]={f=K3,[K2]=1728,[K1]=K13,x=31,y=320,},[3]={f=K3,[K2]=1729,[K1]=K13,x=41,y=320,},},[K2]=1730,[K1]=K4,x=48,y=320,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2373,[K1]=K5,[K5]="A",x=50,y=320,},},[K2]=1731,[K1]=K4,x=48,y=320,},[K2]=1724,[K1]=K7,x=12,y=320,},[K19]={[1]={f=K3,[K10]="A",[K2]=1725,[K1]=K10,x=21,y=320,},xend=22,yend=320,},[K2]=1733,[K1]=K20,x=4,y=321,},[2]={f=K3,t={args={f=K3,[K4]={[1]={f=K3,[K2]=1738,[K1]=K[25],[K18]={[1]={f=K3,[K2]=2384,[K1]=K5,[K5]="A1",x=30,y=321,},[2]={f=K3,[K2]=2385,[K1]=K5,[K5]="A2",x=34,y=321,},},x=29,xend=36,y=321,yend=321,},},[K2]=1740,[K1]=K4,x=38,y=321,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2386,[K1]=K5,[K5]="A1",x=40,y=321,},[2]={f=K3,[K2]=2387,[K1]=K5,[K5]="A2",x=44,y=321,},},[K2]=1741,[K1]=K4,x=38,y=321,},[K2]=1734,[K1]=K7,x=12,y=321,},[K19]={[1]={f=K3,[K10]="A1",[K2]=1735,[K1]=K10,x=21,y=321,},[2]={f=K3,[K10]="A2",[K2]=1736,[K1]=K10,x=25,y=321,},xend=27,yend=321,},[K2]=1744,[K1]=K20,x=4,y=322,},[3]={f=K3,t={args={f=K3,[K4]={[1]={f=K3,[K2]=1751,[K1]=K[25],[K18]={[1]={f=K3,[K2]=2402,[K1]=K5,[K5]="A1",x=34,y=322,},[2]={f=K3,[K2]=2403,[K1]=K5,[K5]="A2",x=38,y=322,},[3]={f=K3,[K2]=2404,[K1]=K5,[K5]="A3",x=42,y=322,},},x=33,xend=44,y=322,yend=322,},},[K2]=1754,[K1]=K4,x=46,y=322,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2405,[K1]=K5,[K5]="A1",x=48,y=322,},[2]={f=K3,[K2]=2406,[K1]=K5,[K5]="A2",x=52,y=322,},[3]={f=K3,[K2]=2407,[K1]=K5,[K5]="A3",x=56,y=322,},},[K2]=1755,[K1]=K4,x=46,y=322,},[K2]=1746,[K1]=K7,x=12,y=322,},[K19]={[1]={f=K3,[K10]="A1",[K2]=1747,[K1]=K10,x=21,y=322,},[2]={f=K3,[K10]="A2",[K2]=1748,[K1]=K10,x=25,y=322,},[3]={f=K3,[K10]="A3",[K2]=1749,[K1]=K10,x=29,y=322,},xend=31,yend=322,},[K2]=1759,[K1]=K20,x=4,y=323,},[4]={f=K3,t={args={f=K3,[K4]={[1]={f=K3,[K2]=1766,[K1]=K[25],[K18]={[1]={f=K3,[K2]=2426,[K1]=K5,[K5]="A1",x=38,y=323,},[2]={f=K3,[K2]=2427,[K1]=K5,[K5]="A2",x=42,y=323,},[3]={f=K3,[K2]=2428,[K1]=K5,[K5]="A3",x=46,y=323,},[4]={f=K3,[K2]=2429,[K1]=K5,[K5]="A4",x=50,y=323,},},x=37,xend=52,y=323,yend=323,},},[K2]=1770,[K1]=K4,x=54,y=323,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2430,[K1]=K5,[K5]="A1",x=56,y=323,},[2]={f=K3,[K2]=2431,[K1]=K5,[K5]="A2",x=60,y=323,},[3]={f=K3,[K2]=2432,[K1]=K5,[K5]="A3",x=64,y=323,},[4]={f=K3,[K2]=2433,[K1]=K5,[K5]="A4",x=68,y=323,},},[K2]=1771,[K1]=K4,x=54,y=323,},[K2]=1760,[K1]=K7,x=12,y=323,},[K19]={[1]={f=K3,[K10]="A1",[K2]=1761,[K1]=K10,x=21,y=323,},[2]={f=K3,[K10]="A2",[K2]=1762,[K1]=K10,x=25,y=323,},[3]={f=K3,[K10]="A3",[K2]=1763,[K1]=K10,x=29,y=323,},[4]={f=K3,[K10]="A4",[K2]=1764,[K1]=K10,x=33,y=323,},xend=35,yend=323,},[K2]=1776,[K1]=K20,x=4,y=324,},[5]={f=K3,t={args={f=K3,[K4]={[1]={f=K3,[K2]=1784,[K1]=K[25],[K18]={[1]={f=K3,[K2]=2456,[K1]=K5,[K5]="A1",x=42,y=324,},[2]={f=K3,[K2]=2457,[K1]=K5,[K5]="A2",x=46,y=324,},[3]={f=K3,[K2]=2458,[K1]=K5,[K5]="A3",x=50,y=324,},[4]={f=K3,[K2]=2459,[K1]=K5,[K5]="A4",x=54,y=324,},[5]={f=K3,[K2]=2460,[K1]=K5,[K5]="A5",x=58,y=324,},},x=41,xend=60,y=324,yend=324,},},[K2]=1789,[K1]=K4,x=62,y=324,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2461,[K1]=K5,[K5]="A1",x=64,y=324,},[2]={f=K3,[K2]=2462,[K1]=K5,[K5]="A2",x=68,y=324,},[3]={f=K3,[K2]=2463,[K1]=K5,[K5]="A3",x=72,y=324,},[4]={f=K3,[K2]=2464,[K1]=K5,[K5]="A4",x=76,y=324,},[5]={f=K3,[K2]=2465,[K1]=K5,[K5]="A5",x=80,y=324,},},[K2]=1790,[K1]=K4,x=62,y=324,},[K2]=1777,[K1]=K7,x=12,y=324,},[K19]={[1]={f=K3,[K10]="A1",[K2]=1778,[K1]=K10,x=21,y=324,},[2]={f=K3,[K10]="A2",[K2]=1779,[K1]=K10,x=25,y=324,},[3]={f=K3,[K10]="A3",[K2]=1780,[K1]=K10,x=29,y=324,},[4]={f=K3,[K10]="A4",[K2]=1781,[K1]=K10,x=33,y=324,},[5]={f=K3,[K10]="A5",[K2]=1782,[K1]=K10,x=37,y=324,},xend=39,yend=324,},[K2]=1796,[K1]=K20,x=1,y=325,},},x=10,y=321,},},[K2]=1630,[K1]=K37,x=1,y=298,},f=K3,[K2]=1797,[K1]=K21,x=1,y=298,},},[K35]={t={[K34]=true,def={[K24]=K35,f=K8,[K31]={},[K30]={},[K2]=7,[K1]=K[11],x=1,y=11,},f=K8,[K2]=8,[K1]=K21,x=1,y=11,},},["tonumber"]={[K27]=K26,t={f=K3,[K2]=2147,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=2137,[K1]="any",x=23,y=407,},},[K2]=2138,[K1]=K4,x=27,y=407,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2140,[K1]=K13,x=29,y=407,},},[K2]=2139,[K1]=K4,x=27,y=407,},[K2]=2136,[K1]=K7,x=14,y=407,},[2]={args={f=K3,[K4]={[1]={f=K3,[K2]=2142,[K1]="any",x=23,y=408,},[2]={f=K3,[K2]=2143,[K1]=K12,x=28,y=408,},},[K2]=2144,[K1]=K4,x=36,y=408,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=2146,[K1]=K12,x=38,y=408,},},[K2]=2145,[K1]=K4,x=36,y=408,},[K2]=2141,[K1]=K7,x=14,y=408,},},x=12,y=408,},},["tostring"]={[K27]=K26,t={args={f=K3,[K4]={[1]={f=K3,[K2]=2149,[K1]="any",x=23,y=410,},},[K2]=2150,[K1]=K4,x=27,y=410,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2152,[K1]=K11,x=29,y=410,},},[K2]=2151,[K1]=K4,x=27,y=410,},[K2]=2148,[K1]=K7,x=14,y=410,},},type={[K27]=K26,t={args={f=K3,[K4]={[1]={f=K3,[K2]=2154,[K1]="any",x=19,y=411,},},[K2]=2155,[K1]=K4,x=23,y=411,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2157,[K1]=K11,x=25,y=411,},},[K2]=2156,[K1]=K4,x=23,y=411,},[K2]=2153,[K1]=K7,x=10,y=411,},},[K[17]]={t=T14,[K[10]]=true,},utf8={[K28]=true,t={[K34]=true,def={[K24]="utf8",f=K3,[K31]={[1]="char",[2]="charpattern",[3]="codepoint",[4]="codes",[5]="len",[6]="offset",},[K30]={char={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1800,[K1]=K13,x=19,y=328,},},[K2]=1801,[K1]=K4,x=29,y=328,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1803,[K1]=K11,x=31,y=328,},},[K2]=1802,[K1]=K4,x=29,y=328,},[K2]=1799,[K1]=K7,x=10,y=328,},["charpattern"]={f=K3,[K2]=1804,[K1]=K11,x=17,y=329,},["codepoint"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1806,[K1]=K11,x=24,y=330,},[2]={f=K3,[K2]=1807,[K1]=K13,x=34,y=330,},[3]={f=K3,[K2]=1808,[K1]=K13,x=44,y=330,},[4]={f=K3,[K2]=1809,[K1]=K25,x=54,y=330,},},[K2]=1810,[K1]=K4,x=62,y=330,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1812,[K1]=K13,x=64,y=330,},},[K2]=1811,[K1]=K4,x=62,y=330,},[K2]=1805,[K1]=K7,x=15,y=330,},["codes"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1814,[K1]=K11,x=20,y=331,},[2]={f=K3,[K2]=1815,[K1]=K25,x=30,y=331,},},[K2]=1816,[K1]=K4,x=38,y=331,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=1819,[K1]=K11,x=50,y=331,},[2]={f=K3,[K2]=1820,[K1]=K13,x=60,y=331,},},[K2]=1821,[K1]=K4,x=67,y=331,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1823,[K1]=K13,x=70,y=331,},[2]={f=K3,[K2]=1824,[K1]=K13,x=78,y=331,},},[K2]=1822,[K1]=K4,x=67,y=331,},[K2]=1818,[K1]=K7,x=41,y=331,},},[K2]=1817,[K1]=K4,x=38,y=331,},[K2]=1813,[K1]=K7,x=11,y=331,},len={args={f=K3,[K4]={[1]={f=K3,[K2]=1826,[K1]=K11,x=18,y=332,},[2]={f=K3,[K2]=1827,[K1]=K13,x=28,y=332,},[3]={f=K3,[K2]=1828,[K1]=K13,x=38,y=332,},[4]={f=K3,[K2]=1829,[K1]=K25,x=48,y=332,},},[K2]=1830,[K1]=K4,x=56,y=332,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1832,[K1]=K13,x=58,y=332,},},[K2]=1831,[K1]=K4,x=56,y=332,},[K2]=1825,[K1]=K7,x=9,y=332,},["offset"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1834,[K1]=K11,x=21,y=333,},[2]={f=K3,[K2]=1835,[K1]=K13,x=29,y=333,},[3]={f=K3,[K2]=1836,[K1]=K13,x=39,y=333,},},[K2]=1837,[K1]=K4,x=46,y=333,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1839,[K1]=K13,x=48,y=333,},},[K2]=1838,[K1]=K4,x=46,y=333,},[K2]=1833,[K1]=K7,x=12,y=333,},},[K2]=1798,[K1]=K37,x=1,y=327,},f=K3,[K2]=1840,[K1]=K21,x=1,y=327,},},["xpcall"]={[K27]=K26,[K28]=true,t={args={f=K3,[K14]=true,[K4]={[1]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2165,[K1]="any",x=30,y=413,},},[K2]=2166,[K1]=K4,x=37,y=413,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2168,[K1]="any",x=39,y=413,},},[K2]=2167,[K1]=K4,x=37,y=413,},[K2]=2164,[K1]=K7,x=21,y=413,},[2]={f=K3,[K17]={def={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1856,[K1]="any",x=44,y=360,},},[K2]=1857,[K1]=K4,x=48,y=360,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={},[K2]=1858,[K1]=K4,x=47,y=360,},[K2]=1855,[K1]=K7,x=30,y=360,},f=K3,[K2]=1859,[K1]=K21,x=30,y=360,},[K16]={[1]="XpcallMsghFunction",},[K2]=2169,[K1]=K15,x=48,y=413,},[3]={f=K3,[K2]=2170,[K1]="any",x=68,y=413,},},[K2]=2171,[K1]=K4,x=75,y=413,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2173,[K1]=K25,x=77,y=413,},[2]={f=K3,[K2]=2174,[K1]="any",x=86,y=413,},},[K2]=2172,[K1]=K4,x=75,y=413,},[K36]="xpcall",[K2]=2163,[K1]=K7,x=12,y=413,},},}
-T1.def={[K24]="FILE",f=K3,[K31]={[1]=K[2],[2]=K[3],[3]=K42,[4]=K[4],[5]=K[5],[6]="read",[7]="seek",[8]="setvbuf",[9]=K[6],},[K30]={[K[2]]=T2,[K[3]]=T3,[K42]={args={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=566,[K1]=K15,x=20,y=37,},},[K2]=567,[K1]=K4,x=25,y=37,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=569,[K1]=K25,x=27,y=37,},[2]={f=K3,[K2]=570,[K1]=K11,x=36,y=37,},[3]={f=K3,[K2]=571,[K1]=K12,x=44,y=37,},},[K2]=568,[K1]=K4,x=25,y=37,},[K2]=565,[K1]=K7,x=11,y=37,},[K[4]]={args={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=573,[K1]=K15,x=20,y=38,},},[K2]=574,[K1]=K4,x=4,y=40,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={},[K2]=575,[K1]=K4,x=24,y=38,},[K2]=572,[K1]=K7,x=11,y=38,},[K[5]]={f=K3,[K2]=593,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=577,[K1]=K15,x=20,y=40,},},[K2]=578,[K1]=K4,x=25,y=40,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=581,[K1]=K4,x=38,y=40,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=583,[K1]=K11,x=41,y=40,},},[K2]=582,[K1]=K4,x=38,y=40,},[K2]=580,[K1]=K7,x=28,y=40,},},[K2]=579,[K1]=K4,x=25,y=40,},[K2]=576,[K1]=K7,x=11,y=40,},[2]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=585,[K1]=K15,x=20,y=41,},[2]={f=K3,[K17]=T4,[K16]={[1]=K48,},[K2]=586,[K1]=K15,x=26,y=41,},},[K2]=587,[K1]=K4,x=44,y=41,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=590,[K1]=K4,x=57,y=41,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=592,[K1]=K13,x=60,y=41,},},[K2]=591,[K1]=K4,x=57,y=41,},[K2]=589,[K1]=K7,x=47,y=41,},},[K2]=588,[K1]=K4,x=44,y=41,},[K2]=584,[K1]=K7,x=11,y=41,},[3]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=595,[K1]=K15,x=20,y=42,},[2]={f=K3,[K2]=597,[K1]=K22,[K18]={[1]={f=K3,[K2]=596,[K1]=K13,x=27,y=42,},[2]={f=K3,[K17]=T5,[K16]={[1]=K47,},[K2]=598,[K1]=K15,x=36,y=42,},},x=27,y=42,},},[K2]=599,[K1]=K4,x=55,y=42,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=602,[K1]=K4,x=68,y=42,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=604,[K1]=K11,x=71,y=42,},},[K2]=603,[K1]=K4,x=68,y=42,},[K2]=601,[K1]=K7,x=58,y=42,},},[K2]=600,[K1]=K4,x=55,y=42,},[K2]=594,[K1]=K7,x=11,y=42,},[4]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=606,[K1]=K15,x=20,y=43,},[2]={f=K3,[K2]=608,[K1]=K22,[K18]={[1]={f=K3,[K2]=607,[K1]=K13,x=27,y=43,},[2]={f=K3,[K17]=T6,[K16]={[1]=K49,},[K2]=609,[K1]=K15,x=36,y=43,},},x=27,y=43,},},[K2]=610,[K1]=K4,x=49,y=43,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=613,[K1]=K4,x=62,y=43,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=616,[K1]=K22,[K18]={[1]={f=K3,[K2]=615,[K1]=K11,x=66,y=43,},[2]={f=K3,[K2]=617,[K1]=K13,x=75,y=43,},},x=66,y=43,},},[K2]=614,[K1]=K4,x=62,y=43,},[K2]=612,[K1]=K7,x=52,y=43,},},[K2]=611,[K1]=K4,x=49,y=43,},[K2]=605,[K1]=K7,x=11,y=43,},[5]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=619,[K1]=K15,x=20,y=44,},[2]={f=K3,[K2]=621,[K1]=K22,[K18]={[1]={f=K3,[K2]=620,[K1]=K13,x=27,y=44,},[2]={f=K3,[K2]=622,[K1]=K11,x=36,y=44,},},x=27,y=44,},},[K2]=623,[K1]=K4,x=47,y=44,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=626,[K1]=K4,x=60,y=44,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=628,[K1]=K11,x=63,y=44,},},[K2]=627,[K1]=K4,x=60,y=44,},[K2]=625,[K1]=K7,x=50,y=44,},},[K2]=624,[K1]=K4,x=47,y=44,},[K2]=618,[K1]=K7,x=11,y=44,},},x=9,y=41,},read={f=K3,[K2]=640,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=630,[K1]=K15,x=19,y=46,},},[K2]=631,[K1]=K4,x=24,y=46,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=633,[K1]=K11,x=26,y=46,},},[K2]=632,[K1]=K4,x=24,y=46,},[K2]=629,[K1]=K7,x=10,y=46,},[2]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=635,[K1]=K15,x=19,y=47,},[2]={f=K3,[K17]=T4,[K16]={[1]=K48,},[K2]=636,[K1]=K15,x=25,y=47,},},[K2]=637,[K1]=K4,x=43,y=47,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=639,[K1]=K13,x=45,y=47,},},[K2]=638,[K1]=K4,x=43,y=47,},[K2]=634,[K1]=K7,x=10,y=47,},[3]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=642,[K1]=K15,x=19,y=48,},[2]={f=K3,[K2]=644,[K1]=K22,[K18]={[1]={f=K3,[K2]=643,[K1]=K13,x=26,y=48,},[2]={f=K3,[K17]=T5,[K16]={[1]=K47,},[K2]=645,[K1]=K15,x=35,y=48,},},x=26,y=48,},},[K2]=646,[K1]=K4,x=54,y=48,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=648,[K1]=K11,x=56,y=48,},},[K2]=647,[K1]=K4,x=54,y=48,},[K2]=641,[K1]=K7,x=10,y=48,},[4]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=650,[K1]=K15,x=19,y=49,},[2]={f=K3,[K2]=652,[K1]=K22,[K18]={[1]={f=K3,[K2]=651,[K1]=K13,x=26,y=49,},[2]={f=K3,[K17]=T6,[K16]={[1]=K49,},[K2]=653,[K1]=K15,x=35,y=49,},},x=26,y=49,},},[K2]=654,[K1]=K4,x=48,y=49,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=657,[K1]=K22,[K18]={[1]={f=K3,[K2]=656,[K1]=K11,x=52,y=49,},[2]={f=K3,[K2]=658,[K1]=K13,x=61,y=49,},},x=52,y=49,},},[K2]=655,[K1]=K4,x=48,y=49,},[K2]=649,[K1]=K7,x=10,y=49,},[5]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=660,[K1]=K15,x=19,y=50,},[2]={f=K3,[K2]=662,[K1]=K22,[K18]={[1]={f=K3,[K2]=661,[K1]=K13,x=26,y=50,},[2]={f=K3,[K2]=663,[K1]=K11,x=35,y=50,},},x=26,y=50,},},[K2]=664,[K1]=K4,x=46,y=50,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=666,[K1]=K11,x=49,y=50,},},[K2]=665,[K1]=K4,x=46,y=50,},[K2]=659,[K1]=K7,x=10,y=50,},},x=8,y=47,},seek={args={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=668,[K1]=K15,x=19,y=52,},[2]={f=K3,[K17]=T2,[K16]={[1]=K[2],},[K2]=669,[K1]=K15,x=27,y=52,},[3]={f=K3,[K2]=670,[K1]=K12,x=41,y=52,},},[K2]=671,[K1]=K4,x=49,y=52,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=673,[K1]=K12,x=51,y=52,},[2]={f=K3,[K2]=674,[K1]=K11,x=60,y=52,},},[K2]=672,[K1]=K4,x=49,y=52,},[K2]=667,[K1]=K7,x=10,y=52,},["setvbuf"]={args={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=676,[K1]=K15,x=22,y=53,},[2]={f=K3,[K17]=T3,[K16]={[1]=K[3],},[K2]=677,[K1]=K15,x=28,y=53,},[3]={f=K3,[K2]=678,[K1]=K12,x=43,y=53,},},[K2]=679,[K1]=K4,x=4,y=55,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={},[K2]=680,[K1]=K4,x=50,y=53,},[K2]=675,[K1]=K7,x=13,y=53,},[K[6]]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=682,[K1]=K15,x=20,y=55,},[2]={f=K3,[K2]=684,[K1]=K22,[K18]={[1]={f=K3,[K2]=683,[K1]=K11,x=27,y=55,},[2]={f=K3,[K2]=685,[K1]=K13,x=36,y=55,},},x=27,y=55,},},[K2]=686,[K1]=K4,x=47,y=55,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=688,[K1]=K15,x=49,y=55,},[2]={f=K3,[K2]=689,[K1]=K11,x=55,y=55,},},[K2]=687,[K1]=K4,x=47,y=55,},[K2]=681,[K1]=K7,x=11,y=55,},},[K[7]]={},[K[8]]=true,["is_userdata"]=true,["meta_field_order"]={[1]=K[9],},["meta_fields"]={[K[9]]={args={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=691,[K1]=K15,x=33,y=57,},},[K2]=692,[K1]=K4,x=1,y=58,},f=K3,[K44]=true,[K9]=false,[K6]=1,rets={f=K3,[K4]={},[K2]=693,[K1]=K4,x=37,y=57,},[K2]=690,[K1]=K7,x=24,y=57,},},[K2]=559,[K1]=K37,x=1,y=26,}
+local T0 = {["..."]={t={f=K[1],[K14]=true,[K4]={[1]={f=K[1],[K2]=2746,[K1]=K11,x=1,y=1,},},[K2]=2747,[K1]=K4,x=1,y=1,},},["@is_va"]={t={f=K[1],[K2]=2748,[K1]="any",x=1,y=1,},},FILE={t=T1,[K[11]]=true,},["_VERSION"]={[K27]=K26,t={f=K3,[K2]=2175,[K1]=K11,x=14,y=414,},},any={t={[K34]=true,def={[K24]="any",f=K8,[K31]={},[K30]={},[K2]=5,[K1]=K[12],x=1,y=8,},f=K8,[K2]=6,[K1]=K21,x=1,y=8,},},arg={[K27]=K26,t={[K32]={f=K3,[K2]=1860,[K1]=K11,x=10,y=362,},f=K3,[K2]=1861,[K1]=K33,x=9,xend=16,y=362,yend=362,},},["assert"]={[K27]=K26,[K28]=true,t={f=K3,[K23]=true,t={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2495,[K1]=K5,[K5]="A@43",x=27,y=363,},[2]={f=K3,[K2]=2496,[K1]=K5,[K5]="B@43",x=32,y=363,},[3]={f=K3,[K2]=1867,[K1]="any",x=40,y=363,},},[K2]=2497,[K1]=K4,x=44,y=363,},f=K3,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2498,[K1]=K5,[K5]="A@43",x=46,y=363,},},[K2]=2499,[K1]=K4,x=44,y=363,},[K36]="assert",[K2]=2500,[K1]=K7,x=12,y=363,},[K19]={[1]={f=K3,[K10]="A@43",[K2]=2493,[K1]=K10,x=21,y=363,},[2]={f=K3,[K10]="B@43",[K2]=2494,[K1]=K10,x=24,y=363,},},[K2]=2501,[K1]=K20,x=4,y=365,},},["collectgarbage"]={[K27]=K26,t={f=K3,[K2]=1883,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K17]={def={[K24]="CollectGarbageCommand",[K38]={["collect"]=true,["count"]=true,["restart"]=true,stop=true,},f=K3,[K2]=1842,[K1]="enum",x=4,y=337,},f=K3,[K2]=1843,[K1]=K21,x=4,y=337,},[K16]={[1]="CollectGarbageCommand",},[K2]=1873,[K1]=K15,x=31,y=365,},},[K2]=1874,[K1]=K4,x=53,y=365,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1876,[K1]=K13,x=55,y=365,},},[K2]=1875,[K1]=K4,x=53,y=365,},[K2]=1872,[K1]=K7,x=20,y=365,},[2]={args={f=K3,[K4]={[1]={f=K3,[K17]={def={[K24]="CollectGarbageSetValue",[K38]={["setpause"]=true,["setstepmul"]=true,step=true,},f=K3,[K2]=1844,[K1]="enum",x=4,y=344,},f=K3,[K2]=1845,[K1]=K21,x=4,y=344,},[K16]={[1]="CollectGarbageSetValue",},[K2]=1878,[K1]=K15,x=29,y=366,},[2]={f=K3,[K2]=1879,[K1]=K12,x=53,y=366,},},[K2]=1880,[K1]=K4,x=61,y=366,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1882,[K1]=K13,x=63,y=366,},},[K2]=1881,[K1]=K4,x=61,y=366,},[K2]=1877,[K1]=K7,x=20,y=366,},[3]={args={f=K3,[K4]={[1]={f=K3,[K17]={def={[K24]="CollectGarbageIsRunning",[K38]={["isrunning"]=true,},f=K3,[K2]=1846,[K1]="enum",x=4,y=350,},f=K3,[K2]=1847,[K1]=K21,x=4,y=350,},[K16]={[1]="CollectGarbageIsRunning",},[K2]=1885,[K1]=K15,x=29,y=367,},},[K2]=1886,[K1]=K4,x=53,y=367,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1888,[K1]=K25,x=55,y=367,},},[K2]=1887,[K1]=K4,x=53,y=367,},[K2]=1884,[K1]=K7,x=20,y=367,},[4]={args={f=K3,[K4]={[1]={f=K3,[K2]=1890,[K1]=K11,x=29,y=368,},[2]={f=K3,[K2]=1891,[K1]=K13,x=39,y=368,},},[K2]=1892,[K1]=K4,x=46,y=368,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1895,[K1]=K22,[K18]={[1]={f=K3,[K2]=1894,[K1]=K25,x=49,y=368,},[2]={f=K3,[K2]=1896,[K1]=K13,x=59,y=368,},},x=49,y=368,},},[K2]=1893,[K1]=K4,x=46,y=368,},[K2]=1889,[K1]=K7,x=20,y=368,},},x=18,y=366,},},["coroutine"]={[K28]=true,t={[K34]=true,def={[K24]="coroutine",f=K3,[K31]={[1]=K[13],[2]=K43,[3]="create",[4]="isyieldable",[5]="resume",[6]="running",[7]="status",[8]="wrap",[9]="yield",},[K30]={[K[13]]=T7,[K43]={args={f=K3,[K4]={[1]={f=K3,[K2]=703,[K1]=K35,x=20,y=63,},},[K2]=704,[K1]=K4,x=27,y=63,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=706,[K1]=K25,x=29,y=63,},[2]={f=K3,[K2]=707,[K1]=K11,x=38,y=63,},},[K2]=705,[K1]=K4,x=27,y=63,},[K2]=702,[K1]=K7,x=11,y=63,},["create"]={args={f=K3,[K4]={[1]={f=K3,[K17]=T7,[K16]={[1]=K[13],},[K2]=709,[K1]=K15,x=21,y=64,},},[K2]=710,[K1]=K4,x=30,y=64,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=712,[K1]=K35,x=32,y=64,},},[K2]=711,[K1]=K4,x=30,y=64,},[K2]=708,[K1]=K7,x=12,y=64,},["isyieldable"]={args={f=K3,[K4]={},[K2]=714,[K1]=K4,x=27,y=65,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=716,[K1]=K25,x=29,y=65,},},[K2]=715,[K1]=K4,x=27,y=65,},[K2]=713,[K1]=K7,x=17,y=65,},["resume"]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=718,[K1]=K35,x=21,y=66,},[2]={f=K3,[K2]=719,[K1]="any",x=29,y=66,},},[K2]=720,[K1]=K4,x=36,y=66,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=722,[K1]=K25,x=38,y=66,},[2]={f=K3,[K2]=723,[K1]="any",x=47,y=66,},},[K2]=721,[K1]=K4,x=36,y=66,},[K2]=717,[K1]=K7,x=12,y=66,},["running"]={args={f=K3,[K4]={},[K2]=725,[K1]=K4,x=23,y=67,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=727,[K1]=K35,x=25,y=67,},[2]={f=K3,[K2]=728,[K1]=K25,x=33,y=67,},},[K2]=726,[K1]=K4,x=23,y=67,},[K2]=724,[K1]=K7,x=13,y=67,},["status"]={args={f=K3,[K4]={[1]={f=K3,[K2]=730,[K1]=K35,x=21,y=68,},},[K2]=731,[K1]=K4,x=28,y=68,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=733,[K1]=K11,x=30,y=68,},},[K2]=732,[K1]=K4,x=28,y=68,},[K2]=729,[K1]=K7,x=12,y=68,},wrap={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={f=K3,[K2]=2208,[K1]=K5,[K5]="F@23",x=22,y=69,},},[K2]=2209,[K1]=K4,x=24,y=69,},f=K3,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2210,[K1]=K5,[K5]="F@23",x=26,y=69,},},[K2]=2211,[K1]=K4,x=24,y=69,},[K2]=2212,[K1]=K7,x=10,y=69,},[K19]={[1]={f=K3,[K10]="F@23",[K2]=2207,[K1]=K10,x=19,y=69,},},[K2]=2213,[K1]=K20,x=4,y=70,},["yield"]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=742,[K1]="any",x=20,y=70,},},[K2]=743,[K1]=K4,x=27,y=70,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=745,[K1]="any",x=29,y=70,},},[K2]=744,[K1]=K4,x=27,y=70,},[K2]=741,[K1]=K7,x=11,y=70,},},[K2]=695,[K1]=K37,x=1,y=60,},f=K3,[K2]=746,[K1]=K21,x=1,y=60,},},[K[14]]={[K28]=true,t={[K34]=true,def={[K24]=K[14],f=K3,[K31]={[1]=K44,[2]=K[15],[3]=K48,[4]=K39,[5]=K[14],[6]="gethook",[7]="getinfo",[8]="getlocal",[9]=K[16],[10]="getregistry",[11]="getupvalue",[12]="getuservalue",[13]="sethook",[14]="setlocal",[15]=K[17],[16]="setupvalue",[17]="setuservalue",[18]="traceback",[19]="upvalueid",[20]="upvaluejoin",},[K30]={[K39]=T8,[K44]=T9,[K[15]]=T10,[K48]=T11,[K[14]]={args={f=K3,[K4]={},[K2]=781,[K1]=K4,x=4,y=100,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={},[K2]=782,[K1]=K4,x=20,y=99,},[K2]=780,[K1]=K7,x=11,y=99,},["gethook"]={args={f=K3,[K4]={[1]={f=K3,[K2]=784,[K1]=K35,x=24,y=100,},},[K2]=785,[K1]=K4,x=31,y=100,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K17]=T11,[K16]={[1]=K48,},[K2]=787,[K1]=K15,x=33,y=100,},[2]={f=K3,[K2]=788,[K1]=K12,x=47,y=100,},},[K2]=786,[K1]=K4,x=31,y=100,},[K2]=783,[K1]=K7,x=13,y=100,},["getinfo"]={f=K3,[K2]=804,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=791,[K1]=K22,[K18]={[1]={f=K3,[K17]=T8,[K16]={[1]=K39,},[K2]=790,[K1]=K15,x=22,y=102,},[2]={f=K3,[K2]=792,[K1]=K12,x=36,y=102,},},x=22,y=102,},},[K2]=793,[K1]=K4,x=44,y=102,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K17]=T9,[K16]={[1]=K44,},[K2]=795,[K1]=K15,x=46,y=102,},},[K2]=794,[K1]=K4,x=44,y=102,},[K2]=789,[K1]=K7,x=13,y=102,},[2]={args={f=K3,[K4]={[1]={f=K3,[K2]=798,[K1]=K22,[K18]={[1]={f=K3,[K17]=T8,[K16]={[1]=K39,},[K2]=797,[K1]=K15,x=22,y=103,},[2]={f=K3,[K2]=799,[K1]=K12,x=36,y=103,},},x=22,y=103,},[2]={f=K3,[K2]=800,[K1]=K11,x=45,y=103,},},[K2]=801,[K1]=K4,x=52,y=103,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K17]=T9,[K16]={[1]=K44,},[K2]=803,[K1]=K15,x=54,y=103,},},[K2]=802,[K1]=K4,x=52,y=103,},[K2]=796,[K1]=K7,x=13,y=103,},[3]={args={f=K3,[K4]={[1]={f=K3,[K2]=806,[K1]=K35,x=22,y=104,},[2]={f=K3,[K2]=808,[K1]=K22,[K18]={[1]={f=K3,[K17]=T8,[K16]={[1]=K39,},[K2]=807,[K1]=K15,x=30,y=104,},[2]={f=K3,[K2]=809,[K1]=K12,x=44,y=104,},},x=30,y=104,},[3]={f=K3,[K2]=810,[K1]=K11,x=53,y=104,},},[K2]=811,[K1]=K4,x=60,y=104,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,[K17]=T9,[K16]={[1]=K44,},[K2]=813,[K1]=K15,x=62,y=104,},},[K2]=812,[K1]=K4,x=60,y=104,},[K2]=805,[K1]=K7,x=13,y=104,},},x=11,y=103,},["getlocal"]={f=K3,[K2]=829,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=815,[K1]=K35,x=23,y=106,},[2]={f=K3,[K17]=T8,[K16]={[1]=K39,},[K2]=816,[K1]=K15,x=31,y=106,},[3]={f=K3,[K2]=817,[K1]=K12,x=44,y=106,},},[K2]=818,[K1]=K4,x=52,y=106,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,[K2]=820,[K1]=K11,x=54,y=106,},},[K2]=819,[K1]=K4,x=52,y=106,},[K2]=814,[K1]=K7,x=14,y=106,},[2]={args={f=K3,[K4]={[1]={f=K3,[K2]=822,[K1]=K35,x=23,y=107,},[2]={f=K3,[K2]=823,[K1]=K12,x=31,y=107,},[3]={f=K3,[K2]=824,[K1]=K12,x=40,y=107,},},[K2]=825,[K1]=K4,x=48,y=107,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,[K2]=827,[K1]=K11,x=50,y=107,},[2]={f=K3,[K2]=828,[K1]="any",x=58,y=107,},},[K2]=826,[K1]=K4,x=48,y=107,},[K2]=821,[K1]=K7,x=14,y=107,},[3]={args={f=K3,[K4]={[1]={f=K3,[K17]=T8,[K16]={[1]=K39,},[K2]=831,[K1]=K15,x=23,y=108,},[2]={f=K3,[K2]=832,[K1]=K12,x=36,y=108,},},[K2]=833,[K1]=K4,x=44,y=108,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=835,[K1]=K11,x=46,y=108,},},[K2]=834,[K1]=K4,x=44,y=108,},[K2]=830,[K1]=K7,x=14,y=108,},[4]={args={f=K3,[K4]={[1]={f=K3,[K2]=837,[K1]=K12,x=23,y=109,},[2]={f=K3,[K2]=838,[K1]=K12,x=32,y=109,},},[K2]=839,[K1]=K4,x=40,y=109,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=841,[K1]=K11,x=42,y=109,},[2]={f=K3,[K2]=842,[K1]="any",x=50,y=109,},},[K2]=840,[K1]=K4,x=40,y=109,},[K2]=836,[K1]=K7,x=14,y=109,},},x=12,y=107,},[K[16]]={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={f=K3,[K2]=2220,[K1]=K5,[K5]="T@24",x=30,y=111,},},[K2]=2221,[K1]=K4,x=32,y=111,},f=K3,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K17]=T12,[K16]={[1]=K42,},[K2]=2223,[K1]=K15,[K41]={[1]={f=K3,[K2]=2222,[K1]=K5,[K5]="T@24",x=44,y=111,},},x=34,y=111,},},[K2]=2224,[K1]=K4,x=32,y=111,},[K2]=2225,[K1]=K7,x=18,y=111,},[K19]={[1]={f=K3,[K10]="T@24",[K2]=2219,[K1]=K10,x=27,y=111,},},[K2]=2226,[K1]=K20,x=4,y=112,},["getregistry"]={args={f=K3,[K4]={},[K2]=852,[K1]=K4,x=27,y=112,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,keys={f=K3,[K2]=854,[K1]="any",x=30,y=112,},[K2]=855,[K1]="map",[K29]={f=K3,[K2]=856,[K1]="any",x=34,y=112,},x=29,xend=37,y=112,yend=112,},},[K2]=853,[K1]=K4,x=27,y=112,},[K2]=851,[K1]=K7,x=17,y=112,},["getupvalue"]={args={f=K3,[K4]={[1]={f=K3,[K17]=T8,[K16]={[1]=K39,},[K2]=858,[K1]=K15,x=25,y=113,},[2]={f=K3,[K2]=859,[K1]=K12,x=38,y=113,},},[K2]=860,[K1]=K4,x=46,y=113,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=862,[K1]="any",x=48,y=113,},},[K2]=861,[K1]=K4,x=46,y=113,},[K2]=857,[K1]=K7,x=16,y=113,},["getuservalue"]={args={f=K3,[K4]={[1]={f=K3,[K17]=T14,[K16]={[1]=K[20],},[K2]=864,[K1]=K15,x=27,y=114,},[2]={f=K3,[K2]=865,[K1]=K12,x=37,y=114,},},[K2]=866,[K1]=K4,x=45,y=114,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=868,[K1]="any",x=47,y=114,},},[K2]=867,[K1]=K4,x=45,y=114,},[K2]=863,[K1]=K7,x=18,y=114,},["sethook"]={f=K3,[K2]=882,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=870,[K1]=K35,x=22,y=116,},[2]={f=K3,[K17]=T11,[K16]={[1]=K48,},[K2]=871,[K1]=K15,x=30,y=116,},[3]={f=K3,[K2]=872,[K1]=K11,x=44,y=116,},[4]={f=K3,[K2]=873,[K1]=K12,x=54,y=116,},},[K2]=874,[K1]=K4,x=4,y=117,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={},[K2]=875,[K1]=K4,x=61,y=116,},[K2]=869,[K1]=K7,x=13,y=116,},[2]={args={f=K3,[K4]={[1]={f=K3,[K17]=T11,[K16]={[1]=K48,},[K2]=877,[K1]=K15,x=22,y=117,},[2]={f=K3,[K2]=878,[K1]=K11,x=36,y=117,},[3]={f=K3,[K2]=879,[K1]=K12,x=46,y=117,},},[K2]=880,[K1]=K4,x=4,y=119,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={},[K2]=881,[K1]=K4,x=53,y=117,},[K2]=876,[K1]=K7,x=13,y=117,},},x=11,y=117,},["setlocal"]={f=K3,[K2]=898,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=884,[K1]=K35,x=23,y=119,},[2]={f=K3,[K2]=885,[K1]=K12,x=31,y=119,},[3]={f=K3,[K2]=886,[K1]=K12,x=40,y=119,},[4]={f=K3,[K2]=887,[K1]="any",x=49,y=119,},},[K2]=888,[K1]=K4,x=53,y=119,},f=K3,[K9]=false,[K6]=4,rets={f=K3,[K4]={[1]={f=K3,[K2]=890,[K1]=K11,x=55,y=119,},},[K2]=889,[K1]=K4,x=53,y=119,},[K2]=883,[K1]=K7,x=14,y=119,},[2]={args={f=K3,[K4]={[1]={f=K3,[K2]=892,[K1]=K12,x=23,y=120,},[2]={f=K3,[K2]=893,[K1]=K12,x=32,y=120,},[3]={f=K3,[K2]=894,[K1]="any",x=41,y=120,},},[K2]=895,[K1]=K4,x=45,y=120,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,[K2]=897,[K1]=K11,x=47,y=120,},},[K2]=896,[K1]=K4,x=45,y=120,},[K2]=891,[K1]=K7,x=14,y=120,},},x=12,y=120,},[K[17]]={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={f=K3,[K2]=2232,[K1]=K5,[K5]="T@25",x=30,y=122,},[2]={f=K3,[K17]=T12,[K16]={[1]=K42,},[K2]=2234,[K1]=K15,[K41]={[1]={f=K3,[K2]=2233,[K1]=K5,[K5]="T@25",x=43,y=122,},},x=33,y=122,},},[K2]=2235,[K1]=K4,x=46,y=122,},f=K3,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=2236,[K1]=K5,[K5]="T@25",x=48,y=122,},},[K2]=2237,[K1]=K4,x=46,y=122,},[K2]=2238,[K1]=K7,x=18,y=122,},[K19]={[1]={f=K3,[K10]="T@25",[K2]=2231,[K1]=K10,x=27,y=122,},},[K2]=2239,[K1]=K20,x=4,y=123,},["setupvalue"]={args={f=K3,[K4]={[1]={f=K3,[K17]=T8,[K16]={[1]=K39,},[K2]=909,[K1]=K15,x=25,y=123,},[2]={f=K3,[K2]=910,[K1]=K12,x=38,y=123,},[3]={f=K3,[K2]=911,[K1]="any",x=47,y=123,},},[K2]=912,[K1]=K4,x=51,y=123,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,[K2]=914,[K1]=K11,x=53,y=123,},},[K2]=913,[K1]=K4,x=51,y=123,},[K2]=908,[K1]=K7,x=16,y=123,},["setuservalue"]={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={f=K3,[K2]=2244,[K1]=K5,[K5]="U@26",x=30,y=124,},[2]={f=K3,[K2]=918,[K1]="any",x=33,y=124,},[3]={f=K3,[K2]=919,[K1]=K12,x=38,y=124,},},[K2]=2245,[K1]=K4,x=46,y=124,},f=K3,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,[K2]=2246,[K1]=K5,[K5]="U@26",x=48,y=124,},},[K2]=2247,[K1]=K4,x=46,y=124,},[K2]=2248,[K1]=K7,x=18,y=124,},[K19]={[1]={f=K3,[K10]="U@26",[K2]=2243,[K1]=K10,x=27,y=124,},},[K2]=2249,[K1]=K20,x=4,y=126,},["traceback"]={f=K3,[K2]=937,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=925,[K1]=K35,x=24,y=126,},[2]={f=K3,[K2]=926,[K1]=K11,x=34,y=126,},[3]={f=K3,[K2]=927,[K1]=K12,x=44,y=126,},},[K2]=928,[K1]=K4,x=52,y=126,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=930,[K1]=K11,x=54,y=126,},},[K2]=929,[K1]=K4,x=52,y=126,},[K2]=924,[K1]=K7,x=15,y=126,},[2]={args={f=K3,[K4]={[1]={f=K3,[K2]=932,[K1]=K11,x=26,y=127,},[2]={f=K3,[K2]=933,[K1]=K12,x=36,y=127,},},[K2]=934,[K1]=K4,x=44,y=127,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=936,[K1]=K11,x=46,y=127,},},[K2]=935,[K1]=K4,x=44,y=127,},[K2]=931,[K1]=K7,x=15,y=127,},},x=13,y=127,},["upvalueid"]={args={f=K3,[K4]={[1]={f=K3,[K17]=T8,[K16]={[1]=K39,},[K2]=939,[K1]=K15,x=24,y=129,},[2]={f=K3,[K2]=940,[K1]=K12,x=37,y=129,},},[K2]=941,[K1]=K4,x=45,y=129,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K17]=T14,[K16]={[1]=K[20],},[K2]=943,[K1]=K15,x=47,y=129,},},[K2]=942,[K1]=K4,x=45,y=129,},[K2]=938,[K1]=K7,x=15,y=129,},["upvaluejoin"]={args={f=K3,[K4]={[1]={f=K3,[K17]=T8,[K16]={[1]=K39,},[K2]=945,[K1]=K15,x=26,y=130,},[2]={f=K3,[K2]=946,[K1]=K12,x=39,y=130,},[3]={f=K3,[K17]=T8,[K16]={[1]=K39,},[K2]=947,[K1]=K15,x=48,y=130,},[4]={f=K3,[K2]=948,[K1]=K12,x=61,y=130,},},[K2]=949,[K1]=K4,x=1,y=131,},f=K3,[K9]=false,[K6]=4,rets={f=K3,[K4]={},[K2]=950,[K1]=K4,x=68,y=130,},[K2]=944,[K1]=K7,x=17,y=130,},},[K2]=747,[K1]=K37,x=1,y=73,},f=K3,[K2]=951,[K1]=K21,x=1,y=73,},},["dofile"]={[K27]=K26,t={args={f=K3,[K4]={[1]={f=K3,[K2]=1898,[K1]=K11,x=23,y=370,},},[K2]=1899,[K1]=K4,x=30,y=370,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1901,[K1]="any",x=32,y=370,},},[K2]=1900,[K1]=K4,x=30,y=370,},[K2]=1897,[K1]=K7,x=12,y=370,},},["error"]={[K27]=K26,t={args={f=K3,[K4]={[1]={f=K3,[K2]=1903,[K1]="any",x=22,y=372,},[2]={f=K3,[K2]=1904,[K1]=K12,x=29,y=372,},},[K2]=1905,[K1]=K4,x=4,y=373,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={},[K2]=1906,[K1]=K4,x=36,y=372,},[K2]=1902,[K1]=K7,x=11,y=372,},},[K[16]]={[K27]=K26,t={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={f=K3,[K2]=2506,[K1]=K5,[K5]="T@44",x=30,y=373,},},[K2]=2507,[K1]=K4,x=32,y=373,},f=K3,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K17]=T12,[K16]={[1]=K42,},[K2]=2509,[K1]=K15,[K41]={[1]={f=K3,[K2]=2508,[K1]=K5,[K5]="T@44",x=44,y=373,},},x=34,y=373,},},[K2]=2510,[K1]=K4,x=32,y=373,},[K2]=2511,[K1]=K7,x=18,y=373,},[K19]={[1]={f=K3,[K10]="T@44",[K2]=2505,[K1]=K10,x=27,y=373,},},[K2]=2512,[K1]=K20,x=4,y=374,},},io={[K28]=true,t={[K34]=true,def={[K24]="io",f=K3,[K31]={[1]=K47,[2]=K43,[3]="input",[4]=K[4],[5]=K[5],[6]="open",[7]="output",[8]="popen",[9]="read",[10]="stderr",[11]="stdin",[12]="stdout",[13]="tmpfile",[14]="type",[15]=K[6],},[K30]={[K47]=T15,[K43]={args={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=956,[K1]=K15,x=22,y=141,},},[K2]=957,[K1]=K4,x=4,y=142,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={},[K2]=958,[K1]=K4,x=26,y=141,},[K2]=955,[K1]=K7,x=11,y=141,},[K[4]]={args={f=K3,[K4]={},[K2]=967,[K1]=K4,x=4,y=145,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={},[K2]=968,[K1]=K4,x=20,y=143,},[K2]=966,[K1]=K7,x=11,y=143,},["input"]={args={f=K3,[K4]={[1]={f=K3,[K2]=961,[K1]=K22,[K18]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=960,[K1]=K15,x=22,y=142,},[2]={f=K3,[K2]=962,[K1]=K11,x=29,y=142,},},x=22,y=142,},},[K2]=963,[K1]=K4,x=36,y=142,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=965,[K1]=K15,x=38,y=142,},},[K2]=964,[K1]=K4,x=36,y=142,},[K2]=959,[K1]=K7,x=11,y=142,},[K[5]]={f=K3,[K2]=986,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=970,[K1]=K11,x=22,y=145,},},[K2]=971,[K1]=K4,x=29,y=145,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=974,[K1]=K4,x=42,y=145,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=976,[K1]=K11,x=45,y=145,},},[K2]=975,[K1]=K4,x=42,y=145,},[K2]=973,[K1]=K7,x=32,y=145,},},[K2]=972,[K1]=K4,x=29,y=145,},[K2]=969,[K1]=K7,x=11,y=145,},[2]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=978,[K1]=K11,x=22,y=146,},[2]={f=K3,[K17]=T4,[K16]={[1]=K50,},[K2]=979,[K1]=K15,x=30,y=146,},},[K2]=980,[K1]=K4,x=48,y=146,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=983,[K1]=K4,x=61,y=146,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=985,[K1]=K13,x=64,y=146,},},[K2]=984,[K1]=K4,x=61,y=146,},[K2]=982,[K1]=K7,x=51,y=146,},},[K2]=981,[K1]=K4,x=48,y=146,},[K2]=977,[K1]=K7,x=11,y=146,},[3]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=988,[K1]=K11,x=22,y=147,},[2]={f=K3,[K2]=990,[K1]=K22,[K18]={[1]={f=K3,[K2]=989,[K1]=K13,x=31,y=147,},[2]={f=K3,[K17]=T5,[K16]={[1]=K49,},[K2]=991,[K1]=K15,x=40,y=147,},},x=31,y=147,},},[K2]=992,[K1]=K4,x=59,y=147,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=995,[K1]=K4,x=72,y=147,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=997,[K1]=K11,x=75,y=147,},},[K2]=996,[K1]=K4,x=72,y=147,},[K2]=994,[K1]=K7,x=62,y=147,},},[K2]=993,[K1]=K4,x=59,y=147,},[K2]=987,[K1]=K7,x=11,y=147,},[4]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=999,[K1]=K11,x=22,y=148,},[2]={f=K3,[K2]=1001,[K1]=K22,[K18]={[1]={f=K3,[K2]=1000,[K1]=K13,x=31,y=148,},[2]={f=K3,[K17]=T6,[K16]={[1]=K[7],},[K2]=1002,[K1]=K15,x=40,y=148,},},x=31,y=148,},},[K2]=1003,[K1]=K4,x=53,y=148,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=1006,[K1]=K4,x=66,y=148,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1009,[K1]=K22,[K18]={[1]={f=K3,[K2]=1008,[K1]=K11,x=70,y=148,},[2]={f=K3,[K2]=1010,[K1]=K13,x=79,y=148,},},x=70,y=148,},},[K2]=1007,[K1]=K4,x=66,y=148,},[K2]=1005,[K1]=K7,x=56,y=148,},},[K2]=1004,[K1]=K4,x=53,y=148,},[K2]=998,[K1]=K7,x=11,y=148,},[5]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1012,[K1]=K11,x=22,y=149,},[2]={f=K3,[K2]=1014,[K1]=K22,[K18]={[1]={f=K3,[K2]=1013,[K1]=K13,x=31,y=149,},[2]={f=K3,[K2]=1015,[K1]=K11,x=40,y=149,},},x=31,y=149,},},[K2]=1016,[K1]=K4,x=51,y=149,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=1019,[K1]=K4,x=64,y=149,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1021,[K1]=K11,x=67,y=149,},},[K2]=1020,[K1]=K4,x=64,y=149,},[K2]=1018,[K1]=K7,x=54,y=149,},},[K2]=1017,[K1]=K4,x=51,y=149,},[K2]=1011,[K1]=K7,x=11,y=149,},},x=9,y=146,},open={args={f=K3,[K4]={[1]={f=K3,[K2]=1023,[K1]=K11,x=19,y=151,},[2]={f=K3,[K17]=T15,[K16]={[1]=K47,},[K2]=1024,[K1]=K15,x=29,y=151,},},[K2]=1025,[K1]=K4,x=38,y=151,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=1027,[K1]=K15,x=40,y=151,},[2]={f=K3,[K2]=1028,[K1]=K11,x=46,y=151,},[3]={f=K3,[K2]=1029,[K1]=K12,x=54,y=151,},},[K2]=1026,[K1]=K4,x=38,y=151,},[K2]=1022,[K1]=K7,x=10,y=151,},["output"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1032,[K1]=K22,[K18]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=1031,[K1]=K15,x=23,y=152,},[2]={f=K3,[K2]=1033,[K1]=K11,x=30,y=152,},},x=23,y=152,},},[K2]=1034,[K1]=K4,x=37,y=152,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=1036,[K1]=K15,x=39,y=152,},},[K2]=1035,[K1]=K4,x=37,y=152,},[K2]=1030,[K1]=K7,x=12,y=152,},["popen"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1038,[K1]=K11,x=20,y=153,},[2]={f=K3,[K17]=T15,[K16]={[1]=K47,},[K2]=1039,[K1]=K15,x=30,y=153,},},[K2]=1040,[K1]=K4,x=39,y=153,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=1042,[K1]=K15,x=41,y=153,},[2]={f=K3,[K2]=1043,[K1]=K11,x=47,y=153,},},[K2]=1041,[K1]=K4,x=39,y=153,},[K2]=1037,[K1]=K7,x=11,y=153,},read={f=K3,[K2]=1053,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={},[K2]=1045,[K1]=K4,x=20,y=155,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1047,[K1]=K11,x=22,y=155,},},[K2]=1046,[K1]=K4,x=20,y=155,},[K2]=1044,[K1]=K7,x=10,y=155,},[2]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K17]=T4,[K16]={[1]=K50,},[K2]=1049,[K1]=K15,x=19,y=156,},},[K2]=1050,[K1]=K4,x=37,y=156,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1052,[K1]=K13,x=39,y=156,},},[K2]=1051,[K1]=K4,x=37,y=156,},[K2]=1048,[K1]=K7,x=10,y=156,},[3]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1056,[K1]=K22,[K18]={[1]={f=K3,[K2]=1055,[K1]=K13,x=20,y=157,},[2]={f=K3,[K17]=T5,[K16]={[1]=K49,},[K2]=1057,[K1]=K15,x=29,y=157,},},x=20,y=157,},},[K2]=1058,[K1]=K4,x=48,y=157,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1060,[K1]=K11,x=50,y=157,},},[K2]=1059,[K1]=K4,x=48,y=157,},[K2]=1054,[K1]=K7,x=10,y=157,},[4]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1063,[K1]=K22,[K18]={[1]={f=K3,[K2]=1062,[K1]=K13,x=20,y=158,},[2]={f=K3,[K17]=T6,[K16]={[1]=K[7],},[K2]=1064,[K1]=K15,x=29,y=158,},},x=20,y=158,},},[K2]=1065,[K1]=K4,x=42,y=158,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1068,[K1]=K22,[K18]={[1]={f=K3,[K2]=1067,[K1]=K11,x=46,y=158,},[2]={f=K3,[K2]=1069,[K1]=K13,x=55,y=158,},},x=46,y=158,},},[K2]=1066,[K1]=K4,x=42,y=158,},[K2]=1061,[K1]=K7,x=10,y=158,},[5]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1072,[K1]=K22,[K18]={[1]={f=K3,[K2]=1071,[K1]=K13,x=20,y=159,},[2]={f=K3,[K2]=1073,[K1]=K11,x=29,y=159,},},x=20,y=159,},},[K2]=1074,[K1]=K4,x=40,y=159,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1076,[K1]=K11,x=43,y=159,},},[K2]=1075,[K1]=K4,x=40,y=159,},[K2]=1070,[K1]=K7,x=10,y=159,},},x=8,y=156,},["stderr"]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=1077,[K1]=K15,x=12,y=161,},["stdin"]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=1078,[K1]=K15,x=11,y=162,},["stdout"]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=1079,[K1]=K15,x=12,y=163,},["tmpfile"]={args={f=K3,[K4]={},[K2]=1081,[K1]=K4,x=23,y=164,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=1083,[K1]=K15,x=25,y=164,},},[K2]=1082,[K1]=K4,x=23,y=164,},[K2]=1080,[K1]=K7,x=13,y=164,},type={args={f=K3,[K4]={[1]={f=K3,[K2]=1085,[K1]="any",x=19,y=165,},},[K2]=1086,[K1]=K4,x=23,y=165,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1088,[K1]=K11,x=25,y=165,},},[K2]=1087,[K1]=K4,x=23,y=165,},[K2]=1084,[K1]=K7,x=10,y=165,},[K[6]]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1091,[K1]=K22,[K18]={[1]={f=K3,[K2]=1090,[K1]=K11,x=21,y=166,},[2]={f=K3,[K2]=1092,[K1]=K13,x=30,y=166,},},x=21,y=166,},},[K2]=1093,[K1]=K4,x=41,y=166,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=1095,[K1]=K15,x=43,y=166,},[2]={f=K3,[K2]=1096,[K1]=K11,x=49,y=166,},},[K2]=1094,[K1]=K4,x=41,y=166,},[K2]=1089,[K1]=K7,x=11,y=166,},},[K2]=952,[K1]=K37,x=1,y=133,},f=K3,[K2]=1097,[K1]=K21,x=1,y=133,},},["ipairs"]={[K27]=K26,[K28]=true,t={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={[K32]={f=K3,[K2]=2517,[K1]=K5,[K5]="A@45",x=25,y=374,},f=K3,[K2]=2518,[K1]=K33,x=24,y=374,},},[K2]=2519,[K1]=K4,x=28,y=374,},f=K3,[K6]=1,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=1922,[K1]=K4,x=41,y=374,},f=K3,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1924,[K1]=K12,x=43,y=374,},[2]={f=K3,[K2]=2520,[K1]=K5,[K5]="A@45",x=52,y=374,},},[K2]=2521,[K1]=K4,x=41,y=374,},[K2]=2522,[K1]=K7,x=31,y=374,},},[K2]=2523,[K1]=K4,x=28,y=374,},[K36]="ipairs",[K2]=2524,[K1]=K7,x=12,y=374,},[K19]={[1]={f=K3,[K10]="A@45",[K2]=2516,[K1]=K10,x=21,y=374,},},[K2]=2525,[K1]=K20,x=4,y=376,},},load={[K27]=K26,[K28]=true,t={f=K3,[K2]=1961,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=1929,[K1]=K22,[K18]={[1]={f=K3,[K2]=1928,[K1]=K11,x=20,y=376,},[2]={f=K3,[K17]=T16,[K16]={[1]="LoadFunction",},[K2]=1930,[K1]=K15,x=29,y=376,},},x=20,y=376,},[2]={f=K3,[K2]=1931,[K1]=K11,x=46,y=376,},[3]={f=K3,[K17]={def={[K24]="LoadMode",[K38]={b=true,bt=true,t=true,},f=K3,[K2]=1853,[K1]="enum",x=4,y=356,},f=K3,[K2]=1854,[K1]=K21,x=4,y=356,},[K16]={[1]="LoadMode",},[K2]=1932,[K1]=K15,x=56,y=376,},[4]={f=K3,keys={f=K3,[K2]=1934,[K1]="any",x=68,y=376,},[K2]=1933,[K1]="map",[K29]={f=K3,[K2]=1935,[K1]="any",x=68,y=376,},x=68,y=376,},},[K2]=1936,[K1]=K4,x=74,y=376,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1939,[K1]="any",x=85,y=376,},},[K2]=1940,[K1]=K4,x=85,y=376,},f=K3,[K46]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1941,[K1]="any",x=85,y=376,},},[K2]=1942,[K1]=K4,x=85,y=376,},[K2]=1938,[K1]=K7,x=77,y=376,},[2]={f=K3,[K2]=1943,[K1]=K11,x=87,y=376,},},[K2]=1937,[K1]=K4,x=74,y=376,},[K2]=1927,[K1]=K7,x=10,y=376,},[2]={args={f=K3,[K4]={[1]={f=K3,[K2]=1946,[K1]=K22,[K18]={[1]={f=K3,[K2]=1945,[K1]=K11,x=20,y=377,},[2]={f=K3,[K17]=T16,[K16]={[1]="LoadFunction",},[K2]=1947,[K1]=K15,x=29,y=377,},},x=20,y=377,},[2]={f=K3,[K2]=1948,[K1]=K11,x=46,y=377,},[3]={f=K3,[K2]=1949,[K1]=K11,x=56,y=377,},[4]={f=K3,keys={f=K3,[K2]=1951,[K1]="any",x=66,y=377,},[K2]=1950,[K1]="map",[K29]={f=K3,[K2]=1952,[K1]="any",x=66,y=377,},x=66,y=377,},},[K2]=1953,[K1]=K4,x=72,y=377,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1956,[K1]="any",x=83,y=377,},},[K2]=1957,[K1]=K4,x=83,y=377,},f=K3,[K46]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1958,[K1]="any",x=83,y=377,},},[K2]=1959,[K1]=K4,x=83,y=377,},[K2]=1955,[K1]=K7,x=75,y=377,},[2]={f=K3,[K2]=1960,[K1]=K11,x=85,y=377,},},[K2]=1954,[K1]=K4,x=72,y=377,},[K2]=1944,[K1]=K7,x=10,y=377,},},x=8,y=377,},},["loadfile"]={[K27]=K26,[K28]=true,t={args={f=K3,[K4]={[1]={f=K3,[K2]=1963,[K1]=K11,x=25,y=379,},[2]={f=K3,[K2]=1964,[K1]=K11,x=35,y=379,},[3]={f=K3,keys={f=K3,[K2]=1965,[K1]="any",x=46,y=379,},[K2]=1966,[K1]="map",[K29]={f=K3,[K2]=1967,[K1]="any",x=50,y=379,},x=45,xend=53,y=379,yend=379,},},[K2]=1968,[K1]=K4,x=55,y=379,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1971,[K1]="any",x=66,y=379,},},[K2]=1972,[K1]=K4,x=66,y=379,},f=K3,[K46]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1973,[K1]="any",x=66,y=379,},},[K2]=1974,[K1]=K4,x=66,y=379,},[K2]=1970,[K1]=K7,x=58,y=379,},[2]={f=K3,[K2]=1975,[K1]=K11,x=68,y=379,},},[K2]=1969,[K1]=K4,x=55,y=379,},[K2]=1962,[K1]=K7,x=14,y=379,},},math={[K28]=true,t={[K34]=true,def={[K24]="math",f=K3,[K31]={[1]=K[21],[2]="abs",[3]="acos",[4]="asin",[5]="atan",[6]="atan2",[7]="ceil",[8]="cos",[9]="cosh",[10]="deg",[11]="exp",[12]="floor",[13]="fmod",[14]="frexp",[15]="huge",[16]="ldexp",[17]="log",[18]="log10",[19]="max",[20]="maxinteger",[21]="min",[22]="mininteger",[23]="modf",[24]="pi",[25]="pow",[26]="rad",[27]="random",[28]="randomseed",[29]="sin",[30]="sinh",[31]="sqrt",[32]="tan",[33]="tanh",[34]="tointeger",[35]="type",[36]="ult",},[K30]={[K[21]]=T17,abs={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={[K[22]]=T18,f=K3,[K2]=2256,[K1]=K5,[K5]="N@27",x=32,y=172,},},[K2]=2257,[K1]=K4,x=34,y=172,},f=K3,[K6]=1,rets={f=K3,[K4]={[1]={[K[22]]=T18,f=K3,[K2]=2258,[K1]=K5,[K5]="N@27",x=36,y=172,},},[K2]=2259,[K1]=K4,x=34,y=172,},[K2]=2260,[K1]=K7,x=9,y=172,},[K19]={[1]={[K[22]]=T18,f=K3,[K10]="N@27",[K2]=2255,[K1]=K10,x=18,y=172,},},[K2]=2261,[K1]=K20,x=4,y=173,},acos={args={f=K3,[K4]={[1]={f=K3,[K2]=1112,[K1]=K13,x=19,y=173,},},[K2]=1113,[K1]=K4,x=26,y=173,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1115,[K1]=K13,x=28,y=173,},},[K2]=1114,[K1]=K4,x=26,y=173,},[K2]=1111,[K1]=K7,x=10,y=173,},asin={args={f=K3,[K4]={[1]={f=K3,[K2]=1117,[K1]=K13,x=19,y=174,},},[K2]=1118,[K1]=K4,x=26,y=174,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1120,[K1]=K13,x=28,y=174,},},[K2]=1119,[K1]=K4,x=26,y=174,},[K2]=1116,[K1]=K7,x=10,y=174,},atan={args={f=K3,[K4]={[1]={f=K3,[K2]=1122,[K1]=K13,x=19,y=175,},[2]={f=K3,[K2]=1123,[K1]=K13,x=29,y=175,},},[K2]=1124,[K1]=K4,x=36,y=175,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1126,[K1]=K13,x=38,y=175,},},[K2]=1125,[K1]=K4,x=36,y=175,},[K2]=1121,[K1]=K7,x=10,y=175,},["atan2"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1128,[K1]=K13,x=20,y=176,},[2]={f=K3,[K2]=1129,[K1]=K13,x=28,y=176,},},[K2]=1130,[K1]=K4,x=35,y=176,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1132,[K1]=K13,x=37,y=176,},},[K2]=1131,[K1]=K4,x=35,y=176,},[K2]=1127,[K1]=K7,x=11,y=176,},ceil={args={f=K3,[K4]={[1]={f=K3,[K2]=1134,[K1]=K13,x=19,y=177,},},[K2]=1135,[K1]=K4,x=26,y=177,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1137,[K1]=K12,x=28,y=177,},},[K2]=1136,[K1]=K4,x=26,y=177,},[K2]=1133,[K1]=K7,x=10,y=177,},cos={args={f=K3,[K4]={[1]={f=K3,[K2]=1139,[K1]=K13,x=18,y=178,},},[K2]=1140,[K1]=K4,x=25,y=178,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1142,[K1]=K13,x=27,y=178,},},[K2]=1141,[K1]=K4,x=25,y=178,},[K2]=1138,[K1]=K7,x=9,y=178,},cosh={args={f=K3,[K4]={[1]={f=K3,[K2]=1144,[K1]=K13,x=19,y=179,},},[K2]=1145,[K1]=K4,x=26,y=179,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1147,[K1]=K13,x=28,y=179,},},[K2]=1146,[K1]=K4,x=26,y=179,},[K2]=1143,[K1]=K7,x=10,y=179,},deg={args={f=K3,[K4]={[1]={f=K3,[K2]=1149,[K1]=K13,x=18,y=180,},},[K2]=1150,[K1]=K4,x=25,y=180,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1152,[K1]=K13,x=27,y=180,},},[K2]=1151,[K1]=K4,x=25,y=180,},[K2]=1148,[K1]=K7,x=9,y=180,},exp={args={f=K3,[K4]={[1]={f=K3,[K2]=1154,[K1]=K13,x=18,y=181,},},[K2]=1155,[K1]=K4,x=25,y=181,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1157,[K1]=K13,x=27,y=181,},},[K2]=1156,[K1]=K4,x=25,y=181,},[K2]=1153,[K1]=K7,x=9,y=181,},["floor"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1159,[K1]=K13,x=20,y=182,},},[K2]=1160,[K1]=K4,x=27,y=182,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1162,[K1]=K12,x=29,y=182,},},[K2]=1161,[K1]=K4,x=27,y=182,},[K2]=1158,[K1]=K7,x=11,y=182,},fmod={f=K3,[K2]=1175,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=1164,[K1]=K12,x=19,y=184,},[2]={f=K3,[K2]=1165,[K1]=K12,x=28,y=184,},},[K2]=1166,[K1]=K4,x=36,y=184,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1168,[K1]=K12,x=38,y=184,},},[K2]=1167,[K1]=K4,x=36,y=184,},[K2]=1163,[K1]=K7,x=10,y=184,},[2]={args={f=K3,[K4]={[1]={f=K3,[K2]=1170,[K1]=K13,x=19,y=185,},[2]={f=K3,[K2]=1171,[K1]=K13,x=27,y=185,},},[K2]=1172,[K1]=K4,x=34,y=185,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1174,[K1]=K13,x=36,y=185,},},[K2]=1173,[K1]=K4,x=34,y=185,},[K2]=1169,[K1]=K7,x=10,y=185,},},x=8,y=185,},["frexp"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1177,[K1]=K13,x=20,y=187,},},[K2]=1178,[K1]=K4,x=27,y=187,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1180,[K1]=K13,x=29,y=187,},[2]={f=K3,[K2]=1181,[K1]=K12,x=37,y=187,},},[K2]=1179,[K1]=K4,x=27,y=187,},[K2]=1176,[K1]=K7,x=11,y=187,},huge={f=K3,[K2]=1182,[K1]=K13,x=10,y=188,},["ldexp"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1184,[K1]=K13,x=20,y=189,},[2]={f=K3,[K2]=1185,[K1]=K12,x=28,y=189,},},[K2]=1186,[K1]=K4,x=36,y=189,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1188,[K1]=K13,x=38,y=189,},},[K2]=1187,[K1]=K4,x=36,y=189,},[K2]=1183,[K1]=K7,x=11,y=189,},log={args={f=K3,[K4]={[1]={f=K3,[K2]=1190,[K1]=K13,x=18,y=190,},[2]={f=K3,[K2]=1191,[K1]=K13,x=28,y=190,},},[K2]=1192,[K1]=K4,x=35,y=190,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1194,[K1]=K13,x=37,y=190,},},[K2]=1193,[K1]=K4,x=35,y=190,},[K2]=1189,[K1]=K7,x=9,y=190,},["log10"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1196,[K1]=K13,x=20,y=191,},},[K2]=1197,[K1]=K4,x=27,y=191,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1199,[K1]=K13,x=29,y=191,},},[K2]=1198,[K1]=K4,x=27,y=191,},[K2]=1195,[K1]=K7,x=11,y=191,},max={f=K3,[K2]=1212,[K1]="poly",[K18]={[1]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1201,[K1]=K12,x=18,y=193,},},[K2]=1202,[K1]=K4,x=29,y=193,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1204,[K1]=K12,x=31,y=193,},},[K2]=1203,[K1]=K4,x=29,y=193,},[K2]=1200,[K1]=K7,x=9,y=193,},[2]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1207,[K1]=K22,[K18]={[1]={f=K3,[K2]=1206,[K1]=K13,x=19,y=194,},[2]={f=K3,[K2]=1208,[K1]=K12,x=28,y=194,},},x=19,y=194,},},[K2]=1209,[K1]=K4,x=40,y=194,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1211,[K1]=K13,x=42,y=194,},},[K2]=1210,[K1]=K4,x=40,y=194,},[K2]=1205,[K1]=K7,x=9,y=194,},[3]={f=K3,t={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2263,[K1]=K5,[K5]="T",x=21,y=195,},},[K2]=1216,[K1]=K4,x=26,y=195,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=2264,[K1]=K5,[K5]="T",x=28,y=195,},},[K2]=1217,[K1]=K4,x=26,y=195,},[K2]=1213,[K1]=K7,x=9,y=195,},[K19]={[1]={f=K3,[K10]="T",[K2]=1214,[K1]=K10,x=18,y=195,},xend=19,yend=195,},[K2]=1219,[K1]=K20,x=4,y=196,},[4]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1221,[K1]="any",x=18,y=196,},},[K2]=1222,[K1]=K4,x=25,y=196,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1224,[K1]="any",x=27,y=196,},},[K2]=1223,[K1]=K4,x=25,y=196,},[K2]=1220,[K1]=K7,x=9,y=196,},},x=7,y=194,},["maxinteger"]={f=K3,[K28]=true,[K2]=1225,[K1]=K12,x=16,y=198,},min={f=K3,[K2]=1238,[K1]="poly",[K18]={[1]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1227,[K1]=K12,x=18,y=200,},},[K2]=1228,[K1]=K4,x=29,y=200,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1230,[K1]=K12,x=31,y=200,},},[K2]=1229,[K1]=K4,x=29,y=200,},[K2]=1226,[K1]=K7,x=9,y=200,},[2]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1233,[K1]=K22,[K18]={[1]={f=K3,[K2]=1232,[K1]=K13,x=19,y=201,},[2]={f=K3,[K2]=1234,[K1]=K12,x=28,y=201,},},x=19,y=201,},},[K2]=1235,[K1]=K4,x=40,y=201,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1237,[K1]=K13,x=42,y=201,},},[K2]=1236,[K1]=K4,x=40,y=201,},[K2]=1231,[K1]=K7,x=9,y=201,},[3]={f=K3,t={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2273,[K1]=K5,[K5]="T",x=21,y=202,},},[K2]=1242,[K1]=K4,x=26,y=202,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=2274,[K1]=K5,[K5]="T",x=28,y=202,},},[K2]=1243,[K1]=K4,x=26,y=202,},[K2]=1239,[K1]=K7,x=9,y=202,},[K19]={[1]={f=K3,[K10]="T",[K2]=1240,[K1]=K10,x=18,y=202,},xend=19,yend=202,},[K2]=1245,[K1]=K20,x=4,y=203,},[4]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1247,[K1]="any",x=18,y=203,},},[K2]=1248,[K1]=K4,x=25,y=203,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1250,[K1]="any",x=27,y=203,},},[K2]=1249,[K1]=K4,x=25,y=203,},[K2]=1246,[K1]=K7,x=9,y=203,},},x=7,y=201,},["mininteger"]={f=K3,[K28]=true,[K2]=1251,[K1]=K12,x=16,y=205,},modf={args={f=K3,[K4]={[1]={f=K3,[K2]=1253,[K1]=K13,x=19,y=207,},},[K2]=1254,[K1]=K4,x=26,y=207,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1256,[K1]=K12,x=28,y=207,},[2]={f=K3,[K2]=1257,[K1]=K13,x=37,y=207,},},[K2]=1255,[K1]=K4,x=26,y=207,},[K2]=1252,[K1]=K7,x=10,y=207,},pi={f=K3,[K2]=1258,[K1]=K13,x=8,y=208,},pow={args={f=K3,[K4]={[1]={f=K3,[K2]=1260,[K1]=K13,x=18,y=209,},[2]={f=K3,[K2]=1261,[K1]=K13,x=26,y=209,},},[K2]=1262,[K1]=K4,x=33,y=209,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1264,[K1]=K13,x=35,y=209,},},[K2]=1263,[K1]=K4,x=33,y=209,},[K2]=1259,[K1]=K7,x=9,y=209,},rad={args={f=K3,[K4]={[1]={f=K3,[K2]=1266,[K1]=K13,x=18,y=210,},},[K2]=1267,[K1]=K4,x=25,y=210,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1269,[K1]=K13,x=27,y=210,},},[K2]=1268,[K1]=K4,x=25,y=210,},[K2]=1265,[K1]=K7,x=9,y=210,},["random"]={f=K3,[K2]=1280,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=1271,[K1]=K12,x=21,y=212,},[2]={f=K3,[K2]=1272,[K1]=K12,x=32,y=212,},},[K2]=1273,[K1]=K4,x=40,y=212,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1275,[K1]=K12,x=42,y=212,},},[K2]=1274,[K1]=K4,x=40,y=212,},[K2]=1270,[K1]=K7,x=12,y=212,},[2]={args={f=K3,[K4]={},[K2]=1277,[K1]=K4,x=22,y=213,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1279,[K1]=K13,x=24,y=213,},},[K2]=1278,[K1]=K4,x=22,y=213,},[K2]=1276,[K1]=K7,x=12,y=213,},},x=10,y=213,},["randomseed"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1282,[K1]=K12,x=27,y=215,},[2]={f=K3,[K2]=1283,[K1]=K12,x=38,y=215,},},[K2]=1284,[K1]=K4,x=46,y=215,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1286,[K1]=K12,x=48,y=215,},[2]={f=K3,[K2]=1287,[K1]=K12,x=57,y=215,},},[K2]=1285,[K1]=K4,x=46,y=215,},[K2]=1281,[K1]=K7,x=16,y=215,},sin={args={f=K3,[K4]={[1]={f=K3,[K2]=1289,[K1]=K13,x=18,y=216,},},[K2]=1290,[K1]=K4,x=25,y=216,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1292,[K1]=K13,x=27,y=216,},},[K2]=1291,[K1]=K4,x=25,y=216,},[K2]=1288,[K1]=K7,x=9,y=216,},sinh={args={f=K3,[K4]={[1]={f=K3,[K2]=1294,[K1]=K13,x=19,y=217,},},[K2]=1295,[K1]=K4,x=26,y=217,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1297,[K1]=K13,x=28,y=217,},},[K2]=1296,[K1]=K4,x=26,y=217,},[K2]=1293,[K1]=K7,x=10,y=217,},sqrt={args={f=K3,[K4]={[1]={f=K3,[K2]=1299,[K1]=K13,x=19,y=218,},},[K2]=1300,[K1]=K4,x=26,y=218,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1302,[K1]=K13,x=28,y=218,},},[K2]=1301,[K1]=K4,x=26,y=218,},[K2]=1298,[K1]=K7,x=10,y=218,},tan={args={f=K3,[K4]={[1]={f=K3,[K2]=1304,[K1]=K13,x=18,y=219,},},[K2]=1305,[K1]=K4,x=25,y=219,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1307,[K1]=K13,x=27,y=219,},},[K2]=1306,[K1]=K4,x=25,y=219,},[K2]=1303,[K1]=K7,x=9,y=219,},tanh={args={f=K3,[K4]={[1]={f=K3,[K2]=1309,[K1]=K13,x=19,y=220,},},[K2]=1310,[K1]=K4,x=26,y=220,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1312,[K1]=K13,x=28,y=220,},},[K2]=1311,[K1]=K4,x=26,y=220,},[K2]=1308,[K1]=K7,x=10,y=220,},["tointeger"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1314,[K1]="any",x=24,y=221,},},[K2]=1315,[K1]=K4,x=28,y=221,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1317,[K1]=K12,x=30,y=221,},},[K2]=1316,[K1]=K4,x=28,y=221,},[K2]=1313,[K1]=K7,x=15,y=221,},type={args={f=K3,[K4]={[1]={f=K3,[K2]=1319,[K1]="any",x=19,y=222,},},[K2]=1320,[K1]=K4,x=23,y=222,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1322,[K1]=K11,x=25,y=222,},},[K2]=1321,[K1]=K4,x=23,y=222,},[K2]=1318,[K1]=K7,x=10,y=222,},ult={args={f=K3,[K4]={[1]={f=K3,[K2]=1324,[K1]=K13,x=18,y=223,},[2]={f=K3,[K2]=1325,[K1]=K13,x=26,y=223,},},[K2]=1326,[K1]=K4,x=33,y=223,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1328,[K1]=K25,x=35,y=223,},},[K2]=1327,[K1]=K4,x=33,y=223,},[K2]=1323,[K1]=K7,x=9,y=223,},},[K2]=1098,[K1]=K37,x=1,y=169,},f=K3,[K2]=1329,[K1]=K21,x=1,y=169,},},[K42]={t=T12,[K[11]]=true,},next={[K27]=K26,t={f=K3,[K2]=1998,[K1]="poly",[K18]={[1]={f=K3,t={args={f=K3,[K4]={[1]={f=K3,keys={f=K3,[K2]=2528,[K1]=K5,[K5]="K",x=26,y=381,},[K2]=1980,[K1]="map",[K29]={f=K3,[K2]=2529,[K1]=K5,[K5]="V",x=28,y=381,},x=25,xend=29,y=381,yend=381,},[2]={f=K3,[K2]=2530,[K1]=K5,[K5]="K",x=34,y=381,},},[K2]=1983,[K1]=K4,x=36,y=381,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2531,[K1]=K5,[K5]="K",x=39,y=381,},[2]={f=K3,[K2]=2532,[K1]=K5,[K5]="V",x=42,y=381,},},[K2]=1984,[K1]=K4,x=36,y=381,},[K2]=1976,[K1]=K7,x=10,y=381,},[K19]={[1]={f=K3,[K10]="K",[K2]=1977,[K1]=K10,x=19,y=381,},[2]={f=K3,[K10]="V",[K2]=1978,[K1]=K10,x=22,y=381,},xend=23,yend=381,},[K2]=1987,[K1]=K20,x=4,y=382,},[2]={f=K3,t={args={f=K3,[K4]={[1]={[K32]={f=K3,[K2]=2546,[K1]=K5,[K5]="A",x=23,y=382,},f=K3,[K2]=1991,[K1]=K33,x=22,xend=24,y=382,yend=382,},[2]={f=K3,[K2]=1992,[K1]=K12,x=29,y=382,},},[K2]=1993,[K1]=K4,x=37,y=382,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1995,[K1]=K12,x=40,y=382,},[2]={f=K3,[K2]=2547,[K1]=K5,[K5]="A",x=49,y=382,},},[K2]=1994,[K1]=K4,x=37,y=382,},[K2]=1988,[K1]=K7,x=10,y=382,},[K19]={[1]={f=K3,[K10]="A",[K2]=1989,[K1]=K10,x=19,y=382,},xend=20,yend=382,},[K2]=1997,[K1]=K20,x=4,y=384,},},x=8,y=382,},},os={[K28]=true,t={[K34]=true,def={[K24]="os",f=K3,[K31]={[1]=K[23],[2]=K[24],[3]="clock",[4]="date",[5]="difftime",[6]="execute",[7]="exit",[8]="getenv",[9]=K[25],[10]="rename",[11]="setlocale",[12]="time",[13]="tmpname",},[K30]={[K[24]]=T19,[K[23]]=T20,["clock"]={args={f=K3,[K4]={},[K2]=1345,[K1]=K4,x=21,y=243,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1347,[K1]=K13,x=23,y=243,},},[K2]=1346,[K1]=K4,x=21,y=243,},[K2]=1344,[K1]=K7,x=11,y=243,},date={f=K3,[K2]=1360,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K17]=T19,[K16]={[1]=K[24],},[K2]=1349,[K1]=K15,x=19,y=245,},[2]={f=K3,[K2]=1350,[K1]=K13,x=31,y=245,},},[K2]=1351,[K1]=K4,x=38,y=245,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K17]=T20,[K16]={[1]=K[23],},[K2]=1353,[K1]=K15,x=40,y=245,},},[K2]=1352,[K1]=K4,x=38,y=245,},[K2]=1348,[K1]=K7,x=10,y=245,},[2]={args={f=K3,[K4]={[1]={f=K3,[K2]=1355,[K1]=K11,x=21,y=246,},[2]={f=K3,[K2]=1356,[K1]=K13,x=31,y=246,},},[K2]=1357,[K1]=K4,x=38,y=246,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1359,[K1]=K11,x=40,y=246,},},[K2]=1358,[K1]=K4,x=38,y=246,},[K2]=1354,[K1]=K7,x=10,y=246,},},x=8,y=246,},["difftime"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1362,[K1]=K12,x=23,y=248,},[2]={f=K3,[K2]=1363,[K1]=K12,x=32,y=248,},},[K2]=1364,[K1]=K4,x=40,y=248,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1366,[K1]=K13,x=42,y=248,},},[K2]=1365,[K1]=K4,x=40,y=248,},[K2]=1361,[K1]=K7,x=14,y=248,},["execute"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1368,[K1]=K11,x=22,y=249,},},[K2]=1369,[K1]=K4,x=29,y=249,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1371,[K1]=K25,x=31,y=249,},[2]={f=K3,[K2]=1372,[K1]=K11,x=40,y=249,},[3]={f=K3,[K2]=1373,[K1]=K12,x=48,y=249,},},[K2]=1370,[K1]=K4,x=29,y=249,},[K2]=1367,[K1]=K7,x=13,y=249,},exit={args={f=K3,[K4]={[1]={f=K3,[K2]=1376,[K1]=K22,[K18]={[1]={f=K3,[K2]=1375,[K1]=K12,x=22,y=250,},[2]={f=K3,[K2]=1377,[K1]=K25,x=32,y=250,},},x=22,y=250,},[2]={f=K3,[K2]=1378,[K1]=K25,x=44,y=250,},},[K2]=1379,[K1]=K4,x=4,y=251,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={},[K2]=1380,[K1]=K4,x=51,y=250,},[K2]=1374,[K1]=K7,x=10,y=250,},["getenv"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1382,[K1]=K11,x=21,y=251,},},[K2]=1383,[K1]=K4,x=28,y=251,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1385,[K1]=K11,x=30,y=251,},},[K2]=1384,[K1]=K4,x=28,y=251,},[K2]=1381,[K1]=K7,x=12,y=251,},[K[25]]={args={f=K3,[K4]={[1]={f=K3,[K2]=1387,[K1]=K11,x=21,y=252,},},[K2]=1388,[K1]=K4,x=28,y=252,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1390,[K1]=K25,x=30,y=252,},[2]={f=K3,[K2]=1391,[K1]=K11,x=39,y=252,},},[K2]=1389,[K1]=K4,x=28,y=252,},[K2]=1386,[K1]=K7,x=12,y=252,},["rename"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1393,[K1]=K11,x=21,y=253,},[2]={f=K3,[K2]=1394,[K1]=K11,x=29,y=253,},},[K2]=1395,[K1]=K4,x=36,y=253,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1397,[K1]=K25,x=38,y=253,},[2]={f=K3,[K2]=1398,[K1]=K11,x=47,y=253,},},[K2]=1396,[K1]=K4,x=36,y=253,},[K2]=1392,[K1]=K7,x=12,y=253,},["setlocale"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1400,[K1]=K11,x=24,y=254,},[2]={f=K3,[K2]=1401,[K1]=K11,x=34,y=254,},},[K2]=1402,[K1]=K4,x=41,y=254,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1404,[K1]=K11,x=43,y=254,},},[K2]=1403,[K1]=K4,x=41,y=254,},[K2]=1399,[K1]=K7,x=15,y=254,},time={args={f=K3,[K4]={[1]={f=K3,[K17]=T20,[K16]={[1]=K[23],},[K2]=1406,[K1]=K15,x=21,y=255,},},[K2]=1407,[K1]=K4,x=31,y=255,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1409,[K1]=K12,x=33,y=255,},},[K2]=1408,[K1]=K4,x=31,y=255,},[K2]=1405,[K1]=K7,x=10,y=255,},["tmpname"]={args={f=K3,[K4]={},[K2]=1411,[K1]=K4,x=23,y=256,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1413,[K1]=K11,x=25,y=256,},},[K2]=1412,[K1]=K4,x=23,y=256,},[K2]=1410,[K1]=K7,x=13,y=256,},},[K2]=1330,[K1]=K37,x=1,y=226,},f=K3,[K2]=1414,[K1]=K21,x=1,y=226,},},["package"]={[K28]=true,t={[K34]=true,def={[K24]="package",f=K3,[K31]={[1]="config",[2]="cpath",[3]="loaded",[4]="loadlib",[5]="loaders",[6]="path",[7]="preload",[8]="searchers",[9]="searchpath",},[K30]={["config"]={f=K3,[K2]=1416,[K1]=K11,x=12,y=260,},["cpath"]={f=K3,[K2]=1417,[K1]=K11,x=11,y=261,},["loaded"]={f=K3,keys={f=K3,[K2]=1418,[K1]=K11,x=13,y=262,},[K2]=1419,[K1]="map",[K29]={f=K3,[K2]=1420,[K1]="any",x=20,y=262,},x=12,xend=23,y=262,yend=262,},["loaders"]={[K32]={args={f=K3,[K4]={[1]={f=K3,[K2]=1432,[K1]=K11,x=24,y=264,},},[K2]=1433,[K1]=K4,x=31,y=264,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1435,[K1]="any",x=33,y=264,},[2]={f=K3,[K2]=1436,[K1]="any",x=38,y=264,},},[K2]=1434,[K1]=K4,x=31,y=264,},[K2]=1431,[K1]=K7,x=15,y=264,},f=K3,[K2]=1437,[K1]=K33,x=13,xend=42,y=264,yend=264,},["loadlib"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1422,[K1]=K11,x=22,y=263,},[2]={f=K3,[K2]=1423,[K1]=K11,x=30,y=263,},},[K2]=1424,[K1]=K4,x=37,y=263,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1427,[K1]="any",x=48,y=263,},},[K2]=1428,[K1]=K4,x=48,y=263,},f=K3,[K46]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1429,[K1]="any",x=48,y=263,},},[K2]=1430,[K1]=K4,x=48,y=263,},[K2]=1426,[K1]=K7,x=40,y=263,},},[K2]=1425,[K1]=K4,x=37,y=263,},[K2]=1421,[K1]=K7,x=13,y=263,},path={f=K3,[K2]=1438,[K1]=K11,x=10,y=265,},["preload"]={f=K3,keys={f=K3,[K2]=1439,[K1]="any",x=14,y=266,},[K2]=1440,[K1]="map",[K29]={f=K3,[K2]=1441,[K1]="any",x=18,y=266,},x=13,xend=21,y=266,yend=266,},["searchers"]={[K32]={args={f=K3,[K4]={[1]={f=K3,[K2]=1443,[K1]=K11,x=26,y=267,},},[K2]=1444,[K1]=K4,x=33,y=267,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1446,[K1]="any",x=35,y=267,},},[K2]=1445,[K1]=K4,x=33,y=267,},[K2]=1442,[K1]=K7,x=17,y=267,},f=K3,[K2]=1447,[K1]=K33,x=15,xend=39,y=267,yend=267,},["searchpath"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1449,[K1]=K11,x=25,y=268,},[2]={f=K3,[K2]=1450,[K1]=K11,x=33,y=268,},[3]={f=K3,[K2]=1451,[K1]=K11,x=43,y=268,},[4]={f=K3,[K2]=1452,[K1]=K11,x=53,y=268,},},[K2]=1453,[K1]=K4,x=60,y=268,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1455,[K1]=K11,x=62,y=268,},[2]={f=K3,[K2]=1456,[K1]=K11,x=70,y=268,},},[K2]=1454,[K1]=K4,x=60,y=268,},[K2]=1448,[K1]=K7,x=16,y=268,},},[K2]=1415,[K1]=K37,x=1,y=259,},f=K3,[K2]=1457,[K1]=K21,x=1,y=259,},},["pairs"]={[K27]=K26,[K28]=true,t={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={f=K3,keys={f=K3,[K2]=2564,[K1]=K5,[K5]="K@48",x=27,y=384,},[K2]=2566,[K1]="map",[K29]={f=K3,[K2]=2565,[K1]=K5,[K5]="V@48",x=29,y=384,},x=26,y=384,},},[K2]=2567,[K1]=K4,x=32,y=384,},f=K3,[K6]=1,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=2008,[K1]=K4,x=45,y=384,},f=K3,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=2568,[K1]=K5,[K5]="K@48",x=47,y=384,},[2]={f=K3,[K2]=2569,[K1]=K5,[K5]="V@48",x=50,y=384,},},[K2]=2570,[K1]=K4,x=45,y=384,},[K2]=2571,[K1]=K7,x=35,y=384,},},[K2]=2572,[K1]=K4,x=32,y=384,},[K36]="pairs",[K2]=2573,[K1]=K7,x=11,y=384,},[K19]={[1]={f=K3,[K10]="K@48",[K2]=2562,[K1]=K10,x=20,y=384,},[2]={f=K3,[K10]="V@48",[K2]=2563,[K1]=K10,x=23,y=384,},},[K2]=2574,[K1]=K20,x=4,y=385,},},["pcall"]={[K27]=K26,[K28]=true,t={args={f=K3,[K14]=true,[K4]={[1]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2015,[K1]="any",x=29,y=385,},},[K2]=2016,[K1]=K4,x=36,y=385,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2018,[K1]="any",x=38,y=385,},},[K2]=2017,[K1]=K4,x=36,y=385,},[K2]=2014,[K1]=K7,x=20,y=385,},[2]={f=K3,[K2]=2019,[K1]="any",x=47,y=385,},},[K2]=2020,[K1]=K4,x=54,y=385,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2022,[K1]=K25,x=56,y=385,},[2]={f=K3,[K2]=2023,[K1]="any",x=65,y=385,},},[K2]=2021,[K1]=K4,x=54,y=385,},[K36]="pcall",[K2]=2013,[K1]=K7,x=11,y=385,},},["print"]={[K27]=K26,t={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2025,[K1]="any",x=20,y=386,},},[K2]=2026,[K1]=K4,x=4,y=387,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={},[K2]=2027,[K1]=K4,x=26,y=386,},[K2]=2024,[K1]=K7,x=11,y=386,},},["rawequal"]={[K27]=K26,t={args={f=K3,[K4]={[1]={f=K3,[K2]=2029,[K1]="any",x=23,y=387,},[2]={f=K3,[K2]=2030,[K1]="any",x=28,y=387,},},[K2]=2031,[K1]=K4,x=32,y=387,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=2033,[K1]=K25,x=34,y=387,},},[K2]=2032,[K1]=K4,x=32,y=387,},[K2]=2028,[K1]=K7,x=14,y=387,},},["rawget"]={[K27]=K26,t={f=K3,[K2]=2053,[K1]="poly",[K18]={[1]={f=K3,t={args={f=K3,[K4]={[1]={f=K3,keys={f=K3,[K2]=2577,[K1]=K5,[K5]="K",x=28,y=389,},[K2]=2038,[K1]="map",[K29]={f=K3,[K2]=2578,[K1]=K5,[K5]="V",x=30,y=389,},x=27,xend=31,y=389,yend=389,},[2]={f=K3,[K2]=2579,[K1]=K5,[K5]="K",x=34,y=389,},},[K2]=2041,[K1]=K4,x=36,y=389,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=2580,[K1]=K5,[K5]="V",x=38,y=389,},},[K2]=2042,[K1]=K4,x=36,y=389,},[K36]="rawget",[K2]=2034,[K1]=K7,x=12,y=389,},[K19]={[1]={f=K3,[K10]="K",[K2]=2035,[K1]=K10,x=21,y=389,},[2]={f=K3,[K10]="V",[K2]=2036,[K1]=K10,x=24,y=389,},xend=25,yend=389,},[K2]=2044,[K1]=K20,x=4,y=390,},[2]={args={f=K3,[K4]={[1]={f=K3,keys={f=K3,[K2]=2046,[K1]="any",x=22,y=390,},[K2]=2047,[K1]="map",[K29]={f=K3,[K2]=2048,[K1]="any",x=26,y=390,},x=21,xend=29,y=390,yend=390,},[2]={f=K3,[K2]=2049,[K1]="any",x=32,y=390,},},[K2]=2050,[K1]=K4,x=36,y=390,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=2052,[K1]="any",x=38,y=390,},},[K2]=2051,[K1]=K4,x=36,y=390,},[K2]=2045,[K1]=K7,x=12,y=390,},[3]={args={f=K3,[K4]={[1]={f=K3,[K2]=2055,[K1]="any",x=21,y=391,},[2]={f=K3,[K2]=2056,[K1]="any",x=26,y=391,},},[K2]=2057,[K1]=K4,x=30,y=391,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=2059,[K1]="any",x=32,y=391,},},[K2]=2058,[K1]=K4,x=30,y=391,},[K2]=2054,[K1]=K7,x=12,y=391,},},x=10,y=390,},},["rawlen"]={[K27]=K26,[K28]=true,t={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={[K32]={f=K3,[K2]=2595,[K1]=K5,[K5]="A@50",x=25,y=393,},f=K3,[K2]=2596,[K1]=K33,x=24,y=393,},},[K2]=2597,[K1]=K4,x=28,y=393,},f=K3,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2066,[K1]=K12,x=30,y=393,},},[K2]=2065,[K1]=K4,x=28,y=393,},[K2]=2598,[K1]=K7,x=12,y=393,},[K19]={[1]={f=K3,[K10]="A@50",[K2]=2594,[K1]=K10,x=21,y=393,},},[K2]=2599,[K1]=K20,x=4,y=395,},},["rawset"]={[K27]=K26,t={f=K3,[K2]=2093,[K1]="poly",[K18]={[1]={f=K3,t={args={f=K3,[K4]={[1]={f=K3,keys={f=K3,[K2]=2602,[K1]=K5,[K5]="K",x=28,y=395,},[K2]=2072,[K1]="map",[K29]={f=K3,[K2]=2603,[K1]=K5,[K5]="V",x=30,y=395,},x=27,xend=31,y=395,yend=395,},[2]={f=K3,[K2]=2604,[K1]=K5,[K5]="K",x=34,y=395,},[3]={f=K3,[K2]=2605,[K1]=K5,[K5]="V",x=37,y=395,},},[K2]=2076,[K1]=K4,x=39,y=395,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,keys={f=K3,[K2]=2606,[K1]=K5,[K5]="K",x=42,y=395,},[K2]=2079,[K1]="map",[K29]={f=K3,[K2]=2607,[K1]=K5,[K5]="V",x=44,y=395,},x=41,xend=45,y=395,yend=395,},},[K2]=2077,[K1]=K4,x=39,y=395,},[K2]=2068,[K1]=K7,x=12,y=395,},[K19]={[1]={f=K3,[K10]="K",[K2]=2069,[K1]=K10,x=21,y=395,},[2]={f=K3,[K10]="V",[K2]=2070,[K1]=K10,x=24,y=395,},xend=25,yend=395,},[K2]=2081,[K1]=K20,x=4,y=396,},[2]={args={f=K3,[K4]={[1]={f=K3,keys={f=K3,[K2]=2083,[K1]="any",x=22,y=396,},[K2]=2084,[K1]="map",[K29]={f=K3,[K2]=2085,[K1]="any",x=26,y=396,},x=21,xend=29,y=396,yend=396,},[2]={f=K3,[K2]=2086,[K1]="any",x=32,y=396,},[3]={f=K3,[K2]=2087,[K1]="any",x=37,y=396,},},[K2]=2088,[K1]=K4,x=41,y=396,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,keys={f=K3,[K2]=2090,[K1]="any",x=44,y=396,},[K2]=2091,[K1]="map",[K29]={f=K3,[K2]=2092,[K1]="any",x=48,y=396,},x=43,xend=51,y=396,yend=396,},},[K2]=2089,[K1]=K4,x=41,y=396,},[K2]=2082,[K1]=K7,x=12,y=396,},[3]={args={f=K3,[K4]={[1]={f=K3,[K2]=2095,[K1]="any",x=21,y=397,},[2]={f=K3,[K2]=2096,[K1]="any",x=26,y=397,},[3]={f=K3,[K2]=2097,[K1]="any",x=31,y=397,},},[K2]=2098,[K1]=K4,x=35,y=397,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,[K2]=2100,[K1]="any",x=37,y=397,},},[K2]=2099,[K1]=K4,x=35,y=397,},[K2]=2094,[K1]=K7,x=12,y=397,},},x=10,y=396,},},["require"]={[K27]=K26,t={args={f=K3,[K4]={[1]={f=K3,[K2]=2102,[K1]=K11,x=22,y=399,},},[K2]=2103,[K1]=K4,x=29,y=399,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2105,[K1]="any",x=31,y=399,},},[K2]=2104,[K1]=K4,x=29,y=399,},[K36]="require",[K2]=2101,[K1]=K7,x=13,y=399,},},["select"]={[K27]=K26,t={f=K3,[K2]=2120,[K1]="poly",[K18]={[1]={f=K3,t={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2108,[K1]=K12,x=24,y=401,},[2]={f=K3,[K2]=2623,[K1]=K5,[K5]="T",x=33,y=401,},},[K2]=2110,[K1]=K4,x=38,y=401,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2624,[K1]=K5,[K5]="T",x=40,y=401,},},[K2]=2111,[K1]=K4,x=38,y=401,},[K2]=2106,[K1]=K7,x=12,y=401,},[K19]={[1]={f=K3,[K10]="T",[K2]=2107,[K1]=K10,x=21,y=401,},xend=22,yend=401,},[K2]=2113,[K1]=K20,x=4,y=402,},[2]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2115,[K1]=K12,x=21,y=402,},[2]={f=K3,[K2]=2116,[K1]="any",x=30,y=402,},},[K2]=2117,[K1]=K4,x=37,y=402,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2119,[K1]="any",x=39,y=402,},},[K2]=2118,[K1]=K4,x=37,y=402,},[K2]=2114,[K1]=K7,x=12,y=402,},[3]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2122,[K1]=K11,x=21,y=403,},[2]={f=K3,[K2]=2123,[K1]="any",x=29,y=403,},},[K2]=2124,[K1]=K4,x=36,y=403,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2126,[K1]=K12,x=38,y=403,},},[K2]=2125,[K1]=K4,x=36,y=403,},[K2]=2121,[K1]=K7,x=12,y=403,},},x=10,y=402,},},[K[17]]={[K27]=K26,t={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={f=K3,[K2]=2637,[K1]=K5,[K5]="T@53",x=30,y=405,},[2]={f=K3,[K17]=T12,[K16]={[1]=K42,},[K2]=2639,[K1]=K15,[K41]={[1]={f=K3,[K2]=2638,[K1]=K5,[K5]="T@53",x=43,y=405,},},x=33,y=405,},},[K2]=2640,[K1]=K4,x=46,y=405,},f=K3,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=2641,[K1]=K5,[K5]="T@53",x=48,y=405,},},[K2]=2642,[K1]=K4,x=46,y=405,},[K2]=2643,[K1]=K7,x=18,y=405,},[K19]={[1]={f=K3,[K10]="T@53",[K2]=2636,[K1]=K10,x=27,y=405,},},[K2]=2644,[K1]=K20,x=4,y=407,},},[K11]={[K28]=true,t={[K34]=true,def={[K24]=K11,f=K3,[K[19]]={[K[26]]={[1]={[1]={text=K40,x=86,y=277,},},},[K[27]]={[1]={[1]={text=K40,x=45,y=278,},},},gsub={[1]={},[2]={[1]={text=K40,x=71,y=281,},},[3]={[1]={text=K40,x=95,y=282,},},[4]={[1]={text=K40,x=96,y=283,},},},pack={[1]={[1]={text=K40,x=58,y=288,},},},[K[28]]={[1]={[1]={text=K40,x=43,y=289,},},},[K[29]]={[1]={[1]={text=K40,x=56,y=294,},},},},[K31]={[1]="byte",[2]="char",[3]="dump",[4]="find",[5]=K[26],[6]=K[27],[7]="gsub",[8]="len",[9]="lower",[10]="match",[11]="pack",[12]=K[28],[13]="rep",[14]="reverse",[15]="sub",[16]=K45,[17]=K[29],},[K30]={byte={f=K3,[K2]=1472,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=1460,[K1]=K11,x=19,y=272,},[2]={f=K3,[K2]=1461,[K1]=K12,x=29,y=272,},},[K2]=1462,[K1]=K4,x=37,y=272,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1464,[K1]=K12,x=39,y=272,},},[K2]=1463,[K1]=K4,x=37,y=272,},[K2]=1459,[K1]=K7,x=10,y=272,},[2]={args={f=K3,[K4]={[1]={f=K3,[K2]=1466,[K1]=K11,x=19,y=273,},[2]={f=K3,[K2]=1467,[K1]=K12,x=27,y=273,},[3]={f=K3,[K2]=1468,[K1]=K12,x=38,y=273,},},[K2]=1469,[K1]=K4,x=46,y=273,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1471,[K1]=K12,x=48,y=273,},},[K2]=1470,[K1]=K4,x=46,y=273,},[K2]=1465,[K1]=K7,x=10,y=273,},},x=8,y=273,},char={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1474,[K1]=K12,x=19,y=275,},},[K2]=1475,[K1]=K4,x=30,y=275,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1477,[K1]=K11,x=32,y=275,},},[K2]=1476,[K1]=K4,x=30,y=275,},[K2]=1473,[K1]=K7,x=10,y=275,},dump={args={f=K3,[K4]={[1]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1480,[K1]="any",x=28,y=276,},},[K2]=1481,[K1]=K4,x=35,y=276,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1483,[K1]="any",x=38,y=276,},},[K2]=1482,[K1]=K4,x=35,y=276,},[K2]=1479,[K1]=K7,x=19,y=276,},[2]={f=K3,[K2]=1484,[K1]=K25,x=46,y=276,},},[K2]=1485,[K1]=K4,x=54,y=276,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1487,[K1]=K11,x=56,y=276,},},[K2]=1486,[K1]=K4,x=54,y=276,},[K2]=1478,[K1]=K7,x=10,y=276,},find={args={f=K3,[K4]={[1]={f=K3,[K2]=1489,[K1]=K11,x=19,y=277,},[2]={f=K3,[K2]=1490,[K1]=K11,x=27,y=277,},[3]={f=K3,[K2]=1491,[K1]=K12,x=37,y=277,},[4]={f=K3,[K2]=1492,[K1]=K25,x=48,y=277,},},[K2]=1493,[K1]=K4,x=56,y=277,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1495,[K1]=K12,x=58,y=277,},[2]={f=K3,[K2]=1496,[K1]=K12,x=67,y=277,},[3]={f=K3,[K2]=1497,[K1]=K11,x=76,y=277,},},[K2]=1494,[K1]=K4,x=56,y=277,},[K36]="string.find",[K2]=1488,[K1]=K7,x=10,y=277,},[K[26]]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1499,[K1]=K11,x=21,y=278,},[2]={f=K3,[K2]=1500,[K1]="any",x=29,y=278,},},[K2]=1501,[K1]=K4,x=36,y=278,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1503,[K1]=K11,x=38,y=278,},},[K2]=1502,[K1]=K4,x=36,y=278,},[K36]="string.format",[K2]=1498,[K1]=K7,x=12,y=278,},[K[27]]={args={f=K3,[K4]={[1]={f=K3,[K2]=1505,[K1]=K11,x=21,y=279,},[2]={f=K3,[K2]=1506,[K1]=K11,x=29,y=279,},[3]={f=K3,[K2]=1507,[K1]=K12,x=39,y=279,},},[K2]=1508,[K1]=K4,x=47,y=279,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=1511,[K1]=K4,x=60,y=279,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1513,[K1]=K11,x=62,y=279,},},[K2]=1512,[K1]=K4,x=60,y=279,},[K2]=1510,[K1]=K7,x=50,y=279,},},[K2]=1509,[K1]=K4,x=47,y=279,},[K36]="string.gmatch",[K2]=1504,[K1]=K7,x=12,y=279,},gsub={f=K3,[K2]=1537,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=1515,[K1]=K11,x=19,y=281,},[2]={f=K3,[K2]=1516,[K1]=K11,x=27,y=281,},[3]={f=K3,[K2]=1517,[K1]=K11,x=35,y=281,},[4]={f=K3,[K2]=1518,[K1]=K12,x=45,y=281,},},[K2]=1519,[K1]=K4,x=53,y=281,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,[K2]=1521,[K1]=K11,x=55,y=281,},[2]={f=K3,[K2]=1522,[K1]=K12,x=63,y=281,},},[K2]=1520,[K1]=K4,x=53,y=281,},[K36]="string.gsub",[K2]=1514,[K1]=K7,x=10,y=281,},[2]={args={f=K3,[K4]={[1]={f=K3,[K2]=1524,[K1]=K11,x=19,y=282,},[2]={f=K3,[K2]=1525,[K1]=K11,x=27,y=282,},[3]={f=K3,keys={f=K3,[K2]=1526,[K1]=K11,x=36,y=282,},[K2]=1527,[K1]="map",[K29]={f=K3,[K2]=1529,[K1]=K22,[K18]={[1]={f=K3,[K2]=1528,[K1]=K11,x=43,y=282,},[2]={f=K3,[K2]=1530,[K1]=K12,x=50,y=282,},[3]={f=K3,[K2]=1531,[K1]=K13,x=58,y=282,},},x=43,y=282,},x=35,xend=64,y=282,yend=282,},[4]={f=K3,[K2]=1532,[K1]=K12,x=69,y=282,},},[K2]=1533,[K1]=K4,x=77,y=282,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,[K2]=1535,[K1]=K11,x=79,y=282,},[2]={f=K3,[K2]=1536,[K1]=K12,x=87,y=282,},},[K2]=1534,[K1]=K4,x=77,y=282,},[K2]=1523,[K1]=K7,x=10,y=282,},[3]={args={f=K3,[K4]={[1]={f=K3,[K2]=1539,[K1]=K11,x=19,y=283,},[2]={f=K3,[K2]=1540,[K1]=K11,x=27,y=283,},[3]={f=K3,keys={f=K3,[K2]=1541,[K1]=K12,x=36,y=283,},[K2]=1542,[K1]="map",[K29]={f=K3,[K2]=1544,[K1]=K22,[K18]={[1]={f=K3,[K2]=1543,[K1]=K11,x=44,y=283,},[2]={f=K3,[K2]=1545,[K1]=K12,x=51,y=283,},[3]={f=K3,[K2]=1546,[K1]=K13,x=59,y=283,},},x=44,y=283,},x=35,xend=65,y=283,yend=283,},[4]={f=K3,[K2]=1547,[K1]=K12,x=70,y=283,},},[K2]=1548,[K1]=K4,x=78,y=283,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,[K2]=1550,[K1]=K11,x=80,y=283,},[2]={f=K3,[K2]=1551,[K1]=K12,x=88,y=283,},},[K2]=1549,[K1]=K4,x=78,y=283,},[K2]=1538,[K1]=K7,x=10,y=283,},[4]={args={f=K3,[K4]={[1]={f=K3,[K2]=1553,[K1]=K11,x=19,y=284,},[2]={f=K3,[K2]=1554,[K1]=K11,x=27,y=284,},[3]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1557,[K1]=K22,[K18]={[1]={f=K3,[K2]=1556,[K1]=K11,x=45,y=284,},[2]={f=K3,[K2]=1558,[K1]=K12,x=52,y=284,},},x=45,y=284,},},[K2]=1559,[K1]=K4,x=64,y=284,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1562,[K1]=K22,[K18]={[1]={f=K3,[K2]=1561,[K1]=K11,x=68,y=284,},[2]={f=K3,[K2]=1563,[K1]=K12,x=75,y=284,},[3]={f=K3,[K2]=1564,[K1]=K13,x=83,y=284,},},x=68,y=284,},},[K2]=1560,[K1]=K4,x=64,y=284,},[K2]=1555,[K1]=K7,x=35,y=284,},[4]={f=K3,[K2]=1565,[K1]=K12,x=98,y=284,},},[K2]=1566,[K1]=K4,x=106,y=284,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={[1]={f=K3,[K2]=1568,[K1]=K11,x=108,y=284,},[2]={f=K3,[K2]=1569,[K1]=K12,x=116,y=284,},},[K2]=1567,[K1]=K4,x=106,y=284,},[K2]=1552,[K1]=K7,x=10,y=284,},},x=8,y=282,},len={args={f=K3,[K4]={[1]={f=K3,[K2]=1571,[K1]=K11,x=18,y=286,},},[K2]=1572,[K1]=K4,x=25,y=286,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1574,[K1]=K12,x=27,y=286,},},[K2]=1573,[K1]=K4,x=25,y=286,},[K2]=1570,[K1]=K7,x=9,y=286,},["lower"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1576,[K1]=K11,x=20,y=287,},},[K2]=1577,[K1]=K4,x=27,y=287,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1579,[K1]=K11,x=29,y=287,},},[K2]=1578,[K1]=K4,x=27,y=287,},[K2]=1575,[K1]=K7,x=11,y=287,},["match"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1581,[K1]=K11,x=20,y=288,},[2]={f=K3,[K2]=1582,[K1]=K11,x=28,y=288,},[3]={f=K3,[K2]=1583,[K1]=K12,x=38,y=288,},},[K2]=1584,[K1]=K4,x=46,y=288,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1586,[K1]=K11,x=48,y=288,},},[K2]=1585,[K1]=K4,x=46,y=288,},[K36]="string.match",[K2]=1580,[K1]=K7,x=11,y=288,},pack={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1588,[K1]=K11,x=19,y=289,},[2]={f=K3,[K2]=1589,[K1]="any",x=27,y=289,},},[K2]=1590,[K1]=K4,x=34,y=289,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1592,[K1]=K11,x=36,y=289,},},[K2]=1591,[K1]=K4,x=34,y=289,},[K36]="string.pack",[K2]=1587,[K1]=K7,x=10,y=289,},[K[28]]={args={f=K3,[K4]={[1]={f=K3,[K2]=1594,[K1]=K11,x=23,y=290,},},[K2]=1595,[K1]=K4,x=30,y=290,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1597,[K1]=K12,x=32,y=290,},},[K2]=1596,[K1]=K4,x=30,y=290,},[K2]=1593,[K1]=K7,x=14,y=290,},rep={args={f=K3,[K4]={[1]={f=K3,[K2]=1599,[K1]=K11,x=18,y=291,},[2]={f=K3,[K2]=1600,[K1]=K12,x=26,y=291,},[3]={f=K3,[K2]=1601,[K1]=K11,x=37,y=291,},},[K2]=1602,[K1]=K4,x=44,y=291,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1604,[K1]=K11,x=46,y=291,},},[K2]=1603,[K1]=K4,x=44,y=291,},[K2]=1598,[K1]=K7,x=9,y=291,},["reverse"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1606,[K1]=K11,x=22,y=292,},},[K2]=1607,[K1]=K4,x=29,y=292,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1609,[K1]=K11,x=31,y=292,},},[K2]=1608,[K1]=K4,x=29,y=292,},[K2]=1605,[K1]=K7,x=13,y=292,},sub={args={f=K3,[K4]={[1]={f=K3,[K2]=1611,[K1]=K11,x=18,y=293,},[2]={f=K3,[K2]=1612,[K1]=K12,x=26,y=293,},[3]={f=K3,[K2]=1613,[K1]=K12,x=37,y=293,},},[K2]=1614,[K1]=K4,x=45,y=293,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1616,[K1]=K11,x=47,y=293,},},[K2]=1615,[K1]=K4,x=45,y=293,},[K2]=1610,[K1]=K7,x=9,y=293,},[K45]={args={f=K3,[K4]={[1]={f=K3,[K2]=1618,[K1]=K11,x=21,y=294,},[2]={f=K3,[K2]=1619,[K1]=K11,x=29,y=294,},[3]={f=K3,[K2]=1620,[K1]=K12,x=39,y=294,},},[K2]=1621,[K1]=K4,x=47,y=294,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1623,[K1]="any",x=49,y=294,},},[K2]=1622,[K1]=K4,x=47,y=294,},[K36]="string.unpack",[K2]=1617,[K1]=K7,x=12,y=294,},[K[29]]={args={f=K3,[K4]={[1]={f=K3,[K2]=1625,[K1]=K11,x=20,y=295,},},[K2]=1626,[K1]=K4,x=27,y=295,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1628,[K1]=K11,x=29,y=295,},},[K2]=1627,[K1]=K4,x=27,y=295,},[K2]=1624,[K1]=K7,x=11,y=295,},},[K2]=1458,[K1]=K37,x=1,y=271,},f=K3,[K2]=1629,[K1]=K21,x=1,y=271,},},["table"]={[K28]=true,t={[K34]=true,def={[K24]="table",f=K3,[K[19]]={pack={[1]={},[2]={[1]={text=K[30],x=42,y=314,},},},[K45]={[1]={},[2]={[1]={text=K[30],x=55,y=320,},},[3]={[1]={text=K[30],x=47,y=321,},},[4]={[1]={text=K[30],x=59,y=322,},},[5]={[1]={text=K[30],x=71,y=323,},},},},[K31]={[1]=K[31],[2]=K[32],[3]="concat",[4]="insert",[5]="move",[6]="pack",[7]=K[25],[8]="sort",[9]=K45,},[K30]={[K[32]]=T21,[K[31]]=T23,["concat"]={args={f=K3,[K4]={[1]={[K32]={f=K3,[K2]=1649,[K1]=K22,[K18]={[1]={f=K3,[K2]=1648,[K1]=K11,x=23,y=307,},[2]={f=K3,[K2]=1650,[K1]=K13,x=32,y=307,},},x=23,y=307,},f=K3,[K2]=1651,[K1]=K33,x=21,xend=39,y=307,yend=307,},[2]={f=K3,[K2]=1652,[K1]=K11,x=44,y=307,},[3]={f=K3,[K2]=1653,[K1]=K12,x=54,y=307,},[4]={f=K3,[K2]=1654,[K1]=K12,x=65,y=307,},},[K2]=1655,[K1]=K4,x=73,y=307,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1657,[K1]=K11,x=75,y=307,},},[K2]=1656,[K1]=K4,x=73,y=307,},[K2]=1647,[K1]=K7,x=12,y=307,},["insert"]={f=K3,[K2]=1675,[K1]="poly",[K18]={[1]={f=K3,t={args={f=K3,[K4]={[1]={[K32]={f=K3,[K2]=2304,[K1]=K5,[K5]="A",x=25,y=309,},f=K3,[K2]=1661,[K1]=K33,x=24,xend=26,y=309,yend=309,},[2]={f=K3,[K2]=1662,[K1]=K12,x=29,y=309,},[3]={f=K3,[K2]=2305,[K1]=K5,[K5]="A",x=38,y=309,},},[K2]=1664,[K1]=K4,x=4,y=310,},f=K3,[K9]=false,[K6]=3,rets={f=K3,[K4]={},[K2]=1665,[K1]=K4,x=39,y=309,},[K2]=1658,[K1]=K7,x=12,y=309,},[K19]={[1]={f=K3,[K10]="A",[K2]=1659,[K1]=K10,x=21,y=309,},xend=22,yend=309,},[K2]=1666,[K1]=K20,x=4,y=310,},[2]={f=K3,t={args={f=K3,[K4]={[1]={[K32]={f=K3,[K2]=2314,[K1]=K5,[K5]="A",x=25,y=310,},f=K3,[K2]=1670,[K1]=K33,x=24,xend=26,y=310,yend=310,},[2]={f=K3,[K2]=2315,[K1]=K5,[K5]="A",x=29,y=310,},},[K2]=1672,[K1]=K4,x=4,y=312,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={},[K2]=1673,[K1]=K4,x=30,y=310,},[K2]=1667,[K1]=K7,x=12,y=310,},[K19]={[1]={f=K3,[K10]="A",[K2]=1668,[K1]=K10,x=21,y=310,},xend=22,yend=310,},[K2]=1674,[K1]=K20,x=4,y=312,},},x=10,y=310,},move={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={[K32]={f=K3,[K2]=2328,[K1]=K5,[K5]="A@34",x=23,y=312,},f=K3,[K2]=2329,[K1]=K33,x=22,y=312,},[2]={f=K3,[K2]=1680,[K1]=K12,x=27,y=312,},[3]={f=K3,[K2]=1681,[K1]=K12,x=36,y=312,},[4]={f=K3,[K2]=1682,[K1]=K12,x=45,y=312,},[5]={[K32]={f=K3,[K2]=2330,[K1]=K5,[K5]="A@34",x=57,y=312,},f=K3,[K2]=2331,[K1]=K33,x=56,y=312,},},[K2]=2332,[K1]=K4,x=60,y=312,},f=K3,[K6]=4,rets={f=K3,[K4]={[1]={[K32]={f=K3,[K2]=2333,[K1]=K5,[K5]="A@34",x=63,y=312,},f=K3,[K2]=2334,[K1]=K33,x=62,y=312,},},[K2]=2335,[K1]=K4,x=60,y=312,},[K2]=2336,[K1]=K7,x=10,y=312,},[K19]={[1]={f=K3,[K10]="A@34",[K2]=2327,[K1]=K10,x=19,y=312,},},[K2]=2337,[K1]=K20,x=4,y=314,},pack={f=K3,[K28]=true,[K2]=1705,[K1]="poly",[K18]={[1]={f=K3,t={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2339,[K1]=K5,[K5]="T",x=22,y=314,},},[K2]=1693,[K1]=K4,x=27,y=314,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K17]=T21,[K16]={[1]=K[32],},[K2]=1695,[K1]=K15,[K41]={[1]={f=K3,[K2]=2340,[K1]=K5,[K5]="T",x=39,y=314,},xend=40,yend=314,},x=29,y=314,},},[K2]=1694,[K1]=K4,x=27,y=314,},[K2]=1690,[K1]=K7,x=10,y=314,},[K19]={[1]={f=K3,[K10]="T",[K2]=1691,[K1]=K10,x=19,y=314,},xend=20,yend=314,},[K2]=1697,[K1]=K20,x=4,y=315,},[2]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1699,[K1]="any",x=19,y=315,},},[K2]=1700,[K1]=K4,x=26,y=315,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,keys={f=K3,[K2]=1702,[K1]="any",x=29,y=315,},[K2]=1703,[K1]="map",[K29]={f=K3,[K2]=1704,[K1]="any",x=33,y=315,},x=28,xend=36,y=315,yend=315,},},[K2]=1701,[K1]=K4,x=26,y=315,},[K2]=1698,[K1]=K7,x=10,y=315,},},x=8,y=315,},[K[25]]={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={[K32]={f=K3,[K2]=2353,[K1]=K5,[K5]="A@36",x=25,y=317,},f=K3,[K2]=2354,[K1]=K33,x=24,y=317,},[2]={f=K3,[K2]=1710,[K1]=K12,x=31,y=317,},},[K2]=2355,[K1]=K4,x=39,y=317,},f=K3,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2356,[K1]=K5,[K5]="A@36",x=41,y=317,},},[K2]=2357,[K1]=K4,x=39,y=317,},[K2]=2358,[K1]=K7,x=12,y=317,},[K19]={[1]={f=K3,[K10]="A@36",[K2]=2352,[K1]=K10,x=21,y=317,},},[K2]=2359,[K1]=K20,x=4,y=318,},sort={f=K3,[K23]=true,t={args={f=K3,[K4]={[1]={[K32]={f=K3,[K2]=2364,[K1]=K5,[K5]="A@37",x=23,y=318,},f=K3,[K2]=2365,[K1]=K33,x=22,y=318,},[2]={f=K3,[K17]=T23,[K16]={[1]=K[31],},[K2]=2367,[K1]=K15,[K41]={[1]={f=K3,[K2]=2366,[K1]=K5,[K5]="A@37",x=42,y=318,},},x=29,y=318,},},[K2]=2368,[K1]=K4,x=4,y=320,},f=K3,[K6]=1,rets={f=K3,[K4]={},[K2]=1722,[K1]=K4,x=44,y=318,},[K2]=2369,[K1]=K7,x=10,y=318,},[K19]={[1]={f=K3,[K10]="A@37",[K2]=2363,[K1]=K10,x=19,y=318,},},[K2]=2370,[K1]=K20,x=4,y=320,},[K45]={f=K3,[K28]=true,[K2]=1745,[K1]="poly",[K18]={[1]={f=K3,t={args={f=K3,[K4]={[1]={[K32]={f=K3,[K2]=2372,[K1]=K5,[K5]="A",x=25,y=320,},f=K3,[K2]=1727,[K1]=K33,x=24,xend=26,y=320,yend=320,},[2]={f=K3,[K2]=1728,[K1]=K13,x=31,y=320,},[3]={f=K3,[K2]=1729,[K1]=K13,x=41,y=320,},},[K2]=1730,[K1]=K4,x=48,y=320,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2373,[K1]=K5,[K5]="A",x=50,y=320,},},[K2]=1731,[K1]=K4,x=48,y=320,},[K2]=1724,[K1]=K7,x=12,y=320,},[K19]={[1]={f=K3,[K10]="A",[K2]=1725,[K1]=K10,x=21,y=320,},xend=22,yend=320,},[K2]=1733,[K1]=K20,x=4,y=321,},[2]={f=K3,t={args={f=K3,[K4]={[1]={f=K3,[K2]=1738,[K1]=K[33],[K18]={[1]={f=K3,[K2]=2384,[K1]=K5,[K5]="A1",x=30,y=321,},[2]={f=K3,[K2]=2385,[K1]=K5,[K5]="A2",x=34,y=321,},},x=29,xend=36,y=321,yend=321,},},[K2]=1740,[K1]=K4,x=38,y=321,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2386,[K1]=K5,[K5]="A1",x=40,y=321,},[2]={f=K3,[K2]=2387,[K1]=K5,[K5]="A2",x=44,y=321,},},[K2]=1741,[K1]=K4,x=38,y=321,},[K2]=1734,[K1]=K7,x=12,y=321,},[K19]={[1]={f=K3,[K10]="A1",[K2]=1735,[K1]=K10,x=21,y=321,},[2]={f=K3,[K10]="A2",[K2]=1736,[K1]=K10,x=25,y=321,},xend=27,yend=321,},[K2]=1744,[K1]=K20,x=4,y=322,},[3]={f=K3,t={args={f=K3,[K4]={[1]={f=K3,[K2]=1751,[K1]=K[33],[K18]={[1]={f=K3,[K2]=2402,[K1]=K5,[K5]="A1",x=34,y=322,},[2]={f=K3,[K2]=2403,[K1]=K5,[K5]="A2",x=38,y=322,},[3]={f=K3,[K2]=2404,[K1]=K5,[K5]="A3",x=42,y=322,},},x=33,xend=44,y=322,yend=322,},},[K2]=1754,[K1]=K4,x=46,y=322,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2405,[K1]=K5,[K5]="A1",x=48,y=322,},[2]={f=K3,[K2]=2406,[K1]=K5,[K5]="A2",x=52,y=322,},[3]={f=K3,[K2]=2407,[K1]=K5,[K5]="A3",x=56,y=322,},},[K2]=1755,[K1]=K4,x=46,y=322,},[K2]=1746,[K1]=K7,x=12,y=322,},[K19]={[1]={f=K3,[K10]="A1",[K2]=1747,[K1]=K10,x=21,y=322,},[2]={f=K3,[K10]="A2",[K2]=1748,[K1]=K10,x=25,y=322,},[3]={f=K3,[K10]="A3",[K2]=1749,[K1]=K10,x=29,y=322,},xend=31,yend=322,},[K2]=1759,[K1]=K20,x=4,y=323,},[4]={f=K3,t={args={f=K3,[K4]={[1]={f=K3,[K2]=1766,[K1]=K[33],[K18]={[1]={f=K3,[K2]=2426,[K1]=K5,[K5]="A1",x=38,y=323,},[2]={f=K3,[K2]=2427,[K1]=K5,[K5]="A2",x=42,y=323,},[3]={f=K3,[K2]=2428,[K1]=K5,[K5]="A3",x=46,y=323,},[4]={f=K3,[K2]=2429,[K1]=K5,[K5]="A4",x=50,y=323,},},x=37,xend=52,y=323,yend=323,},},[K2]=1770,[K1]=K4,x=54,y=323,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2430,[K1]=K5,[K5]="A1",x=56,y=323,},[2]={f=K3,[K2]=2431,[K1]=K5,[K5]="A2",x=60,y=323,},[3]={f=K3,[K2]=2432,[K1]=K5,[K5]="A3",x=64,y=323,},[4]={f=K3,[K2]=2433,[K1]=K5,[K5]="A4",x=68,y=323,},},[K2]=1771,[K1]=K4,x=54,y=323,},[K2]=1760,[K1]=K7,x=12,y=323,},[K19]={[1]={f=K3,[K10]="A1",[K2]=1761,[K1]=K10,x=21,y=323,},[2]={f=K3,[K10]="A2",[K2]=1762,[K1]=K10,x=25,y=323,},[3]={f=K3,[K10]="A3",[K2]=1763,[K1]=K10,x=29,y=323,},[4]={f=K3,[K10]="A4",[K2]=1764,[K1]=K10,x=33,y=323,},xend=35,yend=323,},[K2]=1776,[K1]=K20,x=4,y=324,},[5]={f=K3,t={args={f=K3,[K4]={[1]={f=K3,[K2]=1784,[K1]=K[33],[K18]={[1]={f=K3,[K2]=2456,[K1]=K5,[K5]="A1",x=42,y=324,},[2]={f=K3,[K2]=2457,[K1]=K5,[K5]="A2",x=46,y=324,},[3]={f=K3,[K2]=2458,[K1]=K5,[K5]="A3",x=50,y=324,},[4]={f=K3,[K2]=2459,[K1]=K5,[K5]="A4",x=54,y=324,},[5]={f=K3,[K2]=2460,[K1]=K5,[K5]="A5",x=58,y=324,},},x=41,xend=60,y=324,yend=324,},},[K2]=1789,[K1]=K4,x=62,y=324,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2461,[K1]=K5,[K5]="A1",x=64,y=324,},[2]={f=K3,[K2]=2462,[K1]=K5,[K5]="A2",x=68,y=324,},[3]={f=K3,[K2]=2463,[K1]=K5,[K5]="A3",x=72,y=324,},[4]={f=K3,[K2]=2464,[K1]=K5,[K5]="A4",x=76,y=324,},[5]={f=K3,[K2]=2465,[K1]=K5,[K5]="A5",x=80,y=324,},},[K2]=1790,[K1]=K4,x=62,y=324,},[K2]=1777,[K1]=K7,x=12,y=324,},[K19]={[1]={f=K3,[K10]="A1",[K2]=1778,[K1]=K10,x=21,y=324,},[2]={f=K3,[K10]="A2",[K2]=1779,[K1]=K10,x=25,y=324,},[3]={f=K3,[K10]="A3",[K2]=1780,[K1]=K10,x=29,y=324,},[4]={f=K3,[K10]="A4",[K2]=1781,[K1]=K10,x=33,y=324,},[5]={f=K3,[K10]="A5",[K2]=1782,[K1]=K10,x=37,y=324,},xend=39,yend=324,},[K2]=1796,[K1]=K20,x=1,y=325,},},x=10,y=321,},},[K2]=1630,[K1]=K37,x=1,y=298,},f=K3,[K2]=1797,[K1]=K21,x=1,y=298,},},[K35]={t={[K34]=true,def={[K24]=K35,f=K8,[K31]={},[K30]={},[K2]=7,[K1]=K[12],x=1,y=11,},f=K8,[K2]=8,[K1]=K21,x=1,y=11,},},["tonumber"]={[K27]=K26,t={f=K3,[K2]=2147,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=2137,[K1]="any",x=23,y=407,},},[K2]=2138,[K1]=K4,x=27,y=407,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2140,[K1]=K13,x=29,y=407,},},[K2]=2139,[K1]=K4,x=27,y=407,},[K2]=2136,[K1]=K7,x=14,y=407,},[2]={args={f=K3,[K4]={[1]={f=K3,[K2]=2142,[K1]="any",x=23,y=408,},[2]={f=K3,[K2]=2143,[K1]=K12,x=28,y=408,},},[K2]=2144,[K1]=K4,x=36,y=408,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=2146,[K1]=K12,x=38,y=408,},},[K2]=2145,[K1]=K4,x=36,y=408,},[K2]=2141,[K1]=K7,x=14,y=408,},},x=12,y=408,},},["tostring"]={[K27]=K26,t={args={f=K3,[K4]={[1]={f=K3,[K2]=2149,[K1]="any",x=23,y=410,},},[K2]=2150,[K1]=K4,x=27,y=410,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2152,[K1]=K11,x=29,y=410,},},[K2]=2151,[K1]=K4,x=27,y=410,},[K2]=2148,[K1]=K7,x=14,y=410,},},type={[K27]=K26,t={args={f=K3,[K4]={[1]={f=K3,[K2]=2154,[K1]="any",x=19,y=411,},},[K2]=2155,[K1]=K4,x=23,y=411,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=2157,[K1]=K11,x=25,y=411,},},[K2]=2156,[K1]=K4,x=23,y=411,},[K2]=2153,[K1]=K7,x=10,y=411,},},[K[20]]={t=T14,[K[11]]=true,},utf8={[K28]=true,t={[K34]=true,def={[K24]="utf8",f=K3,[K31]={[1]="char",[2]="charpattern",[3]="codepoint",[4]="codes",[5]="len",[6]="offset",},[K30]={char={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1800,[K1]=K13,x=19,y=328,},},[K2]=1801,[K1]=K4,x=29,y=328,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1803,[K1]=K11,x=31,y=328,},},[K2]=1802,[K1]=K4,x=29,y=328,},[K2]=1799,[K1]=K7,x=10,y=328,},["charpattern"]={f=K3,[K2]=1804,[K1]=K11,x=17,y=329,},["codepoint"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1806,[K1]=K11,x=24,y=330,},[2]={f=K3,[K2]=1807,[K1]=K13,x=34,y=330,},[3]={f=K3,[K2]=1808,[K1]=K13,x=44,y=330,},[4]={f=K3,[K2]=1809,[K1]=K25,x=54,y=330,},},[K2]=1810,[K1]=K4,x=62,y=330,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1812,[K1]=K13,x=64,y=330,},},[K2]=1811,[K1]=K4,x=62,y=330,},[K2]=1805,[K1]=K7,x=15,y=330,},["codes"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1814,[K1]=K11,x=20,y=331,},[2]={f=K3,[K2]=1815,[K1]=K25,x=30,y=331,},},[K2]=1816,[K1]=K4,x=38,y=331,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={[1]={f=K3,[K2]=1819,[K1]=K11,x=50,y=331,},[2]={f=K3,[K2]=1820,[K1]=K13,x=60,y=331,},},[K2]=1821,[K1]=K4,x=67,y=331,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1823,[K1]=K13,x=70,y=331,},[2]={f=K3,[K2]=1824,[K1]=K13,x=78,y=331,},},[K2]=1822,[K1]=K4,x=67,y=331,},[K2]=1818,[K1]=K7,x=41,y=331,},},[K2]=1817,[K1]=K4,x=38,y=331,},[K2]=1813,[K1]=K7,x=11,y=331,},len={args={f=K3,[K4]={[1]={f=K3,[K2]=1826,[K1]=K11,x=18,y=332,},[2]={f=K3,[K2]=1827,[K1]=K13,x=28,y=332,},[3]={f=K3,[K2]=1828,[K1]=K13,x=38,y=332,},[4]={f=K3,[K2]=1829,[K1]=K25,x=48,y=332,},},[K2]=1830,[K1]=K4,x=56,y=332,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=1832,[K1]=K13,x=58,y=332,},},[K2]=1831,[K1]=K4,x=56,y=332,},[K2]=1825,[K1]=K7,x=9,y=332,},["offset"]={args={f=K3,[K4]={[1]={f=K3,[K2]=1834,[K1]=K11,x=21,y=333,},[2]={f=K3,[K2]=1835,[K1]=K13,x=29,y=333,},[3]={f=K3,[K2]=1836,[K1]=K13,x=39,y=333,},},[K2]=1837,[K1]=K4,x=46,y=333,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1839,[K1]=K13,x=48,y=333,},},[K2]=1838,[K1]=K4,x=46,y=333,},[K2]=1833,[K1]=K7,x=12,y=333,},},[K2]=1798,[K1]=K37,x=1,y=327,},f=K3,[K2]=1840,[K1]=K21,x=1,y=327,},},["xpcall"]={[K27]=K26,[K28]=true,t={args={f=K3,[K14]=true,[K4]={[1]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2165,[K1]="any",x=30,y=413,},},[K2]=2166,[K1]=K4,x=37,y=413,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2168,[K1]="any",x=39,y=413,},},[K2]=2167,[K1]=K4,x=37,y=413,},[K2]=2164,[K1]=K7,x=21,y=413,},[2]={f=K3,[K17]={def={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=1856,[K1]="any",x=44,y=360,},},[K2]=1857,[K1]=K4,x=48,y=360,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={},[K2]=1858,[K1]=K4,x=47,y=360,},[K2]=1855,[K1]=K7,x=30,y=360,},f=K3,[K2]=1859,[K1]=K21,x=30,y=360,},[K16]={[1]="XpcallMsghFunction",},[K2]=2169,[K1]=K15,x=48,y=413,},[3]={f=K3,[K2]=2170,[K1]="any",x=68,y=413,},},[K2]=2171,[K1]=K4,x=75,y=413,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=2173,[K1]=K25,x=77,y=413,},[2]={f=K3,[K2]=2174,[K1]="any",x=86,y=413,},},[K2]=2172,[K1]=K4,x=75,y=413,},[K36]="xpcall",[K2]=2163,[K1]=K7,x=12,y=413,},},}
+T1.def={[K24]="FILE",f=K3,[K31]={[1]=K[2],[2]=K[3],[3]=K43,[4]=K[4],[5]=K[5],[6]="read",[7]="seek",[8]="setvbuf",[9]=K[6],},[K30]={[K[2]]=T2,[K[3]]=T3,[K43]={args={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=566,[K1]=K15,x=20,y=37,},},[K2]=567,[K1]=K4,x=25,y=37,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=569,[K1]=K25,x=27,y=37,},[2]={f=K3,[K2]=570,[K1]=K11,x=36,y=37,},[3]={f=K3,[K2]=571,[K1]=K12,x=44,y=37,},},[K2]=568,[K1]=K4,x=25,y=37,},[K2]=565,[K1]=K7,x=11,y=37,},[K[4]]={args={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=573,[K1]=K15,x=20,y=38,},},[K2]=574,[K1]=K4,x=4,y=40,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={},[K2]=575,[K1]=K4,x=24,y=38,},[K2]=572,[K1]=K7,x=11,y=38,},[K[5]]={f=K3,[K2]=593,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=577,[K1]=K15,x=20,y=40,},},[K2]=578,[K1]=K4,x=25,y=40,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=581,[K1]=K4,x=38,y=40,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=583,[K1]=K11,x=41,y=40,},},[K2]=582,[K1]=K4,x=38,y=40,},[K2]=580,[K1]=K7,x=28,y=40,},},[K2]=579,[K1]=K4,x=25,y=40,},[K2]=576,[K1]=K7,x=11,y=40,},[2]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=585,[K1]=K15,x=20,y=41,},[2]={f=K3,[K17]=T4,[K16]={[1]=K50,},[K2]=586,[K1]=K15,x=26,y=41,},},[K2]=587,[K1]=K4,x=44,y=41,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=590,[K1]=K4,x=57,y=41,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=592,[K1]=K13,x=60,y=41,},},[K2]=591,[K1]=K4,x=57,y=41,},[K2]=589,[K1]=K7,x=47,y=41,},},[K2]=588,[K1]=K4,x=44,y=41,},[K2]=584,[K1]=K7,x=11,y=41,},[3]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=595,[K1]=K15,x=20,y=42,},[2]={f=K3,[K2]=597,[K1]=K22,[K18]={[1]={f=K3,[K2]=596,[K1]=K13,x=27,y=42,},[2]={f=K3,[K17]=T5,[K16]={[1]=K49,},[K2]=598,[K1]=K15,x=36,y=42,},},x=27,y=42,},},[K2]=599,[K1]=K4,x=55,y=42,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=602,[K1]=K4,x=68,y=42,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=604,[K1]=K11,x=71,y=42,},},[K2]=603,[K1]=K4,x=68,y=42,},[K2]=601,[K1]=K7,x=58,y=42,},},[K2]=600,[K1]=K4,x=55,y=42,},[K2]=594,[K1]=K7,x=11,y=42,},[4]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=606,[K1]=K15,x=20,y=43,},[2]={f=K3,[K2]=608,[K1]=K22,[K18]={[1]={f=K3,[K2]=607,[K1]=K13,x=27,y=43,},[2]={f=K3,[K17]=T6,[K16]={[1]=K[7],},[K2]=609,[K1]=K15,x=36,y=43,},},x=27,y=43,},},[K2]=610,[K1]=K4,x=49,y=43,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=613,[K1]=K4,x=62,y=43,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=616,[K1]=K22,[K18]={[1]={f=K3,[K2]=615,[K1]=K11,x=66,y=43,},[2]={f=K3,[K2]=617,[K1]=K13,x=75,y=43,},},x=66,y=43,},},[K2]=614,[K1]=K4,x=62,y=43,},[K2]=612,[K1]=K7,x=52,y=43,},},[K2]=611,[K1]=K4,x=49,y=43,},[K2]=605,[K1]=K7,x=11,y=43,},[5]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=619,[K1]=K15,x=20,y=44,},[2]={f=K3,[K2]=621,[K1]=K22,[K18]={[1]={f=K3,[K2]=620,[K1]=K13,x=27,y=44,},[2]={f=K3,[K2]=622,[K1]=K11,x=36,y=44,},},x=27,y=44,},},[K2]=623,[K1]=K4,x=47,y=44,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={args={f=K3,[K4]={},[K2]=626,[K1]=K4,x=60,y=44,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=628,[K1]=K11,x=63,y=44,},},[K2]=627,[K1]=K4,x=60,y=44,},[K2]=625,[K1]=K7,x=50,y=44,},},[K2]=624,[K1]=K4,x=47,y=44,},[K2]=618,[K1]=K7,x=11,y=44,},},x=9,y=41,},read={f=K3,[K2]=640,[K1]="poly",[K18]={[1]={args={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=630,[K1]=K15,x=19,y=46,},},[K2]=631,[K1]=K4,x=24,y=46,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=633,[K1]=K11,x=26,y=46,},},[K2]=632,[K1]=K4,x=24,y=46,},[K2]=629,[K1]=K7,x=10,y=46,},[2]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=635,[K1]=K15,x=19,y=47,},[2]={f=K3,[K17]=T4,[K16]={[1]=K50,},[K2]=636,[K1]=K15,x=25,y=47,},},[K2]=637,[K1]=K4,x=43,y=47,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=639,[K1]=K13,x=45,y=47,},},[K2]=638,[K1]=K4,x=43,y=47,},[K2]=634,[K1]=K7,x=10,y=47,},[3]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=642,[K1]=K15,x=19,y=48,},[2]={f=K3,[K2]=644,[K1]=K22,[K18]={[1]={f=K3,[K2]=643,[K1]=K13,x=26,y=48,},[2]={f=K3,[K17]=T5,[K16]={[1]=K49,},[K2]=645,[K1]=K15,x=35,y=48,},},x=26,y=48,},},[K2]=646,[K1]=K4,x=54,y=48,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=648,[K1]=K11,x=56,y=48,},},[K2]=647,[K1]=K4,x=54,y=48,},[K2]=641,[K1]=K7,x=10,y=48,},[4]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=650,[K1]=K15,x=19,y=49,},[2]={f=K3,[K2]=652,[K1]=K22,[K18]={[1]={f=K3,[K2]=651,[K1]=K13,x=26,y=49,},[2]={f=K3,[K17]=T6,[K16]={[1]=K[7],},[K2]=653,[K1]=K15,x=35,y=49,},},x=26,y=49,},},[K2]=654,[K1]=K4,x=48,y=49,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=657,[K1]=K22,[K18]={[1]={f=K3,[K2]=656,[K1]=K11,x=52,y=49,},[2]={f=K3,[K2]=658,[K1]=K13,x=61,y=49,},},x=52,y=49,},},[K2]=655,[K1]=K4,x=48,y=49,},[K2]=649,[K1]=K7,x=10,y=49,},[5]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=660,[K1]=K15,x=19,y=50,},[2]={f=K3,[K2]=662,[K1]=K22,[K18]={[1]={f=K3,[K2]=661,[K1]=K13,x=26,y=50,},[2]={f=K3,[K2]=663,[K1]=K11,x=35,y=50,},},x=26,y=50,},},[K2]=664,[K1]=K4,x=46,y=50,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=666,[K1]=K11,x=49,y=50,},},[K2]=665,[K1]=K4,x=46,y=50,},[K2]=659,[K1]=K7,x=10,y=50,},},x=8,y=47,},seek={args={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=668,[K1]=K15,x=19,y=52,},[2]={f=K3,[K17]=T2,[K16]={[1]=K[2],},[K2]=669,[K1]=K15,x=27,y=52,},[3]={f=K3,[K2]=670,[K1]=K12,x=41,y=52,},},[K2]=671,[K1]=K4,x=49,y=52,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K2]=673,[K1]=K12,x=51,y=52,},[2]={f=K3,[K2]=674,[K1]=K11,x=60,y=52,},},[K2]=672,[K1]=K4,x=49,y=52,},[K2]=667,[K1]=K7,x=10,y=52,},["setvbuf"]={args={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=676,[K1]=K15,x=22,y=53,},[2]={f=K3,[K17]=T3,[K16]={[1]=K[3],},[K2]=677,[K1]=K15,x=28,y=53,},[3]={f=K3,[K2]=678,[K1]=K12,x=43,y=53,},},[K2]=679,[K1]=K4,x=4,y=55,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={},[K2]=680,[K1]=K4,x=50,y=53,},[K2]=675,[K1]=K7,x=13,y=53,},[K[6]]={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=682,[K1]=K15,x=20,y=55,},[2]={f=K3,[K2]=684,[K1]=K22,[K18]={[1]={f=K3,[K2]=683,[K1]=K11,x=27,y=55,},[2]={f=K3,[K2]=685,[K1]=K13,x=36,y=55,},},x=27,y=55,},},[K2]=686,[K1]=K4,x=47,y=55,},f=K3,[K9]=false,[K6]=1,rets={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=688,[K1]=K15,x=49,y=55,},[2]={f=K3,[K2]=689,[K1]=K11,x=55,y=55,},},[K2]=687,[K1]=K4,x=47,y=55,},[K2]=681,[K1]=K7,x=11,y=55,},},[K[8]]={},[K[9]]=true,["is_userdata"]=true,["meta_field_order"]={[1]=K[10],},["meta_fields"]={[K[10]]={args={f=K3,[K4]={[1]={f=K3,[K17]=T1,[K16]={[1]="FILE",},[K2]=691,[K1]=K15,x=33,y=57,},},[K2]=692,[K1]=K4,x=1,y=58,},f=K3,[K46]=true,[K9]=false,[K6]=1,rets={f=K3,[K4]={},[K2]=693,[K1]=K4,x=37,y=57,},[K2]=690,[K1]=K7,x=24,y=57,},},[K2]=559,[K1]=K37,x=1,y=26,}
 T2.def={[K24]=K[2],[K38]={cur=true,["end"]=true,set=true,},f=K3,[K2]=561,[K1]="enum",x=4,y=29,}
 T3.def={[K24]=K[3],[K38]={full=true,line=true,no=true,},f=K3,[K2]=563,[K1]="enum",x=4,y=33,}
-T4.def={[K24]=K48,[K38]={["*n"]=true,n=true,},f=K3,[K2]=555,[K1]="enum",x=1,y=18,}
-T5.def={[K24]=K47,[K38]={["*L"]=true,["*a"]=true,["*l"]=true,L=true,a=true,l=true,},f=K3,[K2]=553,[K1]="enum",x=1,y=14,}
-T6.def={[K24]=K49,[K38]={["*L"]=true,["*a"]=true,["*l"]=true,["*n"]=true,L=true,a=true,l=true,n=true,},f=K3,[K2]=557,[K1]="enum",x=1,y=22,}
+T4.def={[K24]=K50,[K38]={["*n"]=true,n=true,},f=K3,[K2]=555,[K1]="enum",x=1,y=18,}
+T5.def={[K24]=K49,[K38]={["*L"]=true,["*a"]=true,["*l"]=true,L=true,a=true,l=true,},f=K3,[K2]=553,[K1]="enum",x=1,y=14,}
+T6.def={[K24]=K[7],[K38]={["*L"]=true,["*a"]=true,["*l"]=true,["*n"]=true,L=true,a=true,l=true,n=true,},f=K3,[K2]=557,[K1]="enum",x=1,y=22,}
 T7.def={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=697,[K1]="any",x=29,y=61,},},[K2]=698,[K1]=K4,x=36,y=61,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=700,[K1]="any",x=38,y=61,},},[K2]=699,[K1]=K4,x=36,y=61,},[K2]=696,[K1]=K7,x=20,y=61,}
 T8.def={args={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=775,[K1]="any",x=32,y=97,},},[K2]=776,[K1]=K4,x=39,y=97,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K14]=true,[K4]={[1]={f=K3,[K2]=778,[K1]="any",x=40,y=97,},},[K2]=777,[K1]=K4,x=39,y=97,},[K2]=774,[K1]=K7,x=23,y=97,}
-T9.def={[K24]=K43,f=K3,[K31]={[1]="name",[2]="namewhat",[3]="source",[4]="short_src",[5]="linedefined",[6]="lastlinedefined",[7]="what",[8]="currentline",[9]="istailcall",[10]="nups",[11]="nparams",[12]="isvararg",[13]="func",[14]="activelines",},[K30]={["activelines"]={f=K3,keys={f=K3,[K2]=762,[K1]=K12,x=21,y=88,},[K2]=763,[K1]="map",[K29]={f=K3,[K2]=764,[K1]=K25,x=29,y=88,},x=20,xend=36,y=88,yend=88,},["currentline"]={f=K3,[K2]=756,[K1]=K12,x=20,y=82,},func={f=K3,[K2]=761,[K1]="any",x=13,y=87,},["istailcall"]={f=K3,[K2]=757,[K1]=K25,x=19,y=83,},["isvararg"]={f=K3,[K2]=760,[K1]=K25,x=17,y=86,},["lastlinedefined"]={f=K3,[K2]=754,[K1]=K12,x=24,y=80,},["linedefined"]={f=K3,[K2]=753,[K1]=K12,x=20,y=79,},name={f=K3,[K2]=749,[K1]=K11,x=13,y=75,},["namewhat"]={f=K3,[K2]=750,[K1]=K11,x=17,y=76,},["nparams"]={f=K3,[K2]=759,[K1]=K12,x=16,y=85,},nups={f=K3,[K2]=758,[K1]=K12,x=13,y=84,},["short_src"]={f=K3,[K2]=752,[K1]=K11,x=18,y=78,},["source"]={f=K3,[K2]=751,[K1]=K11,x=15,y=77,},what={f=K3,[K2]=755,[K1]=K11,x=13,y=81,},},[K2]=748,[K1]=K37,x=4,y=74,}
-T10.def={[K24]=K[14],[K38]={call=true,["count"]=true,line=true,["return"]=true,["tail call"]=true,},f=K3,[K2]=766,[K1]="enum",x=4,y=91,}
-T11.def={args={f=K3,[K4]={[1]={f=K3,[K17]=T10,[K16]={[1]=K[14],},[K2]=769,[K1]=K15,x=33,y=95,},[2]={f=K3,[K2]=770,[K1]=K12,x=44,y=95,},},[K2]=771,[K1]=K4,x=4,y=97,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={},[K2]=772,[K1]=K4,x=51,y=95,},[K2]=768,[K1]=K7,x=24,y=95,}
-T12.def={f=K8,t={[K24]=K41,f=K8,[K31]={[1]="Mode",[2]="__call",[3]="__mode",[4]="__name",[5]="__tostring",[6]="__pairs",[7]="__index",[8]="__newindex",[9]="__gc",[10]=K[9],[11]="__add",[12]="__sub",[13]="__mul",[14]="__div",[15]="__idiv",[16]="__mod",[17]="__pow",[18]="__band",[19]="__bor",[20]="__bxor",[21]="__shl",[22]="__shr",[23]="__concat",[24]="__len",[25]="__unm",[26]="__bnot",[27]="__eq",[28]="__lt",[29]="__le",},[K30]={Mode=T13,["__add"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=263,[K1]=K5,[K5]="A@3",x=29,y=35,},[2]={f=K8,[K2]=264,[K1]=K5,[K5]="B@3",x=32,y=35,},},[K2]=265,[K1]=K4,x=34,y=35,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=266,[K1]=K5,[K5]="C@3",x=36,y=35,},},[K2]=267,[K1]=K4,x=34,y=35,},[K2]=268,[K1]=K7,x=11,y=35,},[K19]={[1]={f=K8,[K10]="A@3",[K2]=260,[K1]=K10,x=20,y=35,},[2]={f=K8,[K10]="B@3",[K2]=261,[K1]=K10,x=23,y=35,},[3]={f=K8,[K10]="C@3",[K2]=262,[K1]=K10,x=26,y=35,},},[K2]=269,[K1]=K20,x=4,y=36,},["__band"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=375,[K1]=K5,[K5]="A@10",x=30,y=42,},[2]={f=K8,[K2]=376,[K1]=K5,[K5]="B@10",x=33,y=42,},},[K2]=377,[K1]=K4,x=35,y=42,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=378,[K1]=K5,[K5]="C@10",x=37,y=42,},},[K2]=379,[K1]=K4,x=35,y=42,},[K2]=380,[K1]=K7,x=12,y=42,},[K19]={[1]={f=K8,[K10]="A@10",[K2]=372,[K1]=K10,x=21,y=42,},[2]={f=K8,[K10]="B@10",[K2]=373,[K1]=K10,x=24,y=42,},[3]={f=K8,[K10]="C@10",[K2]=374,[K1]=K10,x=27,y=42,},},[K2]=381,[K1]=K20,x=4,y=43,},["__bnot"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=479,[K1]=K5,[K5]="T",x=24,y=51,},},[K2]=198,[K1]=K4,x=26,y=51,},f=K8,[K6]=1,rets={f=K8,[K4]={[1]={f=K8,[K2]=482,[K1]=K5,[K5]="A@18",x=28,y=51,},},[K2]=483,[K1]=K4,x=26,y=51,},[K2]=484,[K1]=K7,x=12,y=51,},[K19]={[1]={f=K8,[K10]="A@18",[K2]=481,[K1]=K10,x=21,y=51,},},[K2]=485,[K1]=K20,x=4,y=53,},["__bor"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=391,[K1]=K5,[K5]="A@11",x=29,y=43,},[2]={f=K8,[K2]=392,[K1]=K5,[K5]="B@11",x=32,y=43,},},[K2]=393,[K1]=K4,x=34,y=43,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=394,[K1]=K5,[K5]="C@11",x=36,y=43,},},[K2]=395,[K1]=K4,x=34,y=43,},[K2]=396,[K1]=K7,x=11,y=43,},[K19]={[1]={f=K8,[K10]="A@11",[K2]=388,[K1]=K10,x=20,y=43,},[2]={f=K8,[K10]="B@11",[K2]=389,[K1]=K10,x=23,y=43,},[3]={f=K8,[K10]="C@11",[K2]=390,[K1]=K10,x=26,y=43,},},[K2]=397,[K1]=K20,x=4,y=44,},["__bxor"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=407,[K1]=K5,[K5]="A@12",x=30,y=44,},[2]={f=K8,[K2]=408,[K1]=K5,[K5]="B@12",x=33,y=44,},},[K2]=409,[K1]=K4,x=35,y=44,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=410,[K1]=K5,[K5]="C@12",x=37,y=44,},},[K2]=411,[K1]=K4,x=35,y=44,},[K2]=412,[K1]=K7,x=12,y=44,},[K19]={[1]={f=K8,[K10]="A@12",[K2]=404,[K1]=K10,x=21,y=44,},[2]={f=K8,[K10]="B@12",[K2]=405,[K1]=K10,x=24,y=44,},[3]={f=K8,[K10]="C@12",[K2]=406,[K1]=K10,x=27,y=44,},},[K2]=413,[K1]=K20,x=4,y=45,},["__call"]={args={f=K8,[K14]=true,[K4]={[1]={f=K8,[K2]=236,[K1]=K5,[K5]="T",x=21,y=23,},[2]={f=K8,[K2]=18,[K1]="any",x=24,y=23,},},[K2]=19,[K1]=K4,x=31,y=23,},f=K8,[K9]=false,[K6]=1,rets={f=K8,[K14]=true,[K4]={[1]={f=K8,[K2]=21,[K1]="any",x=33,y=23,},},[K2]=20,[K1]=K4,x=31,y=23,},[K2]=16,[K1]=K7,x=12,y=23,},[K[9]]={args={f=K8,[K4]={[1]={f=K8,[K2]=253,[K1]=K5,[K5]="T",x=22,y=33,},},[K2]=49,[K1]=K4,x=4,y=35,},f=K8,[K9]=false,[K6]=1,rets={f=K8,[K4]={},[K2]=50,[K1]=K4,x=23,y=33,},[K2]=47,[K1]=K7,x=13,y=33,},["__concat"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=455,[K1]=K5,[K5]="A@15",x=32,y=47,},[2]={f=K8,[K2]=456,[K1]=K5,[K5]="B@15",x=35,y=47,},},[K2]=457,[K1]=K4,x=37,y=47,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=458,[K1]=K5,[K5]="C@15",x=39,y=47,},},[K2]=459,[K1]=K4,x=37,y=47,},[K2]=460,[K1]=K7,x=14,y=47,},[K19]={[1]={f=K8,[K10]="A@15",[K2]=452,[K1]=K10,x=23,y=47,},[2]={f=K8,[K10]="B@15",[K2]=453,[K1]=K10,x=26,y=47,},[3]={f=K8,[K10]="C@15",[K2]=454,[K1]=K10,x=29,y=47,},},[K2]=461,[K1]=K20,x=4,y=49,},["__div"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=311,[K1]=K5,[K5]="A@6",x=29,y=38,},[2]={f=K8,[K2]=312,[K1]=K5,[K5]="B@6",x=32,y=38,},},[K2]=313,[K1]=K4,x=34,y=38,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=314,[K1]=K5,[K5]="C@6",x=36,y=38,},},[K2]=315,[K1]=K4,x=34,y=38,},[K2]=316,[K1]=K7,x=11,y=38,},[K19]={[1]={f=K8,[K10]="A@6",[K2]=308,[K1]=K10,x=20,y=38,},[2]={f=K8,[K10]="B@6",[K2]=309,[K1]=K10,x=23,y=38,},[3]={f=K8,[K10]="C@6",[K2]=310,[K1]=K10,x=26,y=38,},},[K2]=317,[K1]=K20,x=4,y=39,},["__eq"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=492,[K1]=K5,[K5]="A@19",x=25,y=53,},[2]={f=K8,[K2]=493,[K1]=K5,[K5]="B@19",x=28,y=53,},},[K2]=494,[K1]=K4,x=30,y=53,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=209,[K1]=K25,x=32,y=53,},},[K2]=208,[K1]=K4,x=30,y=53,},[K2]=495,[K1]=K7,x=10,y=53,},[K19]={[1]={f=K8,[K10]="A@19",[K2]=490,[K1]=K10,x=19,y=53,},[2]={f=K8,[K10]="B@19",[K2]=491,[K1]=K10,x=22,y=53,},},[K2]=496,[K1]=K20,x=4,y=54,},["__gc"]={args={f=K8,[K4]={[1]={f=K8,[K2]=252,[K1]=K5,[K5]="T",x=19,y=32,},},[K2]=45,[K1]=K4,x=4,y=33,},f=K8,[K9]=false,[K6]=1,rets={f=K8,[K4]={},[K2]=46,[K1]=K4,x=20,y=32,},[K2]=43,[K1]=K7,x=10,y=32,},["__idiv"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=327,[K1]=K5,[K5]="A@7",x=30,y=39,},[2]={f=K8,[K2]=328,[K1]=K5,[K5]="B@7",x=33,y=39,},},[K2]=329,[K1]=K4,x=35,y=39,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=330,[K1]=K5,[K5]="C@7",x=37,y=39,},},[K2]=331,[K1]=K4,x=35,y=39,},[K2]=332,[K1]=K7,x=12,y=39,},[K19]={[1]={f=K8,[K10]="A@7",[K2]=324,[K1]=K10,x=21,y=39,},[2]={f=K8,[K10]="B@7",[K2]=325,[K1]=K10,x=24,y=39,},[3]={f=K8,[K10]="C@7",[K2]=326,[K1]=K10,x=27,y=39,},},[K2]=333,[K1]=K20,x=4,y=40,},["__index"]={f=K8,[K2]=41,[K1]="any",x=13,y=29,},["__le"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=514,[K1]=K5,[K5]="A@21",x=25,y=55,},[2]={f=K8,[K2]=515,[K1]=K5,[K5]="B@21",x=28,y=55,},},[K2]=516,[K1]=K4,x=30,y=55,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=227,[K1]=K25,x=32,y=55,},},[K2]=226,[K1]=K4,x=30,y=55,},[K2]=517,[K1]=K7,x=10,y=55,},[K19]={[1]={f=K8,[K10]="A@21",[K2]=512,[K1]=K10,x=19,y=55,},[2]={f=K8,[K10]="B@21",[K2]=513,[K1]=K10,x=22,y=55,},},[K2]=518,[K1]=K20,x=1,y=56,},["__len"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=463,[K1]=K5,[K5]="T",x=23,y=49,},},[K2]=184,[K1]=K4,x=25,y=49,},f=K8,[K6]=1,rets={f=K8,[K4]={[1]={f=K8,[K2]=466,[K1]=K5,[K5]="A@16",x=27,y=49,},},[K2]=467,[K1]=K4,x=25,y=49,},[K2]=468,[K1]=K7,x=11,y=49,},[K19]={[1]={f=K8,[K10]="A@16",[K2]=465,[K1]=K10,x=20,y=49,},},[K2]=469,[K1]=K20,x=4,y=50,},["__lt"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=503,[K1]=K5,[K5]="A@20",x=25,y=54,},[2]={f=K8,[K2]=504,[K1]=K5,[K5]="B@20",x=28,y=54,},},[K2]=505,[K1]=K4,x=30,y=54,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=218,[K1]=K25,x=32,y=54,},},[K2]=217,[K1]=K4,x=30,y=54,},[K2]=506,[K1]=K7,x=10,y=54,},[K19]={[1]={f=K8,[K10]="A@20",[K2]=501,[K1]=K10,x=19,y=54,},[2]={f=K8,[K10]="B@20",[K2]=502,[K1]=K10,x=22,y=54,},},[K2]=507,[K1]=K20,x=4,y=55,},["__mod"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=343,[K1]=K5,[K5]="A@8",x=29,y=40,},[2]={f=K8,[K2]=344,[K1]=K5,[K5]="B@8",x=32,y=40,},},[K2]=345,[K1]=K4,x=34,y=40,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=346,[K1]=K5,[K5]="C@8",x=36,y=40,},},[K2]=347,[K1]=K4,x=34,y=40,},[K2]=348,[K1]=K7,x=11,y=40,},[K19]={[1]={f=K8,[K10]="A@8",[K2]=340,[K1]=K10,x=20,y=40,},[2]={f=K8,[K10]="B@8",[K2]=341,[K1]=K10,x=23,y=40,},[3]={f=K8,[K10]="C@8",[K2]=342,[K1]=K10,x=26,y=40,},},[K2]=349,[K1]=K20,x=4,y=41,},["__mode"]={f=K8,[K17]=T13,[K16]={[1]="Mode",},[K2]=22,[K1]=K15,x=12,y=24,},["__mul"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=295,[K1]=K5,[K5]="A@5",x=29,y=37,},[2]={f=K8,[K2]=296,[K1]=K5,[K5]="B@5",x=32,y=37,},},[K2]=297,[K1]=K4,x=34,y=37,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=298,[K1]=K5,[K5]="C@5",x=36,y=37,},},[K2]=299,[K1]=K4,x=34,y=37,},[K2]=300,[K1]=K7,x=11,y=37,},[K19]={[1]={f=K8,[K10]="A@5",[K2]=292,[K1]=K10,x=20,y=37,},[2]={f=K8,[K10]="B@5",[K2]=293,[K1]=K10,x=23,y=37,},[3]={f=K8,[K10]="C@5",[K2]=294,[K1]=K10,x=26,y=37,},},[K2]=301,[K1]=K20,x=4,y=38,},["__name"]={f=K8,[K2]=23,[K1]=K11,x=12,y=25,},["__newindex"]={f=K8,[K2]=42,[K1]="any",x=16,y=30,},["__pairs"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=240,[K1]=K5,[K5]="T",x=28,y=27,},},[K2]=33,[K1]=K4,x=30,y=27,},f=K8,[K6]=1,rets={f=K8,[K4]={[1]={args={f=K8,[K4]={},[K2]=36,[K1]=K4,x=42,y=27,},f=K8,[K6]=0,rets={f=K8,[K4]={[1]={f=K8,[K2]=245,[K1]=K5,[K5]="K@2",x=45,y=27,},[2]={f=K8,[K2]=246,[K1]=K5,[K5]="V@2",x=48,y=27,},},[K2]=247,[K1]=K4,x=42,y=27,},[K2]=248,[K1]=K7,x=32,y=27,},},[K2]=249,[K1]=K4,x=30,y=27,},[K2]=250,[K1]=K7,x=13,y=27,},[K19]={[1]={f=K8,[K10]="K@2",[K2]=243,[K1]=K10,x=22,y=27,},[2]={f=K8,[K10]="V@2",[K2]=244,[K1]=K10,x=25,y=27,},},[K2]=251,[K1]=K20,x=4,y=29,},["__pow"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=359,[K1]=K5,[K5]="A@9",x=29,y=41,},[2]={f=K8,[K2]=360,[K1]=K5,[K5]="B@9",x=32,y=41,},},[K2]=361,[K1]=K4,x=34,y=41,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=362,[K1]=K5,[K5]="C@9",x=36,y=41,},},[K2]=363,[K1]=K4,x=34,y=41,},[K2]=364,[K1]=K7,x=11,y=41,},[K19]={[1]={f=K8,[K10]="A@9",[K2]=356,[K1]=K10,x=20,y=41,},[2]={f=K8,[K10]="B@9",[K2]=357,[K1]=K10,x=23,y=41,},[3]={f=K8,[K10]="C@9",[K2]=358,[K1]=K10,x=26,y=41,},},[K2]=365,[K1]=K20,x=4,y=42,},["__shl"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=423,[K1]=K5,[K5]="A@13",x=29,y=45,},[2]={f=K8,[K2]=424,[K1]=K5,[K5]="B@13",x=32,y=45,},},[K2]=425,[K1]=K4,x=34,y=45,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=426,[K1]=K5,[K5]="C@13",x=36,y=45,},},[K2]=427,[K1]=K4,x=34,y=45,},[K2]=428,[K1]=K7,x=11,y=45,},[K19]={[1]={f=K8,[K10]="A@13",[K2]=420,[K1]=K10,x=20,y=45,},[2]={f=K8,[K10]="B@13",[K2]=421,[K1]=K10,x=23,y=45,},[3]={f=K8,[K10]="C@13",[K2]=422,[K1]=K10,x=26,y=45,},},[K2]=429,[K1]=K20,x=4,y=46,},["__shr"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=439,[K1]=K5,[K5]="A@14",x=29,y=46,},[2]={f=K8,[K2]=440,[K1]=K5,[K5]="B@14",x=32,y=46,},},[K2]=441,[K1]=K4,x=34,y=46,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=442,[K1]=K5,[K5]="C@14",x=36,y=46,},},[K2]=443,[K1]=K4,x=34,y=46,},[K2]=444,[K1]=K7,x=11,y=46,},[K19]={[1]={f=K8,[K10]="A@14",[K2]=436,[K1]=K10,x=20,y=46,},[2]={f=K8,[K10]="B@14",[K2]=437,[K1]=K10,x=23,y=46,},[3]={f=K8,[K10]="C@14",[K2]=438,[K1]=K10,x=26,y=46,},},[K2]=445,[K1]=K20,x=4,y=47,},["__sub"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=279,[K1]=K5,[K5]="A@4",x=29,y=36,},[2]={f=K8,[K2]=280,[K1]=K5,[K5]="B@4",x=32,y=36,},},[K2]=281,[K1]=K4,x=34,y=36,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=282,[K1]=K5,[K5]="C@4",x=36,y=36,},},[K2]=283,[K1]=K4,x=34,y=36,},[K2]=284,[K1]=K7,x=11,y=36,},[K19]={[1]={f=K8,[K10]="A@4",[K2]=276,[K1]=K10,x=20,y=36,},[2]={f=K8,[K10]="B@4",[K2]=277,[K1]=K10,x=23,y=36,},[3]={f=K8,[K10]="C@4",[K2]=278,[K1]=K10,x=26,y=36,},},[K2]=285,[K1]=K20,x=4,y=37,},["__tostring"]={args={f=K8,[K4]={[1]={f=K8,[K2]=237,[K1]=K5,[K5]="T",x=25,y=26,},},[K2]=26,[K1]=K4,x=27,y=26,},f=K8,[K9]=false,[K6]=1,rets={f=K8,[K4]={[1]={f=K8,[K2]=28,[K1]=K11,x=29,y=26,},},[K2]=27,[K1]=K4,x=27,y=26,},[K2]=24,[K1]=K7,x=16,y=26,},["__unm"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=471,[K1]=K5,[K5]="T",x=23,y=50,},},[K2]=191,[K1]=K4,x=25,y=50,},f=K8,[K6]=1,rets={f=K8,[K4]={[1]={f=K8,[K2]=474,[K1]=K5,[K5]="A@17",x=27,y=50,},},[K2]=475,[K1]=K4,x=25,y=50,},[K2]=476,[K1]=K7,x=11,y=50,},[K19]={[1]={f=K8,[K10]="A@17",[K2]=473,[K1]=K10,x=20,y=50,},},[K2]=477,[K1]=K20,x=4,y=51,},},[K2]=13,[K1]=K37,x=1,y=18,},[K19]={[1]={f=K8,[K10]="T",[K2]=12,[K1]=K10,x=25,y=18,},xend=26,yend=18,},[K2]=229,[K1]=K20,x=1,y=18,}
+T9.def={[K24]=K44,f=K3,[K31]={[1]="name",[2]="namewhat",[3]="source",[4]="short_src",[5]="linedefined",[6]="lastlinedefined",[7]="what",[8]="currentline",[9]="istailcall",[10]="nups",[11]="nparams",[12]="isvararg",[13]="func",[14]="activelines",},[K30]={["activelines"]={f=K3,keys={f=K3,[K2]=762,[K1]=K12,x=21,y=88,},[K2]=763,[K1]="map",[K29]={f=K3,[K2]=764,[K1]=K25,x=29,y=88,},x=20,xend=36,y=88,yend=88,},["currentline"]={f=K3,[K2]=756,[K1]=K12,x=20,y=82,},func={f=K3,[K2]=761,[K1]="any",x=13,y=87,},["istailcall"]={f=K3,[K2]=757,[K1]=K25,x=19,y=83,},["isvararg"]={f=K3,[K2]=760,[K1]=K25,x=17,y=86,},["lastlinedefined"]={f=K3,[K2]=754,[K1]=K12,x=24,y=80,},["linedefined"]={f=K3,[K2]=753,[K1]=K12,x=20,y=79,},name={f=K3,[K2]=749,[K1]=K11,x=13,y=75,},["namewhat"]={f=K3,[K2]=750,[K1]=K11,x=17,y=76,},["nparams"]={f=K3,[K2]=759,[K1]=K12,x=16,y=85,},nups={f=K3,[K2]=758,[K1]=K12,x=13,y=84,},["short_src"]={f=K3,[K2]=752,[K1]=K11,x=18,y=78,},["source"]={f=K3,[K2]=751,[K1]=K11,x=15,y=77,},what={f=K3,[K2]=755,[K1]=K11,x=13,y=81,},},[K2]=748,[K1]=K37,x=4,y=74,}
+T10.def={[K24]=K[15],[K38]={call=true,["count"]=true,line=true,["return"]=true,["tail call"]=true,},f=K3,[K2]=766,[K1]="enum",x=4,y=91,}
+T11.def={args={f=K3,[K4]={[1]={f=K3,[K17]=T10,[K16]={[1]=K[15],},[K2]=769,[K1]=K15,x=33,y=95,},[2]={f=K3,[K2]=770,[K1]=K12,x=44,y=95,},},[K2]=771,[K1]=K4,x=4,y=97,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={},[K2]=772,[K1]=K4,x=51,y=95,},[K2]=768,[K1]=K7,x=24,y=95,}
+T12.def={f=K8,t={[K24]=K42,f=K8,[K[19]]={[K[18]]={[1]={[1]={text="--[[FIXME: function | table | anything with an __index metamethod]]",x=17,y=29,},},},},[K31]={[1]="Mode",[2]="__call",[3]="__mode",[4]="__name",[5]="__tostring",[6]="__pairs",[7]="__index",[8]=K[18],[9]="__gc",[10]=K[10],[11]="__add",[12]="__sub",[13]="__mul",[14]="__div",[15]="__idiv",[16]="__mod",[17]="__pow",[18]="__band",[19]="__bor",[20]="__bxor",[21]="__shl",[22]="__shr",[23]="__concat",[24]="__len",[25]="__unm",[26]="__bnot",[27]="__eq",[28]="__lt",[29]="__le",},[K30]={Mode=T13,["__add"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=263,[K1]=K5,[K5]="A@3",x=29,y=35,},[2]={f=K8,[K2]=264,[K1]=K5,[K5]="B@3",x=32,y=35,},},[K2]=265,[K1]=K4,x=34,y=35,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=266,[K1]=K5,[K5]="C@3",x=36,y=35,},},[K2]=267,[K1]=K4,x=34,y=35,},[K2]=268,[K1]=K7,x=11,y=35,},[K19]={[1]={f=K8,[K10]="A@3",[K2]=260,[K1]=K10,x=20,y=35,},[2]={f=K8,[K10]="B@3",[K2]=261,[K1]=K10,x=23,y=35,},[3]={f=K8,[K10]="C@3",[K2]=262,[K1]=K10,x=26,y=35,},},[K2]=269,[K1]=K20,x=4,y=36,},["__band"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=375,[K1]=K5,[K5]="A@10",x=30,y=42,},[2]={f=K8,[K2]=376,[K1]=K5,[K5]="B@10",x=33,y=42,},},[K2]=377,[K1]=K4,x=35,y=42,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=378,[K1]=K5,[K5]="C@10",x=37,y=42,},},[K2]=379,[K1]=K4,x=35,y=42,},[K2]=380,[K1]=K7,x=12,y=42,},[K19]={[1]={f=K8,[K10]="A@10",[K2]=372,[K1]=K10,x=21,y=42,},[2]={f=K8,[K10]="B@10",[K2]=373,[K1]=K10,x=24,y=42,},[3]={f=K8,[K10]="C@10",[K2]=374,[K1]=K10,x=27,y=42,},},[K2]=381,[K1]=K20,x=4,y=43,},["__bnot"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=479,[K1]=K5,[K5]="T",x=24,y=51,},},[K2]=198,[K1]=K4,x=26,y=51,},f=K8,[K6]=1,rets={f=K8,[K4]={[1]={f=K8,[K2]=482,[K1]=K5,[K5]="A@18",x=28,y=51,},},[K2]=483,[K1]=K4,x=26,y=51,},[K2]=484,[K1]=K7,x=12,y=51,},[K19]={[1]={f=K8,[K10]="A@18",[K2]=481,[K1]=K10,x=21,y=51,},},[K2]=485,[K1]=K20,x=4,y=53,},["__bor"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=391,[K1]=K5,[K5]="A@11",x=29,y=43,},[2]={f=K8,[K2]=392,[K1]=K5,[K5]="B@11",x=32,y=43,},},[K2]=393,[K1]=K4,x=34,y=43,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=394,[K1]=K5,[K5]="C@11",x=36,y=43,},},[K2]=395,[K1]=K4,x=34,y=43,},[K2]=396,[K1]=K7,x=11,y=43,},[K19]={[1]={f=K8,[K10]="A@11",[K2]=388,[K1]=K10,x=20,y=43,},[2]={f=K8,[K10]="B@11",[K2]=389,[K1]=K10,x=23,y=43,},[3]={f=K8,[K10]="C@11",[K2]=390,[K1]=K10,x=26,y=43,},},[K2]=397,[K1]=K20,x=4,y=44,},["__bxor"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=407,[K1]=K5,[K5]="A@12",x=30,y=44,},[2]={f=K8,[K2]=408,[K1]=K5,[K5]="B@12",x=33,y=44,},},[K2]=409,[K1]=K4,x=35,y=44,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=410,[K1]=K5,[K5]="C@12",x=37,y=44,},},[K2]=411,[K1]=K4,x=35,y=44,},[K2]=412,[K1]=K7,x=12,y=44,},[K19]={[1]={f=K8,[K10]="A@12",[K2]=404,[K1]=K10,x=21,y=44,},[2]={f=K8,[K10]="B@12",[K2]=405,[K1]=K10,x=24,y=44,},[3]={f=K8,[K10]="C@12",[K2]=406,[K1]=K10,x=27,y=44,},},[K2]=413,[K1]=K20,x=4,y=45,},["__call"]={args={f=K8,[K14]=true,[K4]={[1]={f=K8,[K2]=236,[K1]=K5,[K5]="T",x=21,y=23,},[2]={f=K8,[K2]=18,[K1]="any",x=24,y=23,},},[K2]=19,[K1]=K4,x=31,y=23,},f=K8,[K9]=false,[K6]=1,rets={f=K8,[K14]=true,[K4]={[1]={f=K8,[K2]=21,[K1]="any",x=33,y=23,},},[K2]=20,[K1]=K4,x=31,y=23,},[K2]=16,[K1]=K7,x=12,y=23,},[K[10]]={args={f=K8,[K4]={[1]={f=K8,[K2]=253,[K1]=K5,[K5]="T",x=22,y=33,},},[K2]=49,[K1]=K4,x=4,y=35,},f=K8,[K9]=false,[K6]=1,rets={f=K8,[K4]={},[K2]=50,[K1]=K4,x=23,y=33,},[K2]=47,[K1]=K7,x=13,y=33,},["__concat"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=455,[K1]=K5,[K5]="A@15",x=32,y=47,},[2]={f=K8,[K2]=456,[K1]=K5,[K5]="B@15",x=35,y=47,},},[K2]=457,[K1]=K4,x=37,y=47,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=458,[K1]=K5,[K5]="C@15",x=39,y=47,},},[K2]=459,[K1]=K4,x=37,y=47,},[K2]=460,[K1]=K7,x=14,y=47,},[K19]={[1]={f=K8,[K10]="A@15",[K2]=452,[K1]=K10,x=23,y=47,},[2]={f=K8,[K10]="B@15",[K2]=453,[K1]=K10,x=26,y=47,},[3]={f=K8,[K10]="C@15",[K2]=454,[K1]=K10,x=29,y=47,},},[K2]=461,[K1]=K20,x=4,y=49,},["__div"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=311,[K1]=K5,[K5]="A@6",x=29,y=38,},[2]={f=K8,[K2]=312,[K1]=K5,[K5]="B@6",x=32,y=38,},},[K2]=313,[K1]=K4,x=34,y=38,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=314,[K1]=K5,[K5]="C@6",x=36,y=38,},},[K2]=315,[K1]=K4,x=34,y=38,},[K2]=316,[K1]=K7,x=11,y=38,},[K19]={[1]={f=K8,[K10]="A@6",[K2]=308,[K1]=K10,x=20,y=38,},[2]={f=K8,[K10]="B@6",[K2]=309,[K1]=K10,x=23,y=38,},[3]={f=K8,[K10]="C@6",[K2]=310,[K1]=K10,x=26,y=38,},},[K2]=317,[K1]=K20,x=4,y=39,},["__eq"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=492,[K1]=K5,[K5]="A@19",x=25,y=53,},[2]={f=K8,[K2]=493,[K1]=K5,[K5]="B@19",x=28,y=53,},},[K2]=494,[K1]=K4,x=30,y=53,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=209,[K1]=K25,x=32,y=53,},},[K2]=208,[K1]=K4,x=30,y=53,},[K2]=495,[K1]=K7,x=10,y=53,},[K19]={[1]={f=K8,[K10]="A@19",[K2]=490,[K1]=K10,x=19,y=53,},[2]={f=K8,[K10]="B@19",[K2]=491,[K1]=K10,x=22,y=53,},},[K2]=496,[K1]=K20,x=4,y=54,},["__gc"]={args={f=K8,[K4]={[1]={f=K8,[K2]=252,[K1]=K5,[K5]="T",x=19,y=32,},},[K2]=45,[K1]=K4,x=4,y=33,},f=K8,[K9]=false,[K6]=1,rets={f=K8,[K4]={},[K2]=46,[K1]=K4,x=20,y=32,},[K2]=43,[K1]=K7,x=10,y=32,},["__idiv"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=327,[K1]=K5,[K5]="A@7",x=30,y=39,},[2]={f=K8,[K2]=328,[K1]=K5,[K5]="B@7",x=33,y=39,},},[K2]=329,[K1]=K4,x=35,y=39,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=330,[K1]=K5,[K5]="C@7",x=37,y=39,},},[K2]=331,[K1]=K4,x=35,y=39,},[K2]=332,[K1]=K7,x=12,y=39,},[K19]={[1]={f=K8,[K10]="A@7",[K2]=324,[K1]=K10,x=21,y=39,},[2]={f=K8,[K10]="B@7",[K2]=325,[K1]=K10,x=24,y=39,},[3]={f=K8,[K10]="C@7",[K2]=326,[K1]=K10,x=27,y=39,},},[K2]=333,[K1]=K20,x=4,y=40,},["__index"]={f=K8,[K2]=41,[K1]="any",x=13,y=29,},["__le"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=514,[K1]=K5,[K5]="A@21",x=25,y=55,},[2]={f=K8,[K2]=515,[K1]=K5,[K5]="B@21",x=28,y=55,},},[K2]=516,[K1]=K4,x=30,y=55,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=227,[K1]=K25,x=32,y=55,},},[K2]=226,[K1]=K4,x=30,y=55,},[K2]=517,[K1]=K7,x=10,y=55,},[K19]={[1]={f=K8,[K10]="A@21",[K2]=512,[K1]=K10,x=19,y=55,},[2]={f=K8,[K10]="B@21",[K2]=513,[K1]=K10,x=22,y=55,},},[K2]=518,[K1]=K20,x=1,y=56,},["__len"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=463,[K1]=K5,[K5]="T",x=23,y=49,},},[K2]=184,[K1]=K4,x=25,y=49,},f=K8,[K6]=1,rets={f=K8,[K4]={[1]={f=K8,[K2]=466,[K1]=K5,[K5]="A@16",x=27,y=49,},},[K2]=467,[K1]=K4,x=25,y=49,},[K2]=468,[K1]=K7,x=11,y=49,},[K19]={[1]={f=K8,[K10]="A@16",[K2]=465,[K1]=K10,x=20,y=49,},},[K2]=469,[K1]=K20,x=4,y=50,},["__lt"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=503,[K1]=K5,[K5]="A@20",x=25,y=54,},[2]={f=K8,[K2]=504,[K1]=K5,[K5]="B@20",x=28,y=54,},},[K2]=505,[K1]=K4,x=30,y=54,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=218,[K1]=K25,x=32,y=54,},},[K2]=217,[K1]=K4,x=30,y=54,},[K2]=506,[K1]=K7,x=10,y=54,},[K19]={[1]={f=K8,[K10]="A@20",[K2]=501,[K1]=K10,x=19,y=54,},[2]={f=K8,[K10]="B@20",[K2]=502,[K1]=K10,x=22,y=54,},},[K2]=507,[K1]=K20,x=4,y=55,},["__mod"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=343,[K1]=K5,[K5]="A@8",x=29,y=40,},[2]={f=K8,[K2]=344,[K1]=K5,[K5]="B@8",x=32,y=40,},},[K2]=345,[K1]=K4,x=34,y=40,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=346,[K1]=K5,[K5]="C@8",x=36,y=40,},},[K2]=347,[K1]=K4,x=34,y=40,},[K2]=348,[K1]=K7,x=11,y=40,},[K19]={[1]={f=K8,[K10]="A@8",[K2]=340,[K1]=K10,x=20,y=40,},[2]={f=K8,[K10]="B@8",[K2]=341,[K1]=K10,x=23,y=40,},[3]={f=K8,[K10]="C@8",[K2]=342,[K1]=K10,x=26,y=40,},},[K2]=349,[K1]=K20,x=4,y=41,},["__mode"]={f=K8,[K17]=T13,[K16]={[1]="Mode",},[K2]=22,[K1]=K15,x=12,y=24,},["__mul"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=295,[K1]=K5,[K5]="A@5",x=29,y=37,},[2]={f=K8,[K2]=296,[K1]=K5,[K5]="B@5",x=32,y=37,},},[K2]=297,[K1]=K4,x=34,y=37,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=298,[K1]=K5,[K5]="C@5",x=36,y=37,},},[K2]=299,[K1]=K4,x=34,y=37,},[K2]=300,[K1]=K7,x=11,y=37,},[K19]={[1]={f=K8,[K10]="A@5",[K2]=292,[K1]=K10,x=20,y=37,},[2]={f=K8,[K10]="B@5",[K2]=293,[K1]=K10,x=23,y=37,},[3]={f=K8,[K10]="C@5",[K2]=294,[K1]=K10,x=26,y=37,},},[K2]=301,[K1]=K20,x=4,y=38,},["__name"]={f=K8,[K2]=23,[K1]=K11,x=12,y=25,},[K[18]]={f=K8,[K2]=42,[K1]="any",x=16,y=30,},["__pairs"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=240,[K1]=K5,[K5]="T",x=28,y=27,},},[K2]=33,[K1]=K4,x=30,y=27,},f=K8,[K6]=1,rets={f=K8,[K4]={[1]={args={f=K8,[K4]={},[K2]=36,[K1]=K4,x=42,y=27,},f=K8,[K6]=0,rets={f=K8,[K4]={[1]={f=K8,[K2]=245,[K1]=K5,[K5]="K@2",x=45,y=27,},[2]={f=K8,[K2]=246,[K1]=K5,[K5]="V@2",x=48,y=27,},},[K2]=247,[K1]=K4,x=42,y=27,},[K2]=248,[K1]=K7,x=32,y=27,},},[K2]=249,[K1]=K4,x=30,y=27,},[K2]=250,[K1]=K7,x=13,y=27,},[K19]={[1]={f=K8,[K10]="K@2",[K2]=243,[K1]=K10,x=22,y=27,},[2]={f=K8,[K10]="V@2",[K2]=244,[K1]=K10,x=25,y=27,},},[K2]=251,[K1]=K20,x=4,y=29,},["__pow"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=359,[K1]=K5,[K5]="A@9",x=29,y=41,},[2]={f=K8,[K2]=360,[K1]=K5,[K5]="B@9",x=32,y=41,},},[K2]=361,[K1]=K4,x=34,y=41,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=362,[K1]=K5,[K5]="C@9",x=36,y=41,},},[K2]=363,[K1]=K4,x=34,y=41,},[K2]=364,[K1]=K7,x=11,y=41,},[K19]={[1]={f=K8,[K10]="A@9",[K2]=356,[K1]=K10,x=20,y=41,},[2]={f=K8,[K10]="B@9",[K2]=357,[K1]=K10,x=23,y=41,},[3]={f=K8,[K10]="C@9",[K2]=358,[K1]=K10,x=26,y=41,},},[K2]=365,[K1]=K20,x=4,y=42,},["__shl"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=423,[K1]=K5,[K5]="A@13",x=29,y=45,},[2]={f=K8,[K2]=424,[K1]=K5,[K5]="B@13",x=32,y=45,},},[K2]=425,[K1]=K4,x=34,y=45,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=426,[K1]=K5,[K5]="C@13",x=36,y=45,},},[K2]=427,[K1]=K4,x=34,y=45,},[K2]=428,[K1]=K7,x=11,y=45,},[K19]={[1]={f=K8,[K10]="A@13",[K2]=420,[K1]=K10,x=20,y=45,},[2]={f=K8,[K10]="B@13",[K2]=421,[K1]=K10,x=23,y=45,},[3]={f=K8,[K10]="C@13",[K2]=422,[K1]=K10,x=26,y=45,},},[K2]=429,[K1]=K20,x=4,y=46,},["__shr"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=439,[K1]=K5,[K5]="A@14",x=29,y=46,},[2]={f=K8,[K2]=440,[K1]=K5,[K5]="B@14",x=32,y=46,},},[K2]=441,[K1]=K4,x=34,y=46,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=442,[K1]=K5,[K5]="C@14",x=36,y=46,},},[K2]=443,[K1]=K4,x=34,y=46,},[K2]=444,[K1]=K7,x=11,y=46,},[K19]={[1]={f=K8,[K10]="A@14",[K2]=436,[K1]=K10,x=20,y=46,},[2]={f=K8,[K10]="B@14",[K2]=437,[K1]=K10,x=23,y=46,},[3]={f=K8,[K10]="C@14",[K2]=438,[K1]=K10,x=26,y=46,},},[K2]=445,[K1]=K20,x=4,y=47,},["__sub"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=279,[K1]=K5,[K5]="A@4",x=29,y=36,},[2]={f=K8,[K2]=280,[K1]=K5,[K5]="B@4",x=32,y=36,},},[K2]=281,[K1]=K4,x=34,y=36,},f=K8,[K6]=2,rets={f=K8,[K4]={[1]={f=K8,[K2]=282,[K1]=K5,[K5]="C@4",x=36,y=36,},},[K2]=283,[K1]=K4,x=34,y=36,},[K2]=284,[K1]=K7,x=11,y=36,},[K19]={[1]={f=K8,[K10]="A@4",[K2]=276,[K1]=K10,x=20,y=36,},[2]={f=K8,[K10]="B@4",[K2]=277,[K1]=K10,x=23,y=36,},[3]={f=K8,[K10]="C@4",[K2]=278,[K1]=K10,x=26,y=36,},},[K2]=285,[K1]=K20,x=4,y=37,},["__tostring"]={args={f=K8,[K4]={[1]={f=K8,[K2]=237,[K1]=K5,[K5]="T",x=25,y=26,},},[K2]=26,[K1]=K4,x=27,y=26,},f=K8,[K9]=false,[K6]=1,rets={f=K8,[K4]={[1]={f=K8,[K2]=28,[K1]=K11,x=29,y=26,},},[K2]=27,[K1]=K4,x=27,y=26,},[K2]=24,[K1]=K7,x=16,y=26,},["__unm"]={f=K8,[K23]=true,t={args={f=K8,[K4]={[1]={f=K8,[K2]=471,[K1]=K5,[K5]="T",x=23,y=50,},},[K2]=191,[K1]=K4,x=25,y=50,},f=K8,[K6]=1,rets={f=K8,[K4]={[1]={f=K8,[K2]=474,[K1]=K5,[K5]="A@17",x=27,y=50,},},[K2]=475,[K1]=K4,x=25,y=50,},[K2]=476,[K1]=K7,x=11,y=50,},[K19]={[1]={f=K8,[K10]="A@17",[K2]=473,[K1]=K10,x=20,y=50,},},[K2]=477,[K1]=K20,x=4,y=51,},},[K2]=13,[K1]=K37,x=1,y=18,},[K19]={[1]={f=K8,[K10]="T",[K2]=12,[K1]=K10,x=25,y=18,},xend=26,yend=18,},[K2]=229,[K1]=K20,x=1,y=18,}
 T13.def={[K24]="Mode",[K38]={k=true,kv=true,v=true,},f=K8,[K2]=14,[K1]="enum",x=4,y=19,}
-T14.def={[K24]=K[17],f=K8,[K31]={},[K30]={},[K[7]]={},[K[8]]=true,["is_userdata"]=true,[K2]=9,[K1]=K[11],x=1,y=14,}
-T15.def={[K24]=K45,[K38]={["*a"]=true,["*a+"]=true,["*a+b"]=true,["*ab"]=true,["*r"]=true,["*r+"]=true,["*r+b"]=true,["*rb"]=true,["*w"]=true,["*w+"]=true,["*w+b"]=true,["*wb"]=true,a=true,["a+"]=true,["a+b"]=true,ab=true,r=true,["r+"]=true,["r+b"]=true,rb=true,w=true,["w+"]=true,["w+b"]=true,wb=true,},f=K3,[K2]=953,[K1]="enum",x=4,y=134,}
+T14.def={[K24]=K[20],f=K8,[K31]={},[K30]={},[K[8]]={},[K[9]]=true,["is_userdata"]=true,[K2]=9,[K1]=K[12],x=1,y=14,}
+T15.def={[K24]=K47,[K38]={["*a"]=true,["*a+"]=true,["*a+b"]=true,["*ab"]=true,["*r"]=true,["*r+"]=true,["*r+b"]=true,["*rb"]=true,["*w"]=true,["*w+"]=true,["*w+b"]=true,["*wb"]=true,a=true,["a+"]=true,["a+b"]=true,ab=true,r=true,["r+"]=true,["r+b"]=true,rb=true,w=true,["w+"]=true,["w+b"]=true,wb=true,},f=K3,[K2]=953,[K1]="enum",x=4,y=134,}
 T16.def={args={f=K3,[K4]={},[K2]=1849,[K1]=K4,x=34,y=354,},f=K3,[K9]=false,[K6]=0,rets={f=K3,[K4]={[1]={f=K3,[K2]=1851,[K1]=K11,x=36,y=354,},},[K2]=1850,[K1]=K4,x=34,y=354,},[K2]=1848,[K1]=K7,x=24,y=354,}
 T17.def={f=K3,[K2]=1100,[K1]=K22,[K18]={[1]={f=K3,[K2]=1099,[K1]=K13,x=19,y=170,},[2]={f=K3,[K2]=1101,[K1]=K12,x=28,y=170,},},x=19,y=170,}
 T18[K17]=T17
-T18[K16]={[1]=K[18],}
-T19.def={[K24]=K[20],[K38]={["!*t"]=true,["*t"]=true,},f=K3,[K2]=1342,[K1]="enum",x=4,y=239,}
-T20.def={[K24]=K50,f=K3,[K31]={[1]="year",[2]="month",[3]="day",[4]="hour",[5]="min",[6]="sec",[7]="wday",[8]="yday",[9]="isdst",},[K30]={day={f=K3,[K2]=1334,[K1]=K12,x=12,y=230,},hour={f=K3,[K2]=1335,[K1]=K12,x=13,y=231,},["isdst"]={f=K3,[K2]=1340,[K1]=K25,x=14,y=236,},min={f=K3,[K2]=1336,[K1]=K12,x=12,y=232,},["month"]={f=K3,[K2]=1333,[K1]=K12,x=14,y=229,},sec={f=K3,[K2]=1337,[K1]=K12,x=12,y=233,},wday={f=K3,[K2]=1338,[K1]=K12,x=13,y=234,},yday={f=K3,[K2]=1339,[K1]=K12,x=13,y=235,},year={f=K3,[K2]=1332,[K1]=K12,x=13,y=228,},},[K2]=1331,[K1]=K37,x=4,y=227,}
-T21.def={f=K3,t={[K24]=K[24],[K32]=T22,f=K3,[K31]={[1]="n",},[K30]={n={f=K3,[K2]=1644,[K1]=K12,x=10,y=304,},},[K[7]]={[1]={[K32]=T22,f=K3,[K2]=1643,[K1]=K33,x=10,xend=12,y=302,yend=302,},},[K[8]]=true,[K2]=1641,[K1]=K37,x=4,y=301,},[K19]={[1]={f=K3,[K10]="A",[K2]=1640,[K1]=K10,x=21,y=301,},xend=22,yend=301,},[K2]=1645,[K1]=K20,x=4,y=301,}
+T18[K16]={[1]=K[21],}
+T19.def={[K24]=K[24],[K38]={["!*t"]=true,["*t"]=true,},f=K3,[K2]=1342,[K1]="enum",x=4,y=239,}
+T20.def={[K24]=K[23],f=K3,[K31]={[1]="year",[2]="month",[3]="day",[4]="hour",[5]="min",[6]="sec",[7]="wday",[8]="yday",[9]="isdst",},[K30]={day={f=K3,[K2]=1334,[K1]=K12,x=12,y=230,},hour={f=K3,[K2]=1335,[K1]=K12,x=13,y=231,},["isdst"]={f=K3,[K2]=1340,[K1]=K25,x=14,y=236,},min={f=K3,[K2]=1336,[K1]=K12,x=12,y=232,},["month"]={f=K3,[K2]=1333,[K1]=K12,x=14,y=229,},sec={f=K3,[K2]=1337,[K1]=K12,x=12,y=233,},wday={f=K3,[K2]=1338,[K1]=K12,x=13,y=234,},yday={f=K3,[K2]=1339,[K1]=K12,x=13,y=235,},year={f=K3,[K2]=1332,[K1]=K12,x=13,y=228,},},[K2]=1331,[K1]=K37,x=4,y=227,}
+T21.def={f=K3,t={[K24]=K[32],[K32]=T22,f=K3,[K31]={[1]="n",},[K30]={n={f=K3,[K2]=1644,[K1]=K12,x=10,y=304,},},[K[8]]={[1]={[K32]=T22,f=K3,[K2]=1643,[K1]=K33,x=10,xend=12,y=302,yend=302,},},[K[9]]=true,[K2]=1641,[K1]=K37,x=4,y=301,},[K19]={[1]={f=K3,[K10]="A",[K2]=1640,[K1]=K10,x=21,y=301,},xend=22,yend=301,},[K2]=1645,[K1]=K20,x=4,y=301,}
 T23.def={f=K3,t={args={f=K3,[K4]={[1]={f=K3,[K2]=2288,[K1]=K5,[K5]="A",x=36,y=299,},[2]={f=K3,[K2]=2289,[K1]=K5,[K5]="A",x=39,y=299,},},[K2]=1635,[K1]=K4,x=41,y=299,},f=K3,[K9]=false,[K6]=2,rets={f=K3,[K4]={[1]={f=K3,[K2]=1637,[K1]=K25,x=43,y=299,},},[K2]=1636,[K1]=K4,x=41,y=299,},[K2]=1631,[K1]=K7,x=24,y=299,},[K19]={[1]={f=K3,[K10]="A",[K2]=1632,[K1]=K10,x=33,y=299,},xend=34,yend=299,},[K2]=1638,[K1]=K20,x=4,y=301,}
 
 return { globals = T0, typeid_ctr = 2748, typevar_ctr = 53}
