+++ tl
@@ -172,14 +172,7 @@ local function setup_env(tlconfig, filename)
       predefined_modules = tlconfig._init_env_modules,
    }
 
-   local env, err
-   if #tlconfig._init_env_modules == 0 then
-      env, err = tl.default_env()
-      env.defaults = opts.defaults
-   else
-      env, err = tl.new_env(opts)
-   end
-
+   local env, err = tl.new_env(opts)
    if not env then
       die(err)
    end
